==> scripts/00_helpers.R <==

### -----------------------------------------------------------------
### This script contains helpers for the SubsidyExplorer analysis  
### Last updated on January 23, 2020
### -----------------------------------------------------------------
