==> scripts/01_data_wrangling.Rmd <==
---
output:
  html_document: default

title: "SubsidyExplorer - Step 1: Data Wrangling"
author: "Kat Millage, Vienna Saccomanno, Laura Lea Rubino, Christopher Costello"
date: "`r format(Sys.time(), '%B %d, %Y')`"
fontsize: 11pt
---

```{r}
# Packages
library(knitr) # Necessary to convert file to .html
library(countrycode) # Country name matching
library(here) # File path handling
#remotes::install_github('vincentarelbundock/WDI')
library(WDI) # World Bank World Development Indicators data
library(janitor) # Column name handling
library(tidyverse) # General data wrangling
library(readxl) # Load .xlsx files

# Code chunk defaults
knitr::opts_chunk$set(echo = FALSE, error = FALSE, message = FALSE, warning = FALSE, include = FALSE)

# Source file with helper functions (used throughout analysis)
source(here::here("scripts", "00_helpers.R"))

# Directories for the output files generated by this script
data_dir <- here::here("data", "edited/")
if (dir.exists(data_dir) == F) {
  dir.create(data_dir, recursive = T)
}

app_data_dir <- here::here("app", "data/")
if (dir.exists(app_data_dir) == F) {
  dir.create(app_data_dir, recursive = T)
}

results_dir <- here::here("results", "01-data-wrangling/")
if (dir.exists(results_dir) == F) {
  dir.create(results_dir, recursive = T)
}
```

# Description

This script handles the preliminary data wrangling for the raw data sets underlying the SubsidyExplorer analysis. It outputs tidied data files to `r data_dir`. Some of these tidied data files are used directly as input data by the SubsidyExplorer toolkit (saved to `r app_data_dir`). 

```{r}
### Paths to demographic data files
# WDI data - no path, accessed through package
fao_fishers_path <- here::here("data", "raw", "FAO-2019-yearbook", "fao_fishers_yearbook_070619.csv")
teh_fte_path <- here::here("data", "raw", "Teh-and-Sumaila-2013-fte", "Fisher employment FTE.xlsx")

### Paths to fisheries subsidies data files
sumaila_subsidies_path <- here::here("data", "raw", "Sumaila-et-al-2019-subsidies", "Subsidies_PEW_Final.xlsx")
oecd_fse_path <- here::here("data", "raw", "OECD-fisheries-support-estimates", "FISH_FSE_10032020183349581.csv")
schuhbauer_subsidies_ssf_path <- here::here("data", "raw", "Schuhbauer-et-al-2017-ssf-subsidies", "Subsidies_SSF_split_2018.csv")

### Paths to other fisheries data files
fao_production_path <- here::here("data", "raw", "FAO-global-capture-production", "Capture_2020.1.0/")
#fao_value_path
melnychuk_price_path <- here::here("data", "raw", "Melnychuk-et-al-2013-price", "Melnychuk_Exvessel_Price_Database.csv")
ram_sadb_path <- here::here("data", "raw", "RAM-legacy-stock-assessment-database", "RAMLDB v4.491", "DB Files With Assessment Data", "RAMCore/")
costello_upsides_path <- here::here("data", "raw", "Costello-et-al-2016-upsides", "ProjectionData_2012.csv")
melnychuk_fmi_path <- here::here("data", "raw", "Melnychuk-et-al-2017-fmi", "fmi_values_country.csv")
iuu_vessel_path <- here::here("data", "raw", "combined-IUU-vessel-list", "IUUList-20190430.csv")

### Paths to other necessary files
cap_tier_path <- here::here("data", "raw", "USA-cap-tier-proposal", "US_cap_tier_data.csv")
country_lookup_path <- here::here("data", "lookup-tables", "country_dependencies.csv")
fao_regions_for_ram_lookup_path <- here::here("data", "lookup-tables", "fao_areas_for_ram_stocks_v4.491.csv")
sumaila_lookup_path <- here::here("data", "lookup-tables", "sumaila_subsidies_classification_lookup_table.csv")
schuhbauer_lookup_path <- here::here("data", "lookup-tables", "schuhbauer_subsidies_classification_lookup_table.csv")
oecd_fse_lookup_path <- here::here("data", "lookup-tables", "oecd_fse_classification_lookup_table.csv")
```

The following data sources are included in this analysis: 

**Demographic Data**
Total GDP, GDP from fisheries, forestry, and agriculture, and population by state from the World Bank's World Development Indicators Database (2000 - 2018). These data were accessed through the `WDI` package. The code used to pull the data is below. 
> World Bank. (2020). World Development Indicators. https://data.worldbank.org/

Total number of fishers by state from the FAO Yearbook of Fishery and Aquaculture Statistics (misc. years between 1995 and 2017). These data were manually transcribed from PDF (`./data/raw/fleet_&_employment.pdf`) to CSV format on July 6, 2019 and are stored as `r fao_fishers_path`. 
> FAO. (2019). FAO Yearbook of Fishery and Aquaculture Statistics 2017 (p. 108).

Estimated number of full-time jobs in fisheries from Teh and Sumaila (2013) in 2003. These data were provided to us directly by the authors and are stored as `r teh_fte_path`. 
> Teh, L. C. L., & Sumaila, U. R. (2011). Contribution of marine fisheries to worldwide employment. Fish and Fisheries, 14(1), 77–88. https://doi.org/10.1111/j.1467-2979.2011.00450.x

**Fisheries Subsidies Data**
Estimates of annual fisheries subsidies provision by state from Sumaila et al. (2019) in 2018 US\$. These data were provided to us directly by the authors and are stored at `r sumaila_subsidies_path`.
> Sumaila, U. R., Ebrahim, N., Schuhbauer, A., Skerritt, D., Li, Y., Kim, H. S., … Pauly, D. (2019). Updated estimates and analysis of global fisheries subsidies. Marine Policy, 109, 103695. https://doi.org/10.1016/j.marpol.2019.103695

Estimates of fisheries support provision by state from the OECD (2008 - 2018). These data were last downloaded from the OECD iLibrary portal on March 10, 2019 and are stored as `r oecd_fse_path`.
> OECD. (2019). Fisheries: Fisheries support estimates. Fisheries and Aquaculture Statistics, OECD Agriculture Statistics (Database). https://doi.org/10.1787/ade64fdc-en

Estimates of annual fisheries subsidies provision to small scale fisheries by state in 2018 US\$. Data used in this analysis are unpublished updates (based on the Sumaila et al. (2019) data) from those reported in Schuhbauer et al. (2017) (which are in 2009 US\$). These data were provided to us directly by the authors and are stored at `r schuhbauer_subsidies_ssf_path`. 
> Schuhbauer, A., Chuenpagdee, R., Cheung, W. W. L., Greer, K., & Sumaila, U. R. (2017). How subsidies affect the economic viability of small-scale fisheries. Marine Policy, 82, 114–121. https://doi.org/10.1016/j.marpol.2017.05.013

**Other Fisheries Data**
Fisheries landings in 2018 from the FAO Global Capture Production database (2020). These data were last downloaded from the FAO website on July 25, 2020 and are stored as `r fao_production_path`. 
> FAO. (2020). Global Capture Production Dataset. http://www.fao.org/fishery/statistics/global-capture-production/en

Estimates of ex-vessel prices created by Melnychuk et al. (2017). These data were provided to us directly by the authors and are stored as `r melnychuk_price_path`. 
> Melnychuk, M. C., Clavelle, T., Owashi, B., & Strauss, K. (2017). Reconstruction of global ex-vessel prices of fished species. ICES Journal of Marine Science, 74(1), 121–133. https://doi.org/10.1093/icesjms/fsw169

The RAM Legacy Stock Assessment Database is a compilation of stock assessments for commercially exploited species from around the world. The latest version of this database (version 4.491) was downloaded directly from the RAM website on July 25, 2020 and are stored as `r ram_sadb_path`. 
> RAM Legacy Stock Assessment Database. (2020). Version 4.491-assessment only. Released 2020-02-20. https://doi.org/10.5281/zenodo.3676087

Estimates of stock status for commercially exploited species not included in the RAM Legacy Stock Assessment Database from Costello et al. (2016). These data were provided to us directly by the authors and are stored as `r costello_upsides_path`. 
> Costello, C., Ovando, D., Clavelle, T., Strauss, C. K., Hilborn, R., Melnychuk, M. C., Branch, T. A., Gaines, S. D., Szuwalski, C. S., Cabral, R. B., Rader, D. N., & Leland, A. (2016). Global fishery prospects under contrasting management regimes. Proceedings of the National Academy of Sciences, 113(18), 5125–5129. https://doi.org/10.1073/pnas.1520420113

Fisheries management indicators for many of the top fishing nations from Melnychuk et al. (2017). These data were manually transcribed from Figure S2 in the supplementary information and are stored as `r melnychuk_fmi_path`. 
> Melnychuk, M. C., Peterson, E., Elliott, M., & Hilborn, R. (2017). Fisheries management impacts on target species status. Proceedings of the National Academy of Sciences, 114(1), 178–183. https://doi.org/10.1073/pnas.1609915114

The IUU Consolidated Vessel List created and maintained by Trygg Mat Tracking. These data were downloaded directly from the website on November 30, 2019 and are stored as `r iuu_vessel_path`. 
> Trygg Mat Tracking. (2020). Combined IUU Fishing Vessel List. https://www.iuu-vessels.org/

We also use the following tables to aid in our analysis: 

- Proposed tier assignments for a cap-based approach to limiting subsidies as included in the proposal made to the WTO by the delegations of Australia and the USA on May 3, 2019 [TN/RL/GEN/197/Rev.1].

- A manually compiled table of overseas dependencies of WTO Member and Observer countries

- All subsidy types used by Sumaila et al. (2019)

- All subsidy types used by the OECD FSE database

- A manually compiled table of FAO statistical regions corresponding to the stocks contained in version 4.491 of the RAM Legacy Stock Assessment Database. 

# Note: Country naming, political entities, and dependencies

The naming of flag and coastal states varies widely across our data sources and we recognize that this can be a politically sensitive matter. For the purposes of merging multiple data sources, we therefore translate all state names from the raw data sources to ISO-3 character codes. Country names are only used for display purposes, and the use of any particular name is not meant to convey any options regarding sovereignty on behalf of the authors. 

```{r}
# Load country dependencies helper file
country_dependencies <- read_csv(country_lookup_path)

# Get all overseas dependencies that are relevant for the WTO - there are 49
territories <- country_dependencies %>%
  dplyr::filter(is_overseas_territory & is_WTO) %>%
  distinct(iso3, sovereign_iso3)

# Get all cooresponding sovereign states (we need to make sure any totals for these states include their dependencies) - there are 9
# Australia, Denmark, France, United Kingdom, Morocco, Netherlands, Norway, New Zealand, and the United States
sovereigns_with_territories <- territories %>%
  distinct(sovereign_iso3)

# EU countries
eu_countries <- country_dependencies$iso3[country_dependencies$is_EU]

# EU overseas territories
eu_territories <- country_dependencies$iso3[country_dependencies$is_overseas_territory & country_dependencies$sovereign_iso3 %in% eu_countries]
```

Additionally, it is important to note that the European Union as a political entity is a Member of the World Trade Organization, rather then the individual states that make up the European Union. We therefore need to create a profile for the European Union as a whole.  The following `r length(eu_country_names)` states are considered to be part of the European Union for the purposes of this analysis: `r paste(country_dependencies$WTO_name[country_dependencies$iso3 %in% eu_countries], collapse = ", ")`. 

The Netherlands, France, Denmark, and the United Kingdom also have overseas dependencies that should be considered to be a part of the European Union for statistical purposes.

Outside of the EU, the other dependencies we need to worry about are those that relate to the Austraila, Morocco, Norway, New Zealand, and the United States.

# Section 1 - Subsidies Data

## Estimates of global fisheries subsidies from Sumaila et al. (2019)

```{r}
### Data Wrangling
# Load and wrangle 2019 data
sumaila_2019_dat <- read_xlsx(sumaila_subsidies_path) %>%
  clean_names(case = "snake") %>%
  mutate(iso3 = case_when(country == "Dominican Rp" ~ "DOM",
                          country == "Micronesia" ~ "FSM",
                          country == "Untd Arab Em" ~ "ARE",
                          TRUE ~ countrycode(country, "country.name", "iso3c"))) %>%
  dplyr::select(iso3, type = class, data_type, value = constant_2018_usd) %>%
  mutate(year = 2018,
         units = "2018 US$",
         source = "Sumaila et al. (2019)") %>%
  arrange(iso3)

# Load lookup table for classification scheme
sumaila_lookup <- read_csv(sumaila_lookup_path)

# Clean up the naming/spelling of the classification scheme as well as data type
sumaila_2019_dat_edit <- sumaila_2019_dat %>%
  left_join(sumaila_lookup, by = "type") %>%
  dplyr::select(iso3, year, category, category_name, type, type_name, value, data_type, units, source)

sumaila_2019_dat_edit$data_type[sumaila_2019_dat_edit$data_type == "not found evidence of subsidy"] <- "Found no evidence of subsidy"

# Deal with the EU
sumaila_2019_dat_eu <- sumaila_2019_dat_edit %>%
  dplyr::filter(iso3 %in% c(eu_countries, eu_territories)) %>%
  mutate(entity = "EU",
         data_type = "Aggregate") %>%
  group_by(entity, year, category, category_name, type, type_name, data_type, units, source) %>%
  summarize(value = sum(value, na.rm = T)) %>%
  ungroup() %>%
  rename(iso3 = entity)

# Add the EU back in
sumaila_2019_dat_edit <- sumaila_2019_dat_edit %>%
  bind_rows(sumaila_2019_dat_eu) %>%
  mutate(variable = "subsidies_Sumaila")

# Test to make sure there isn't data for any other dependencies - nope we're good
sumaila_2019_dat_overseas_territory <- sumaila_2019_dat_edit %>%
  dplyr::filter(iso3 %in% territories$iso3 & !(iso3 %in% eu_territories))

# Save
write_csv(sumaila_2019_dat_edit, paste0(data_dir, "sumaila_et_al_2019_subsidies_tidy.csv"))
write_csv(sumaila_2019_dat_edit, paste0(app_data_dir, "sumaila_et_al_2019_subsidies_tidy.csv"))

```

## Estimates of fisheries support provision by state from the OECD (2008 - 2018)

```{r}
# Load and wrangle OECD data (2009 - 2017)
oecd_fse_dat <- read_csv(oecd_fse_path) %>%
  clean_names(case = "snake") %>%
  dplyr::filter(measure == "USD", # only pull estimates in USD for consistancy
                country != "OECD") %>% # remove totals
  dplyr::select(iso3 = country, code = variable, year, value) 

# Load OECD lookup table
oecd_fse_lookup <- read_csv(oecd_fse_lookup_path) %>%
  distinct(category, category_name, type, type_code, type_name)

# Match based on type - that should be plenty for the figure on SubsidyExplorer
oecd_fse_dat_edit <- oecd_fse_dat %>%
  inner_join(oecd_fse_lookup, by = c("code" = "type_code")) %>%
  mutate(data_type = "Reported",
         units = paste0(year, " US$"),
         source = "OECD FSE database") %>%
  dplyr::select(iso3, year, category, category_name, type, type_name, data_type, units, source, value)

# Deal with the EU
oecd_fse_dat_eu <- oecd_fse_dat_edit %>%
  dplyr::filter(iso3 %in% c(eu_countries, eu_territories)) %>%
  mutate(entity = "EU",
         data_type = "Aggregate") %>%
  group_by(entity, year, category, category_name, type, type_name, data_type, units, source) %>%
  summarize(value = sum(value, na.rm = T)) %>%
  ungroup() %>%
  rename(iso3 = entity) %>%
  arrange(iso3, year, type)

# Add EU back in 
oecd_fse_dat_edit <- oecd_fse_dat_edit %>%
  bind_rows(oecd_fse_dat_eu) %>%
  mutate(variable = "subsidies_OECD")

# Test to make sure there isn't data for any other dependencies - nope we're good
oecd_fse_dat_overseas_territory <- oecd_fse_dat_edit %>%
  dplyr::filter(iso3 %in% territories$iso3 & !(iso3 %in% eu_territories))
  
# Save
write_csv(oecd_fse_dat_edit, paste0(data_dir, "oecd_2019_fse_tidy.csv"))
write_csv(oecd_fse_dat_edit, paste0(app_data_dir, "oecd_2019_fse_tidy.csv"))

```

## Updated estimates of small-scale fisheries subsides from Schuhbauer et al. (2017)

Note: these data are unpublished updates (in 2018 US\$) of the data described in Schuhbauer et al. (2017). 

```{r}
# Load and wrangle data (2018 US$)
schuhbauer_subsidies_ssf_dat <- read_csv(schuhbauer_subsidies_ssf_path) %>%
  clean_names(case = "snake") %>%
  mutate(iso3 = case_when(country == "Dominican Rp" ~ "DOM",
                          country == "Micronesia" ~ "FSM",
                          country == "Untd Arab Em" ~ "ARE",
                          TRUE ~ countrycode(country, "country.name", "iso3c"))) %>%
  mutate(prop_ssf = ssf_usd/constant_2018_usd) %>%
  dplyr::select(iso3, old_type_name = type, prop_ssf) %>%
  mutate(source = "Schuhbauer et al. (2017)") %>%
  arrange(iso3)
schuhbauer_subsidies_ssf_dat$prop_ssf[is.nan(schuhbauer_subsidies_ssf_dat$prop_ssf)] <- 0 # handling dividing by 0

# Load lookup table
schuhbauer_lookup <- read_csv(schuhbauer_lookup_path)

# Clean up the naming/spelling of the classification scheme as well as data type
schuhbauer_subsidies_ssf_dat_edit <- schuhbauer_subsidies_ssf_dat %>%
  left_join(schuhbauer_lookup, by = c("old_type_name")) %>%
  dplyr::select(iso3, category, category_name, type, type_name, prop_ssf, source)

# Deal with the EU
schuhbauer_subsidies_ssf_dat_eu <- read_csv(schuhbauer_subsidies_ssf_path) %>%
  clean_names(case = "snake") %>%
  mutate(iso3 = case_when(country == "Dominican Rp" ~ "DOM",
                          country == "Micronesia" ~ "FSM",
                          country == "Untd Arab Em" ~ "ARE",
                          TRUE ~ countrycode(country, "country.name", "iso3c"))) %>%
  dplyr::filter(iso3 %in% c(eu_countries, eu_territories)) %>%
  mutate(entity = "EU") %>%
  group_by(entity, type) %>%
  summarize(constant_2018_usd = sum(constant_2018_usd, na.rm = T),
            ssf_usd = sum(ssf_usd, na.rm = T)) %>%
  ungroup() %>%
  mutate(prop_ssf = ssf_usd/constant_2018_usd) %>%
  dplyr::select(iso3 = entity, old_type_name = type, prop_ssf) %>%
  mutate(source = "Schuhbauer et al. (2017)") %>%
  left_join(schuhbauer_lookup, by = c("old_type_name")) %>%
  dplyr::select(iso3, category, category_name, type, type_name, prop_ssf, source)

# Add the EU back in
schuhbauer_subsidies_ssf_dat_edit <- schuhbauer_subsidies_ssf_dat_edit %>%
  bind_rows(schuhbauer_subsidies_ssf_dat_eu)

# Test to make sure there isn't data for any other dependencies - nope we're good
schuhbauer_subsidies_ssf_dat_overseas_territory <- schuhbauer_subsidies_ssf_dat_edit %>%
  dplyr::filter(iso3 %in% territories$iso3 & !(iso3 %in% eu_territories))

# Save
write_csv(schuhbauer_subsidies_ssf_dat_edit, paste0(data_dir, "schuhbauer_ssf_subsidies_proportions.csv"))
write_csv(schuhbauer_subsidies_ssf_dat_edit, paste0(app_data_dir, "schuhbauer_ssf_subsidies_proportions.csv"))
```

# Section 2 - Demographic Data

## Demographic statistics by state from the World Bank's World Development Indicators Database (2000 - 2018)

```{r}
#Load data using WDI package
wb_demo_dat <- WDI(country = "all",
                   indicator = c("gdp" = "NY.GDP.MKTP.CD", # GDP (current US$)
                                 "gdp_ffa" = "NV.AGR.TOTL.CD", # Added value coming from fisheries, forestry, and agriculture (current US$)
                                 "population" = "SP.POP.TOTL", # Total population
                                 "gni_per_capita" = "NY.GNP.PCAP.KD"), # GNI per capita (constant 2010 US$)
                                  start = 2000,
                                  end = 2018,
                                  extra = TRUE) %>%
  dplyr::filter(region != "Aggregates") %>%
  dplyr::select(iso3 = iso3c,
                year,
                gdp,
                gdp_ffa,
                population,
                gni_per_capita) %>%
  mutate(iso3 = as.character(iso3)) %>%
  arrange(iso3, year) %>%
  gather(variable, value, c("gdp", "gdp_ffa", "population", "gni_per_capita")) %>%
  dplyr::filter(!is.na(value)) %>%
  mutate(units = case_when(variable == "population" ~ "individuals",
                           variable == "gni_per_capita" ~ "constant 2010 US$",
                           TRUE ~ paste0(year, " US$")),
         source = "World Bank WDI database")

# # Deal with the EU + territories
wb_dat_eu <- wb_demo_dat %>%
  dplyr::filter(iso3 %in% c(eu_countries, eu_territories)) %>%
  mutate(entity = "EU") %>%
  group_by(entity, year, variable, units, source) %>%
  summarize(value = sum(value, na.rm = T)) %>%
  rename(iso3 = entity) %>%
  arrange(iso3, year, variable)

# Deal with data for other dependencies
wb_dat_overseas_territories <- wb_demo_dat %>%
  dplyr::filter(iso3 %in% territories$iso3 | iso3 %in% territories$sovereign_iso3) %>%
  left_join(territories, by = c("iso3")) %>%
  mutate(sovereign_iso3 = case_when(is.na(sovereign_iso3) ~ iso3,
                                    TRUE ~ sovereign_iso3)) %>%
  mutate(entity = paste0(sovereign_iso3, "-T")) %>%
  group_by(entity, year, variable, units, source) %>%
  summarize(value = sum(value, na.rm = T)) %>%
  rename(iso3 = entity) %>%
  arrange(iso3, year, variable)

# Add back in
wb_dat_edit <- wb_demo_dat %>%
  bind_rows(wb_dat_eu) %>%
  bind_rows(wb_dat_overseas_territories)

# Export clean
write_csv(wb_dat_edit, paste0(data_dir, "world_bank_2020_wdi_tidy.csv"))
write_csv(wb_dat_edit, paste0(app_data_dir, "world_bank_2020_wdi_tidy.csv"))
```

## Total number of fishers by state from the FAO Yearbook of Fishery and Aquaculture Statistics (misc. years between 1995 and 2016)

```{r}
# Pull data and wrangle
fao_fishers_dat <- read_csv(fao_fishers_path) %>%
  mutate(iso3 = countrycode(country, "country.name", "iso3c")) %>%
  dplyr::select(-country) %>%
  gather(year, value, c('1995', '2000', '2005', '2010', '2011', '2012', '2013', '2014', '2015', '2016')) %>%
  mutate(year = as.numeric(year)) %>%
  dplyr::filter(!is.na(value)) %>%
  mutate(variable = "fishers",
         units = "individuals",
         source = "FAO Yearbook of Fishery and Aquaculture Statistics (2017)")

# Deal with EU
fao_fishers_dat_eu <- fao_fishers_dat %>%
  dplyr::filter(iso3 %in% c(eu_countries, eu_territories)) %>%
  mutate(entity = "EU") %>%
  group_by(entity, year, variable, units, source) %>%
  summarize(value = sum(value, na.rm = T)) %>%
  ungroup() %>%
  rename(iso3 = entity) %>%
  arrange(iso3, year, variable)

# Add EU back in 
fao_fishers_dat_edit <- fao_fishers_dat %>%
  bind_rows(fao_fishers_dat_eu) %>%
  arrange(year, iso3)

# # Test to make sure there isn't data for any other dependencies - nope we're good
fao_fishers_dat_overseas_territory <- fao_fishers_dat %>%
  dplyr::filter(iso3 %in% territories$iso3 & !(iso3 %in% eu_territories))

# Export clean
write_csv(fao_fishers_dat_edit, paste0(data_dir, "fao_2017_yearbook_fishers_tidy.csv"))
write_csv(fao_fishers_dat_edit, paste0(app_data_dir, "fao_2017_yearbook_fishers_tidy.csv"))
```

## Estimated number of full-time jobs in fisheries from Teh and Sumaila (2013) in 2003

```{r}
# Pull data and wrangle
teh_fte_dat <- read_xlsx(teh_fte_path) %>%
  clean_names(case = "snake") %>%
  mutate(iso3 = case_when(country == "Dominican Rp" ~ "DOM",
                          country == "Micronesia" ~ "FSM",
                          country == "Untd Arab Em" ~ "ARE",
                          TRUE ~ countrycode(country, "country.name", "iso3c"))) %>%
  dplyr::select(-country) %>%
  mutate(variable = "fishers_fte",
         year = 2003,
         units = "full-time equivalent jobs",
         source = "Teh and Sumaila (2013)") %>%
  dplyr::select(iso3, year, variable, value = total_fte, units, source)

# Deal with EU
teh_fte_dat_eu <- teh_fte_dat %>%
  dplyr::filter(iso3 %in% c(eu_countries, eu_territories)) %>%
  mutate(entity = "EU") %>%
  group_by(entity, year, variable, units, source) %>%
  summarize(value = sum(value, na.rm = T)) %>%
  ungroup() %>%
  rename(iso3 = entity) %>%
  arrange(iso3, year, variable)

# Add EU back in 
teh_fte_dat_edit <- teh_fte_dat %>%
  bind_rows(teh_fte_dat_eu)

# # Test to make sure there isn't data for any other dependencies - nope we're good
teh_fte_dat_overseas_territory <- teh_fte_dat_edit %>%
  dplyr::filter(iso3 %in% territories$iso3 & !(iso3 %in% eu_territories))

# Export clean
write_csv(teh_fte_dat_edit, paste0(data_dir, "teh_and_sumaila_2013_fte_tidy.csv"))
write_csv(teh_fte_dat_edit, paste0(app_data_dir, "teh_and_sumaila_2013_fte_tidy.csv"))
```

# Section 3 - Fisheries Data

## Fisheries landings for 2000-2017 from the FAO Global Capture Production database (2019)

```{r}
# FAO stores a lot of characteristics in lookup tables that come with the production dataset. We need to use these tables to match country names, species names, and source
fao_country_lookup <- read_csv(paste0(fao_production_path, "CL_FI_COUNTRY_GROUPS.csv")) %>%
  clean_names(case = "snake") %>%
  dplyr::select(country = un_code,
                iso3 = iso3_code,
                name = name_en) %>%
  mutate(iso3_new = case_when(name == "Belgium-Luxembourg" ~ "BEL",
                              name %in% c("Canton and Enderbury Is", "Czechoslovakia", "Neutral Zone", "Pacific Islands Trust Tr", "Serbia and Montenegro", "Un. Sov. Soc. Rep.", "Yemen Arab Republic", "Yemen, Democratic", "Yugoslavia SFR") ~ "XXX",
                              name %in% c("Channel Islands", "Sark") ~ "GGY",
                              name == "Eswatini" ~ "SWZ",
                              name == "European Union" ~ "EU",
                              name %in% c("Johnston Island", "Midway Islands", "Wake Island", "US Minor Outlying Is.") ~ "UMI",
                              name == "Netherlands Antilles" ~ "ANT",
                              name == "Saint-Martin" ~ "MAF",
                              name == "Zanzibar" ~ "TZA",
                              name == "Other nei" ~ "UNK",
                              is.na(name) ~ "UNK",
                              TRUE ~ countrycode(name, "country.name", "iso3c"))) %>%
  dplyr::filter(iso3_new != "XXX") %>% # remove entities that no longer exist today
  dplyr::select(iso3 = iso3_new, country) %>%
  mutate(country = as.integer(country))

# FAO species lookup 
fao_species_lookup <- read_csv(paste0(fao_production_path, "CL_FI_SPECIES_GROUPS.csv")) %>%
  clean_names(case = "snake") %>%
  dplyr::select(species = x3alpha_code,
                asfis_species = name_en,
                scientific_name,
                yearbook_group,
                isscaap_group)

# FAO Symbol lookup
fao_symbol_lookup <- read_csv(paste0(fao_production_path, "CL_FI_SYMBOL.csv"))

# FAO Area lookup
fao_area_lookup <- read_csv(paste0(fao_production_path, "CL_FI_WATERAREA_GROUPS.csv")) %>%
  clean_names(case = "snake") %>%
  dplyr::select(area = code) %>%
  mutate(area = as.integer(area))
  
# Load data
fao_production_dat <- read.csv(paste0(fao_production_path, "TS_FI_CAPTURE.csv")) %>%
  clean_names(case = "snake") %>%
  dplyr::filter(year >= 2000) %>%
  dplyr::filter(unit == "t") %>% # remove catches measured in number of individuals
  mutate(species = as.character(species)) %>%
  left_join(fao_country_lookup, by = "country") %>%
  inner_join(fao_area_lookup, by = c("fishing_area" = "area")) %>%
  dplyr::filter(fishing_area > 10) %>% # remove inland fishing
  group_by(iso3, year, species, fishing_area) %>%
  summarize(value = sum(quantity, na.rm = T)) %>%
  ungroup() %>%
  left_join(fao_species_lookup, by = "species")

# Deal with the EU + territories
fao_production_dat_eu <- fao_production_dat %>%
  dplyr::filter(iso3 %in% c(eu_countries, eu_territories, "EU")) %>%
  mutate(entity = "EU") %>%
  group_by(entity, year, fishing_area, species, asfis_species, scientific_name, yearbook_group, isscaap_group) %>%
  summarize(value = sum(value, na.rm = T)) %>%
  ungroup() %>%
  rename(iso3 = entity)

# Deal with all territories 
fao_production_dat_overseas_territories <- fao_production_dat %>%
  dplyr::filter(iso3 %in% territories$iso3 | iso3 %in% territories$sovereign_iso3) %>%
  left_join(territories, by = c("iso3")) %>%
  mutate(sovereign_iso3 = case_when(is.na(sovereign_iso3) ~ iso3,
                                    TRUE ~ sovereign_iso3)) %>%
  mutate(entity = paste0(sovereign_iso3, "-T")) %>%
  group_by(entity, year, fishing_area, species, asfis_species, scientific_name, yearbook_group, isscaap_group) %>%
  summarize(value = sum(value, na.rm = T)) %>%
  ungroup() %>%
  rename(iso3 = entity)

# Put them back together
fao_production_dat_edit <- fao_production_dat %>%
  dplyr::filter(iso3 != "EU") %>%
  bind_rows(fao_production_dat_eu) %>%
  bind_rows(fao_production_dat_overseas_territories) %>%
  mutate(variable = "capture_production",
         units = "tonnes",
         source = "FAO (2020)")

write_csv(fao_production_dat_edit, paste0(data_dir, "fao_2020_capture_production_tidy.csv"))

# Aggregate by ISCCAP groups and export for time series plot
fao_production_dat_edit_groups <- fao_production_dat_edit %>%
  group_by(iso3, year, isscaap_group, variable, units, source) %>%
  summarize(value = sum(value, na.rm = T)) %>%
  ungroup() %>%
  spread(year, value) %>% # adding empty rows to make the plotting work better
  gather(year, value, as.character(c(2000:2018))) %>% # second part of that process
  mutate(year = as.numeric(year))
fao_production_dat_edit_groups$value[is.na(fao_production_dat_edit_groups$value)] <- 0 # for plotting purposes - we don't get gaps on area plot this way

write_csv(fao_production_dat_edit_groups, paste0(data_dir, "fao_2020_capture_production_isscaap_groups_tidy.csv"))
write_csv(fao_production_dat_edit_groups, paste0(app_data_dir, "fao_2020_capture_production_isscaap_groups_tidy.csv"))
```

## Ex-vessel prices from Melnychuk et al. (2017)
```{r}
# Price data from Melynchuk et al. (2017)
melnychuk_price_dat <- read_csv(melnychuk_price_path) %>%
  clean_names(case = "snake") %>%
  dplyr::select(-entry) %>%
  dplyr::filter(year >= 2000 & year <= 2012) %>%
  group_by(year, scientific_name, asfis_species, isscaap_group, group_for_pairing) %>%
  summarize(exvessel = mean(exvessel, na.rm = T)) %>%
  ungroup()

# Pull data from 2012 to use for extrapolation to 2017
melnychuk_price_dat_mry <- melnychuk_price_dat %>%
  dplyr::filter(year == 2012) %>%
  rename(exvessel_for_extrap = exvessel) %>%
  dplyr::select(-year)

# CPI for 2012 - 2018 to convert USD
cpi_2012 <- 229.594
cpi_2013 <- 232.957
cpi_2014 <- 236.736
cpi_2015 <- 237.017
cpi_2016 <- 240.007
cpi_2017 <- 245.120
cpi_2018 <- 251.107

# Do extrapolation for missing years
melnychuk_price_dat_extrapolation <- melnychuk_price_dat %>%
  complete(nesting(scientific_name, asfis_species, isscaap_group, group_for_pairing), year = 2000:2018) %>%
  left_join(melnychuk_price_dat_mry, by = c("scientific_name", "asfis_species", "isscaap_group", "group_for_pairing")) %>%
  mutate(exvessel_adjusted = case_when(year <= 2012 ~ exvessel,
                                       year == 2013 ~ exvessel_for_extrap * (cpi_2013/cpi_2012),
                                       year == 2014 ~ exvessel_for_extrap * (cpi_2014/cpi_2012),
                                       year == 2015 ~ exvessel_for_extrap * (cpi_2015/cpi_2012),
                                       year == 2016 ~ exvessel_for_extrap * (cpi_2016/cpi_2012),
                                       year == 2017 ~ exvessel_for_extrap * (cpi_2017/cpi_2012),
                                       year == 2018 ~ exvessel_for_extrap * (cpi_2018/cpi_2012))) %>%
  dplyr::select(-exvessel, -exvessel_for_extrap) %>%
  rename(exvessel = exvessel_adjusted)

# Get rid of NAs and save
melnychuk_price_dat_extrapolation <- melnychuk_price_dat_extrapolation %>%
  dplyr::filter(!is.nan(exvessel)) %>%
  mutate(variable = "ex_vessel_price",
         units = paste(year, "US$"),
         source = "Melnychuk et al. (2017)")

write_csv(melnychuk_price_dat_extrapolation, paste0(data_dir, "melnychuk_2017_price_tidy.csv"))

# Create averages by pairing group for each year
price_dat_group_year <- melnychuk_price_dat_extrapolation %>%
  group_by(year, group_for_pairing) %>%
  summarize(exvessel_group_avg = mean(exvessel, na.rm = T)) %>%
  ungroup()

# Create averages by pairing group (no year) 
price_dat_group <- melnychuk_price_dat_extrapolation %>%
  group_by(group_for_pairing) %>%
  summarize(exvessel_group_avg_allyears = mean(exvessel, na.rm = T)) %>%
  ungroup()
```

## Estimated landed value from FAO/Melnychuk et al. (2017)

```{r}
# Combine landings and price data
combine_catch_price_dat <- fao_production_dat_edit %>%
  left_join(melnychuk_price_dat_extrapolation %>% dplyr::select(-isscaap_group, -variable, -source, -units), by = c("year", "scientific_name", "asfis_species", "isscaap_group" = "group_for_pairing")) %>%
  left_join(price_dat_group_year, by = c("year", "isscaap_group" = "group_for_pairing")) %>%
  left_join(price_dat_group, by = c("isscaap_group" = "group_for_pairing")) %>%
  mutate(exvessel_inclusive = case_when(!is.na(exvessel) ~ exvessel,
                                        !is.na(exvessel_group_avg) ~ exvessel_group_avg,
                                        !is.na(exvessel_group_avg_allyears) ~ exvessel_group_avg_allyears))

# Look at NAs - 11,553 entries
combine_nas <- combine_catch_price_dat %>%
  dplyr::filter(is.na(exvessel_inclusive))

combine_nas_groups <- unique(combine_nas$isscaap_group)

# Set price to 0 for those
combine_catch_price_dat$exvessel_inclusive[is.na(combine_catch_price_dat$exvessel_inclusive)] <- 0

# Calculate revenues and export
landed_value_dat <- combine_catch_price_dat %>%
  dplyr::select(-exvessel, -exvessel_group_avg, -exvessel_group_avg_allyears) %>%
  rename(exvessel = exvessel_inclusive) %>%
  mutate(landed_value = value*exvessel) %>%
  group_by(iso3, year, fishing_area, species, asfis_species, scientific_name, yearbook_group, isscaap_group) %>%
  summarize(value = sum(landed_value, na.rm = T)) %>%
  mutate(variable = "landed_value",
         units = paste(year, "US$"),
         source = "FAO (2020); Melnychuk et al. (2017)")

write_csv(landed_value_dat, paste0(data_dir, "estimated_landed_value_tidy.csv"))

# Calculate revenues and aggregate by ISSCAAP group
landed_value_dat_groups <- landed_value_dat %>%
  group_by(iso3, year, isscaap_group, variable, source) %>%
  summarize(value = sum(value, na.rm = T)) %>%
  ungroup() %>%
  spread(year, value) %>% # adding empty rows to make the plotting work better
  gather(year, value, as.character(c(2000:2018))) %>% # second part of that process
  mutate(year = as.numeric(year)) %>%
  mutate(units = paste0(year, " US$"))

write_csv(landed_value_dat_groups, paste0(data_dir, "estimated_landed_value_isscaap_groups_tidy.csv"))
write_csv(landed_value_dat_groups, paste0(app_data_dir, "estimated_landed_value_isscaap_groups_tidy.csv"))
```

## RAM Legacy Stock Assessment database 

```{r}
# Load B/Bmsy timeseries from the RAM legacy stock assessment database
ram_dat_bv_bmsy <- read_csv(paste0(ram_sadb_path, "BvB.csv")) %>%
  rename(year = X1) %>%
  gather(stockid, bv_bmsy, -year) %>%
  na.omit() %>%
  group_by(stockid) %>%
  mutate(year_max = max(year)) %>%
  ungroup() %>%
  dplyr::filter(year == year_max)

# Load lookup table applying FAO regions to the RAM stocks
ram_lookup <- read_csv(fao_regions_for_ram_lookup_path) 

# Apply to RAM data and create a new row for each FAO region
ram_dat_bv_bmsy_edit <- ram_dat_bv_bmsy %>%
  left_join(ram_lookup, by = "stockid") %>%
  mutate(region_fao = fao_area) %>%
  mutate(region_fao_split = str_split(region_fao, ";")) %>%
  unnest(cols = c(region_fao_split)) %>%
  group_by(stockid) %>%
  mutate(n_regions = n_distinct(region_fao_split)) %>%
  ungroup()

# Save clean
write_csv(ram_dat_bv_bmsy_edit, paste0(data_dir, "RAMLDB_v4.491_bvmsy_tidy.csv"))
```


## Estimates of global stock status from Costello et al. (2016)

```{r}
# Load data
costello_upsides_dat <- read_csv(costello_upsides_path) %>%
  clean_names(case = "snake")

# We need to split the entries that span multiple FAO regions 
costello_upsides_dat_edit <- costello_upsides_dat %>%
  mutate(region_fao_chr = as.character(region_fao),
         n_regions = nchar(region_fao_chr)/2,
         region_fao_split = str_split(gsub("(.{2})", "\\1 ", region_fao_chr), " ")) %>%
  unnest(cols = c(region_fao_split)) %>%
  mutate(region_fao_split = as.numeric(region_fao_split)) %>%
  dplyr::filter(region_fao_split != "") %>%
  mutate(iso3 = case_when(country == "Channel Islands" ~ "JEY",
                          country == "Micronesia" ~ "FSM",
                          country == "Zanzibar" ~ "TZA",
                          TRUE ~ countrycode(country, "country.name", "iso3c")))

# Save clean
write_csv(costello_upsides_dat_edit, paste0(data_dir, "costello_et_al_2016_upsides_tidy.csv"))

```

# Melnychuk FMI
```{r}
# Load data and wrangle
melnychuk_fmi_dat <- read_csv(melnychuk_fmi_path) %>%
  clean_names(case = "snake") %>%
  mutate(iso3 = countrycode(country, "country.name", "iso3c")) %>%
  dplyr::select(-country) %>%
  gather(variable, value, -iso3)

# Create a mean aggregate for the EU
melnychuk_fmi_dat_eu <- melnychuk_fmi_dat  %>%
  dplyr::filter(iso3 %in% c(eu_countries, eu_territories)) %>%
  mutate(entity = "EU") %>%
  group_by(entity, variable) %>%
  summarize(value = mean(value, na.rm = T)) %>%
  rename(iso3 = entity)

# Add EU back in
melnychuk_fmi_dat_edit <- melnychuk_fmi_dat %>%
  bind_rows(melnychuk_fmi_dat_eu)

# Export clean
write_csv(melnychuk_fmi_dat_edit, paste0(data_dir, "melnychuk_et_al_2017_fmi_tidy.csv"))

```

# Section 5 - Relative Metrics

This section is a bit of a relic from when we were doing "country profiles", but there are a couple of relative subsidy metrics we want to calculate from our data sources. This is done here. 

```{r}
# Sumaila dat for matching
subs_dat_match <- sumaila_2019_dat_edit %>%
  dplyr::select(-units, -variable) %>%
  rename(subs = value, subs_year = year, subs_source = source)

# GDP
gdp_dat_mry <- wb_dat_edit %>%
  dplyr::filter(variable == "gdp") %>%
  group_by(iso3, variable) %>%
  dplyr::filter(year == max(year[!is.na(value)])) %>%
  ungroup() %>%
  dplyr::select(iso3, gdp_year = year, gdp = value, gdp_source = source) %>%
  right_join(subs_dat_match, by = c("iso3")) %>%
  mutate(variable = "subsidies_per_gdp",
         data_type = paste0(data_type),
         source = paste0(subs_source, "; ", gdp_source),
         value = subs/gdp,
         units = paste0(subs_year, " US$/", gdp_year, " US$")) %>%
  dplyr::select(iso3, category, category_name, type, type_name, variable, value, units, data_type, source) %>%
  dplyr::filter(!is.na(value))

# Population
pop_dat_mry <- wb_dat_edit %>%
  dplyr::filter(variable == "population") %>%
  group_by(iso3, variable) %>%
  dplyr::filter(year == max(year[!is.na(value)])) %>%
  ungroup() %>%
  dplyr::select(iso3, population_year = year, population = value, population_source = source) %>%
  right_join(subs_dat_match, by = c("iso3")) %>%
  mutate(variable = "subsidies_per_capita",
         data_type = paste0(data_type),
         source = paste0(subs_source, "; ", population_source),
         value = subs/population,
         units = paste0(subs_year, " US$/person (", population_year, ")")) %>%
  dplyr::select(iso3, category, category_name, type, type_name, variable, value, units, data_type, source) %>%
  dplyr::filter(!is.na(value))

# Fishers
fisher_dat_mry <- fao_fishers_dat_edit %>%
  group_by(iso3, variable) %>%
  dplyr::filter(year == max(year[!is.na(value)])) %>%
  ungroup() %>%
  dplyr::select(iso3, fishers_year = year, fishers = value, fishers_source = source) %>%
  right_join(subs_dat_match, by = c("iso3")) %>%
  mutate(variable = "subsidies_per_fisher",
         data_type = paste0(data_type),
         source = paste0(subs_source, "; ", fishers_source),
         value = subs/fishers,
         units = paste0(subs_year, " US$/fisher (", fishers_year, ")")) %>%
  dplyr::select(iso3, category, category_name, type, type_name, variable, value, units, data_type, source) %>%
  dplyr::filter(!is.na(value))

# FTE
fte_dat_mry <- teh_fte_dat_edit %>%
  dplyr::select(iso3, fte_year = year, fte = value, fte_source = source) %>%
  right_join(subs_dat_match, by = c("iso3")) %>%
  mutate(variable = "subsidies_per_fte",
         data_type = paste0(data_type),
         source = paste0(subs_source, "; ", fte_source),
         value = subs/fte,
         units = paste0(subs_year, " US$/full-time-equivalent (", fte_year, ")")) %>%
  dplyr::select(iso3, category, category_name, type, type_name, variable, value, units, data_type, source) %>%
  dplyr::filter(!is.na(value))

# Capture Production
production_dat_mry <- fao_production_dat_edit_groups %>%
  group_by(iso3, year, variable) %>%
  summarize(value = sum(value, na.rm = T),
            source = unique(source)) %>%
  group_by(iso3, variable) %>%
  dplyr::filter(year == max(year[!is.na(value)])) %>%
  ungroup() %>%
  dplyr::select(iso3, production_year = year, production = value, production_source = source) %>%
  right_join(subs_dat_match, by = c("iso3")) %>%
  mutate(variable = "subsidies_per_production",
         data_type = paste0(data_type),
         source = paste0(subs_source, "; ", production_source),
         value = subs/production,
         units = paste0(subs_year, " US$/tonne (", production_year, ")")) %>%
  dplyr::select(iso3, category, category_name, type, type_name, variable, value, units, data_type, source) %>%
  dplyr::filter(!is.na(value))

# Landed Value
landed_value_dat_mry <- landed_value_dat %>%
  group_by(iso3, year, variable) %>%
  summarize(value = sum(value, na.rm = T),
            source = unique(source)) %>%
  group_by(iso3, variable) %>%
  dplyr::filter(year == max(year[!is.na(value)])) %>%
  ungroup() %>%
  dplyr::select(iso3, landed_value_year = year, landed_value = value, landed_value_source = source) %>%
  right_join(subs_dat_match, by = c("iso3")) %>%
  mutate(variable = "subsidies_per_landed_value",
         data_type = paste0(data_type),
         source = paste0(subs_source, "; ", landed_value_source),
         value = subs/landed_value,
         units = paste0(subs_year, " US$/", landed_value_year, " US$")) %>%
  dplyr::select(iso3, category, category_name, type, type_name, variable, value, units, data_type, source) %>%
  dplyr::filter(!is.na(value))

# Combine with subsidy data and calculate relative metrics
relative_dat <-  gdp_dat_mry %>%
  bind_rows(pop_dat_mry) %>%
  bind_rows(fisher_dat_mry) %>%
  bind_rows(fte_dat_mry) %>%
  bind_rows(production_dat_mry) %>%
  bind_rows(landed_value_dat_mry)

write_csv(relative_dat, paste0(data_dir, "relative_subsidy_metrics_tidy.csv"))
write_csv(relative_dat, paste0(app_data_dir, "relative_subsidy_metrics_tidy.csv"))
```

# Section 6 - Cap/Tier Lookup Tables

## Cap and Tier Data from the US Proposal
```{r}
cap_tier_dat <- read_csv(cap_tier_path) %>%
  mutate(iso3 = case_when(country == "Eswatini" ~ "SWZ",
                          country == "EU" ~ "EU",
                          TRUE ~ countrycode(country, "country.name", "iso3c"))) %>%
  dplyr::select(-country)

write_csv(cap_tier_dat, paste0(data_dir, "USA_cap_tier_tidy.csv"))
write_csv(cap_tier_dat, paste0(app_data_dir, "USA_cap_tier_tidy.csv"))
```

Many of the capping proposals utilize some metric to sort Members into tiers. This is often done based upon the average over some period. In order to allow this to be interactive later on, we need to craft a lookup table with all possible date ranges. 

```{r}
# Capture production lookup
capture_production_lookup <- fao_production_dat_edit_groups %>%
  group_by(iso3, year) %>%
  summarize(capture_production = sum(value, na.rm = T)) %>%
  ungroup() %>%
  spread(year, capture_production) %>%
  mutate(variable = "capture_production") %>%
  left_join(country_dependencies %>% dplyr::select(iso3, is_WTO), by = "iso3") %>%
  mutate(is_WTO = case_when(!is.na(is_WTO) ~ is_WTO,
                            iso3 %in% unique(paste0(territories$sovereign_iso3, "-T")) ~ TRUE,
                            iso3 == "EU" ~ TRUE,
                            TRUE ~ FALSE))

# Landed value
landed_value_lookup <- landed_value_dat_groups %>%
  group_by(iso3, year) %>%
  summarize(landed_value = sum(value, na.rm = T)) %>%
  ungroup() %>%
  spread(year, landed_value) %>%
  mutate(variable = "landed_value") %>%
  left_join(country_dependencies %>% dplyr::select(iso3, is_WTO), by = "iso3") %>%
  mutate(is_WTO = case_when(!is.na(is_WTO) ~ is_WTO,
                            iso3 %in% unique(paste0(territories$sovereign_iso3, "-T")) ~ TRUE,
                            iso3 == "EU" ~ TRUE,
                            TRUE ~ FALSE))

# Subsidy lookup 
subsidy_totals <- sumaila_2019_dat_edit %>%
  group_by(iso3, year) %>%
  summarize(good_subs = sum(value[type %in% c("A1", "A2", "A3")], na.rm = T),
            bad_subs = sum(value[type %in% c("B1", "B2", "B3", "B4", "B5", "B6", "B7")], na.rm = T),
            ugly_subs = sum(value[type %in% c("C1", "C2", "C3")], na.rm = T)) %>%
  ungroup() %>%
  mutate(all_subs = rowSums(select(., one_of(c("good_subs", "bad_subs", "ugly_subs")))),
         bad_ugly_subs = rowSums(select(., one_of(c("bad_subs", "ugly_subs"))))) %>%
  gather(variable, value, -c("iso3", "year")) %>%
  spread(year, value) %>%
  left_join(country_dependencies %>% dplyr::select(iso3, is_WTO), by = "iso3") %>%
  mutate(is_WTO = case_when(!is.na(is_WTO) ~ is_WTO,
                            iso3 %in% unique(paste0(territories$sovereign_iso3, "-T")) ~ TRUE,
                            iso3 == "EU" ~ TRUE,
                            TRUE ~ FALSE))

subsidy_lookup <- sumaila_2019_dat_edit %>%
  dplyr::select(iso3, year, type, value) %>%
  mutate(variable = paste0(type, "_subs")) %>%
  dplyr::select(-type) %>%
  group_by(iso3, year, variable) %>%
  summarize(value = sum(value, na.rm = T)) %>%
  spread(year, value) %>%
  left_join(country_dependencies %>% dplyr::select(iso3, is_WTO), by = "iso3") %>%
  mutate(is_WTO = case_when(!is.na(is_WTO) ~ is_WTO,
                            iso3 %in% unique(paste0(territories$sovereign_iso3, "-T")) ~ TRUE,
                            iso3 == "EU" ~ TRUE,
                            TRUE ~ FALSE)) %>%
  bind_rows(subsidy_totals)


# Demographic lookup 
demographic_lookup <- wb_dat_edit %>%
  dplyr::select(-units, -source) %>%
  spread(year, value) %>%
  left_join(country_dependencies %>% dplyr::select(iso3, is_WTO), by = "iso3") %>%
  mutate(is_WTO = case_when(!is.na(is_WTO) ~ is_WTO,
                            iso3 %in% unique(paste0(territories$sovereign_iso3, "-T")) ~ TRUE,
                            iso3 == "EU" ~ TRUE,
                            TRUE ~ FALSE))

# Fishers lookup
fishers_lookup <- fao_fishers_dat_edit %>%
  dplyr::select(-units, -source) %>%
  spread(year, value) %>%
  dplyr::select(-`1995`) %>%
  left_join(country_dependencies %>% dplyr::select(iso3, is_WTO), by = "iso3") %>%
  mutate(is_WTO = case_when(!is.na(is_WTO) ~ is_WTO,
                            iso3 %in% unique(paste0(territories$sovereign_iso3, "-T")) ~ TRUE,
                            iso3 == "EU" ~ TRUE,
                            TRUE ~ FALSE))

# Combine 
cap_tier_lookup <- capture_production_lookup %>%
  bind_rows(landed_value_lookup) %>%
  bind_rows(subsidy_lookup) %>%
  bind_rows(demographic_lookup) %>%
  bind_rows(fishers_lookup)

write_csv(cap_tier_lookup, paste0(data_dir, "cap_tier_lookup_table.csv"))
write_csv(cap_tier_lookup, paste0(app_data_dir, "cap_tier_lookup_table.csv"))

```
