==> scripts/03_management.Rmd <==
---
output:
  html_document: default

title: "SubsidyExplorer - Step 3: Classify vessels into management regimes"
author: "Kat Millage, Vienna Saccomanno, Laura Lea Rubino, Christopher Costello"
date: "`r format(Sys.time(), '%B %d, %Y')`"
fontsize: 11pt
bibliography: "../pew-subsidies.bib"

---

# Introduction

The purpose of this script is to assign management scores to all of our vessels, such that they can later be classified into different management regimes. The idea being that fisheries with effective management controls in place are less likely to see as large of changes in fishing effort resulting from subsidy reforms (e.g., Subsidies are likely to be less coupled with effort in well-managed fisheries; effort is instead likely dictated more by management). 

This script outputs a single data file to `r here::here("results", "03-management")`: 
- vessel-list-2018-management.csv

```{r}
# General
library(knitr) 
library(countrycode) # country name matching
library(janitor)
library(tidyverse)

# Path to vessel list
vessel_list_path <- here::here("results", "02-vessel-list", "vessel_list_2018_raw.csv")

# Path to Melnychuk FMI data
melnychuk_2017_mgmt_dat_path <- here::here("data", "edited", "melnychuk_et_al_2017_fmi_tidy.csv" )

# Path to Costello upsides data
costello_2016_upsides_dat_path <- here::here("data", "edited", "costello_et_al_2016_upsides_tidy.csv")

# Make figures?
make_plots <- T

# Few more packages and dataif you're making figures
if(make_plots){
  
  library(ggridges)
  library(ggrepel)
  library(scales)
  library(sf)
  
  # 1) World map ---
  world <- read_sf(here::here("app", "data", "shapefiles", "ne_50m_admin_SubsidyExplorer"), layer = "land_50m")

  # 2) EEZs/FAO regions ---
  eez_fao <- read_sf(dsn = here::here("app", "data", "shapefiles", "eez-v10-fao-combined-simple"), layer="eez_v10_fao_combined_simple") %>%
    mutate(eez_hs_code = ifelse(!is.na(zone), paste0("HS-",zone), as.character(mrgid)))
  
  # 3) Just FAO regions ---
  fao_areas <- read_sf(here::here("data", "shapefiles", "fao"), layer = "World_Fao_Zones")
  
}

# Save data files? 
save_files <- F
```

```{r}
# Code options
knitr::opts_chunk$set(echo = FALSE, error = FALSE, message = FALSE, warning = F)

# Directories for the output files generated by this script
results_dir <- here::here("results", "03-management/")
if (dir.exists(results_dir) == F) {
  dir.create(results_dir, recursive = T)
}

```

# Get EEZ info from vessel list

```{r}
# Load vessel list
vessel_list_raw <- read.csv(vessel_list_path, stringsAsFactors = F)

```

```{r}
# Get eez codes and corresponding territories/sovereigns
eez_info <- vessel_list_raw %>%
  group_by(eez_id) %>%
  distinct(eez_territory1_iso3, eez_territory2_iso3, eez_territory3_iso3, eez_sovereign1_iso3, eez_sovereign2_iso3, eez_sovereign3_iso3) 

# Tidy eez data
eez_info_tidy <- eez_info %>%
  mutate(n_states = case_when((is.na(eez_territory2_iso3) & is.na(eez_territory3_iso3)) ~ 1,
                            (!is.na(eez_territory2_iso3) & is.na(eez_territory3_iso3)) ~ 2,
                            TRUE ~ 3)) %>%
  mutate(eez_territory_iso3 = case_when(n_states == 1 ~ c(eez_territory1_iso3),
                                   n_states == 2 ~ paste(eez_territory1_iso3, eez_territory2_iso3, sep = " "),
                                   n_states == 3 ~ paste(eez_territory1_iso3, eez_territory2_iso3, eez_territory3_iso3, sep = " ")),
         eez_sovereign_iso3 = case_when(n_states == 1 ~ c(eez_sovereign1_iso3),
                                   n_states == 2 ~ paste(eez_sovereign1_iso3, eez_sovereign2_iso3, sep = " "),
                                   n_states == 3 ~ paste(eez_sovereign1_iso3, eez_sovereign2_iso3, eez_sovereign3_iso3, sep = " "))) %>%
  dplyr::select(eez_id, eez_territory_iso3, eez_sovereign_iso3) %>%
  mutate(territory_split = map(eez_territory_iso3, ~ str_split(.x, " ")[[1]]),
         sovereign_split = map(eez_sovereign_iso3, ~ str_split(.x, " ")[[1]])) %>%
  unnest(cols = c(territory_split, sovereign_split)) %>%
  dplyr::select(eez_id, eez_territory_iso3 = territory_split, eez_sovereign_iso3 = sovereign_split)

eez_info_tidy$eez_sovereign_iso3[eez_info_tidy$eez_id == 49003] <- "CHN"
```

# Management Scores from FMI indices

## Assign FMI indices to regions

```{r setup, include=FALSE}
# Load management indicies by state
melnychuk_2017_mgmt_dat <- read_csv(melnychuk_2017_mgmt_dat_path) %>%
  dplyr::filter(variable == "fmi_weighted")

# Load country lookup
country_dependencies <- read_csv(here::here("data", "lookup-tables", "country_dependencies.csv"))

# EU countries
eu_countries <- country_dependencies$iso3[country_dependencies$is_EU]
eu_territories <- country_dependencies$iso3[country_dependencies$sovereign_iso3 %in% eu_countries & country_dependencies$is_overseas_territory]
```

```{r}
# Get mean management indices for the EU, as well as by region 
regional_management_indices <- melnychuk_2017_mgmt_dat %>%
  mutate(region = countrycode(iso3, "iso3c", "region")) %>%
  group_by(region) %>%
  summarize(fmi_weighted_region = mean(value)) %>%
  na.omit()

# Get EU avg management index
eu_management_index <- melnychuk_2017_mgmt_dat %>%
  dplyr::filter(iso3 == "EU")

eu_df <- tibble(iso3 = c(eu_countries, eu_territories), fmi_weighted_eu = eu_management_index$value)

# Get continent averages 
continent_management_indices <- melnychuk_2017_mgmt_dat %>%
  mutate(continent = countrycode(iso3, "iso3c", "continent")) %>%
  group_by(continent) %>%
  summarize(fmi_weighted_continent = mean(value)) %>%
  na.omit()

# Assign management indices to each EEZ
eez_management_indices <- eez_info_tidy %>%
  arrange(eez_sovereign_iso3) %>%
  mutate(eez_territory_region = countrycode(eez_territory_iso3, "iso3c", "region"),
         eez_sovereign_region = countrycode(eez_sovereign_iso3, "iso3c", "region"),
         eez_continent = countrycode(eez_territory_iso3, "iso3c", "continent")) %>%
  left_join(melnychuk_2017_mgmt_dat, by = c("eez_sovereign_iso3" = "iso3")) %>%
  rename(fmi_weighted = value) %>%
  dplyr::select(-variable) %>%
  left_join(eu_df, by = c("eez_territory_iso3" = "iso3")) %>%
  left_join(regional_management_indices, by = c("eez_territory_region" = "region")) %>%
  rename(fmi_weighted_region_territory = fmi_weighted_region) %>%
  left_join(regional_management_indices, by = c("eez_sovereign_region" = "region")) %>%
  rename(fmi_weighted_region_sovereign = fmi_weighted_region) %>%
  left_join(continent_management_indices, by = c("eez_continent" = "continent")) %>%
  mutate(fmi_best = case_when(!is.na(fmi_weighted) ~ fmi_weighted,
                              !is.na(fmi_weighted_eu) ~ fmi_weighted_eu,
                              !is.na(fmi_weighted_region_territory) ~ fmi_weighted_region_territory,
                              !is.na(fmi_weighted_region_sovereign) ~ fmi_weighted_region_sovereign,
                              !is.na(fmi_weighted_continent) ~ fmi_weighted_continent,
                              TRUE ~ min(melnychuk_2017_mgmt_dat$value))) %>%
  group_by(eez_id) %>%
  summarize(fmi_best = mean(fmi_best)) %>% # Necessary for disputed areas/joint regime areas
  ungroup()

```

```{r}
if(make_plots){
  
  # Summarize data
    fmi_map_dat <- eez_fao %>%
      left_join(eez_management_indices %>% mutate(eez_hs_code = as.character(eez_id)), by = c("eez_hs_code")) %>%
      mutate(fmi_best = ifelse(is.na(fmi_best), 0.21, fmi_best))
  
  # Map by EEZ/FAO area
  fmi_map <- ggplot()+
    geom_sf(data = world, fill = "black", color = "grey", size = 0.25)+
    geom_sf(data = fmi_map_dat, aes(fill = fmi_best), color = "grey", size = 0.25)+
    theme_bw()+
    scale_fill_distiller(palette = "YlGn", 
                         direction = 1,
                         limits = c(0.2, 1))+
    labs(fill = "FMI Score")+
    theme(plot.background = element_rect(color = NA),
          panel.border = element_rect(color = NA))
  
  fmi_map
  
  ggsave(paste0(results_dir, "fmi_score_map.png"), width = 7.5, height = 4, units = "in")
  
  # Map by EEZ with the high seas FAO areas removed for PLOS ONE
  hs_color <- RColorBrewer::brewer.pal(9, "YlGn")[1]
  
  fmi_map_no_hs <- ggplot()+
    geom_sf(data = world, fill = "black", color = "grey", size = 0.25)+
    geom_sf(data = fmi_map_dat %>% dplyr::filter(!is.na(mrgid)), aes(fill = fmi_best), color = "grey", size = 0.25)+
    theme_bw()+
    scale_fill_distiller(palette = "YlGn", 
                         direction = 1,
                         limits = c(0.2, 1))+
    labs(fill = "FMI Score")+
    theme(plot.background = element_rect(color = NA),
          panel.border = element_rect(color = NA),
          panel.background = element_rect(fill = hs_color),
          axis.text = element_blank(),
          axis.ticks = element_blank())+
    coord_sf(expand = F)
  
  fmi_map_no_hs
  
  ggsave(paste0(results_dir, "fmi_scores_no_hs_map.png"), width = 7.5, height = 4, units = "in")
}
```

# Median fishing pressure by FAO region

```{r}
# Load costello stock status database
costello_2016_upsides_dat <- read_csv(costello_2016_upsides_dat_path)

# Median F/Fmsy by FAO region
median_fvfmsy_fao <- costello_2016_upsides_dat %>%
  group_by(region_fao_split) %>%
  summarize(fvfmsy_med_region = median(fv_fmsy, na.rm = T),
            fvfmsy_mean_region = mean(fv_fmsy, na.rm = T),
            fvfmsy_w_mean_region = weighted.mean(fv_fmsy, w = msy, na.rm = T)) %>%
  dplyr::select(fao_region = region_fao_split, fvfmsy_med_region, fvfmsy_mean_region, fvfmsy_w_mean_region)
```

```{r}
if(make_plots){
  
  # Summarize data
    fvfmsy_map_dat <- fao_areas %>%
      rename(fao_region = zone) %>%
      left_join(median_fvfmsy_fao, by = "fao_region") %>%
      gather(metric, value, 3:5) %>%
      mutate(metric = str_replace(str_replace(metric, "fvfmsy_", ""), "_region", ""),
             metric_display = case_when(metric == "med" ~ "Median",
                                        metric == "w_mean" ~ "Weighted Mean",
                                        metric == "mean" ~ "Mean"))
  
  # Map by EEZ/FAO area
  fvfmsy_map <- ggplot()+
    geom_sf(data = world, fill = "black", color = "grey", size = 0.25)+
    geom_sf(data = fvfmsy_map_dat, aes(fill = value), color = "grey", size = 0.25)+
    theme_bw()+
    scale_fill_gradient2(low = "blue",
                         high = 'red',
                         mid = "white",
                         midpoint = 1,
                         limits = c(0,5),
                         oob = scales::squish)+
    labs(fill = "F/Fmsy")+
    theme(plot.background = element_rect(color = NA),
          panel.border = element_rect(color = NA))+
    facet_wrap(~metric_display, ncol = 1)
  
  fvfmsy_map
  
  ggsave(paste0(results_dir, "fvfmsy_map.png"), width = 7.5, height = 9.5, units = "in")
  
  # Make PLOS ONE friendly version 
  fao_region_names <- c("77" = "Eastern Central part of the Pacific Ocean", 
                          "71" = "Western Central part of the Pacific Ocean",
                          "51" = "Western part of the Indian Ocean",
                          "61" = "Northwestern part of the Pacific Ocean",
                          "87" = "Southeastern part of the Pacific Ocean",
                          "41" = "Southwestern part of the Atlantic Ocean",
                          "34" = "Eastern Central part of the Atlantic Ocean",
                          "47" = "Southeastern part of the Atlantic Ocean",
                          "27" = "Northeastern part of the Atlantic Ocean",
                          "81" = "Southwestern part of the Pacific Ocean",
                          "21" = "Northwestern part of the Atlantic Ocean",
                          "57" = "Eastern part of the Indian Ocean",
                          "31" = "Western part of the Atlantic Ocean",
                          "67" = "Northeastern part of the Pacific Ocean",
                          "48" = "Antarctic part of the Atlantic Ocean",
                          "88" = "Antarctic part of the Pacific Ocean",
                          "58" = "Antarctic and Southern parts of the Indian Ocean",
                          "18" = "Arctic Ocean",
                          "37" = "Mediterranean and Black Sea")
    
    fao_region_names_df <- tibble(fao_region = as.integer(names(fao_region_names)),
                                  name = fao_region_names)
    
  fvfmsy_plot_dat <- fvfmsy_map_dat %>%
    st_drop_geometry() %>%
    left_join(fao_region_names_df, by = "fao_region") %>%
    mutate(display_name = paste0("Area ", fao_region, ": ", name))
  
  fvfmsy_plot <- ggplot()+
    geom_hline(yintercept = 1)+
    geom_point(data = fvfmsy_plot_dat, aes(x = as.factor(display_name), y = value, fill = value, shape = metric_display), size = 3)+
    theme_bw()+
    scale_x_discrete(limits=rev)+
    coord_flip()+
    scale_fill_gradient2(low = "blue",
                         high = 'red',
                         mid = "white",
                         midpoint = 1,
                         limits = c(0,5),
                         oob = scales::squish)+
    labs(fill = "F/Fmsy", shape = "Value", x = "", y = "F/Fmsy")+
    guides(fill = "none")+
    scale_shape_manual(values = c(21,22,23))+
    theme(plot.background = element_rect(color = NA),
          #panel.border = element_rect(color = NA),
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          legend.position = "bottom")
  
    ggsave(paste0(results_dir, "fvfmsy_plot.png"), width = 7.5, height = 5, units = "in")

}
```

<!-- # Proportion of formally assessed stocks by FAO region -->

<!-- # ```{r} -->
<!-- # prop_assessed_stocks <- costello_2016_upsides_dat %>% -->
<!-- #   group_by(region_fao_split) %>% -->
<!-- #   summarize(catch_ram = sum(catch[dbase == "RAM"]), -->
<!-- #             catch_other = sum(catch[dbase != "RAM"])) %>% -->
<!-- #   mutate(catch_prop_ram = catch_ram/(catch_ram + catch_other)) %>% -->
<!-- #   dplyr::select(fao_region = region_fao_split, catch_prop_ram) -->
<!-- # ``` -->

<!-- # Proportion of catch share stocks by FAO region -->

<!-- # ```{r} -->
<!-- # prop_catch_share_stocks <- costello_2016_upsides_dat %>% -->
<!-- #   group_by(region_fao_split) %>% -->
<!-- #   summarize(catch_share= sum(catch[catch_share == 1]), -->
<!-- #             catch_other = sum(catch[catch_share != 1])) %>% -->
<!-- #   mutate(catch_prop_cs= catch_share/(catch_share + catch_other)) %>% -->
<!-- #   dplyr::select(fao_region = region_fao_split, catch_prop_cs) -->
<!-- # ``` -->

# Apply management indices to vessels

```{r}
# Apply management scores to vessel list
vessel_list_manage <- vessel_list_raw %>%
  left_join(eez_management_indices, by = c("eez_id")) %>%
  left_join(median_fvfmsy_fao, by = c("fao_region"))

# FAO region 18 is missing stock data, but has some effort
vessel_list_manage$fvfmsy_med_region[is.na(vessel_list_manage$fvfmsy_med_region)] <- 1
vessel_list_manage$fvfmsy_mean_region[is.na(vessel_list_manage$fvfmsy_mean_region)] <- 1
vessel_list_manage$fvfmsy_w_mean_region[is.na(vessel_list_manage$fvfmsy_w_mean_region)] <- 1

if(save_files){
write_csv(vessel_list_manage, paste0(results_dir, "vessel_list_2018_management.csv"))
}
```


```{r}
# Calculate management quantiles for each indicator
fmi_quantile <- tibble("indicator" = "fmi") %>%
  bind_cols(setNames(as.data.frame(matrix(quantile(vessel_list_manage$fmi_best), nrow = 1)), names(quantile(vessel_list_manage$fmi_best))))

fvfmsy_med_quantile <- tibble("indicator" = "fvfmsy_median") %>%
  bind_cols(setNames(as.data.frame(matrix(quantile(vessel_list_manage$fvfmsy_med_region), nrow = 1)), names(quantile(vessel_list_manage$fvfmsy_med_region))))

fvfmsy_mean_quantile <- tibble("indicator" = "fvfmsy_mean") %>%
  bind_cols(setNames(as.data.frame(matrix(quantile(vessel_list_manage$fvfmsy_mean_region), nrow = 1)), names(quantile(vessel_list_manage$fvfmsy_mean_region))))

fvfmsy_w_mean_quantile <- tibble("indicator" = "fvfmsy_w_mean") %>%
  bind_cols(setNames(as.data.frame(matrix(quantile(vessel_list_manage$fvfmsy_w_mean_region), nrow = 1)), names(quantile(vessel_list_manage$fvfmsy_w_mean_region))))

indicator_quantiles <- fmi_quantile %>%
  bind_rows(fvfmsy_med_quantile) %>%
  bind_rows(fvfmsy_mean_quantile) %>%
  bind_rows(fvfmsy_w_mean_quantile)

write_csv(indicator_quantiles, paste0(results_dir, "management_indicator_quantiles.csv"))
```


```{r}
# if(make_plots){
#   
#   # Summarize data
#   density_dat <- vessel_list_manage %>%
#     dplyr::select(fmi_best, fvfmsy_med_region, fvfmsy_mean_region, fvfmsy_w_mean_region)
#   
#   probs <- c(0, 1/3, 2/3, 1)
#   
#   # Make FMI density plot (across all vessels)
#   fmi_density <- density(density_dat$fmi_best)
#   
#   fmi_density_plot_dat <- data.frame(name = "FMI Score", x = fmi_density$x, y = fmi_density$y)
#  
#   fmi_quantiles <- quantile(fmi_density_plot_dat$y, prob=probs)
#   fmi_density_plot_dat$quant <- factor(findInterval(fmi_density_plot_dat$x, fmi_quantiles))
#   
#   fmi_density_plot <- ggplot(fmi_density_plot_dat, aes(x,y)) + 
#     geom_line() + 
#     geom_ribbon(aes(ymin=0, ymax=y, fill=quant)) + 
#     scale_y_continuous(expand = c(0,0)) + 
#     scale_fill_manual(labels = c("0-33%", "33-66%", "66-100%"), values = brewer.pal(3, "Blues"), name = "Quantile")+
#     labs(x = "FMI Score", y = "Kernal Density")+
#     theme_bw()
# 
#   # Make median F/Fmsy density plot (across all vessels)
#   med_density <- density(density_dat$fvfmsy_med_region)
# 
#   med_density_plot_dat <- data.frame(name = "Median F/Fmsy", x = med_density$x, y = med_density$y)
#   
#   med_quantiles <- quantile(med_density_plot_dat$y, prob=probs)
#   med_density_plot_dat$quant <- factor(findInterval(med_density_plot_dat$x, med_quantiles))
#   
#   med_density_plot <- ggplot(med_density_plot_dat, aes(x,y)) + 
#     geom_line() + 
#     geom_ribbon(aes(ymin=0, ymax=y, fill=quant)) + 
#     scale_y_continuous(expand = c(0,0)) + 
#     scale_fill_manual(labels = c("66-100%"), values = brewer.pal(3, "Blues")[3], name = "Quantile")+
#     labs(x = "Median F/Fmsy", y = "Kernal Density")+
#     theme_bw()
#   
#   # Make mean F/Fmsy density plot (across all vessels)
#   mean_density <- density(density_dat$fvfmsy_mean_region)
# 
#   mean_density_plot_dat <- data.frame(name = "Mean F/Fmsy", x = mean_density$x, y = mean_density$y)
#   
#   mean_quantiles <- quantile(mean_density_plot_dat$y, prob=probs)
#   mean_density_plot_dat$quant <- factor(findInterval(mean_density_plot_dat$x, mean_quantiles))
#   
#   mean_density_plot <- ggplot(mean_density_plot_dat, aes(x,y)) + 
#     geom_line() + 
#     geom_ribbon(aes(ymin=0, ymax=y, fill=quant)) + 
#     scale_y_continuous(expand = c(0,0)) + 
#     scale_fill_manual(labels = c("66-100%"), values = brewer.pal(3, "Blues")[3], name = "Quantile")+
#     labs(x = "Mean F/Fmsy", y = "Kernal Density")+
#     theme_bw()
#   
#   # Make weighted mean F/Fmsy density plot (across all vessels)
#   w_mean_density <- density(density_dat$fvfmsy_w_mean_region)
# 
#   w_mean_density_plot_dat <- data.frame(name = "Weight. Mean F/Fmsy", x = w_mean_density$x, y = w_mean_density$y)
#   
#   w_mean_quantiles <- quantile(w_mean_density_plot_dat$y, prob=probs)
#   w_mean_density_plot_dat$quant <- factor(findInterval(w_mean_density_plot_dat$x, w_mean_quantiles))
#   
#   w_mean_density_plot <- ggplot(w_mean_density_plot_dat, aes(x,y)) + 
#     geom_line() + 
#     geom_ribbon(aes(ymin=0, ymax=y, fill=quant)) + 
#     scale_y_continuous(expand = c(0,0)) + 
#     scale_fill_manual(labels = c("66-100%"), values = brewer.pal(3, "Blues")[3], name = "Quantile")+
#     labs(x = "Weight. Mean F/Fmsy", y = "Kernal Density")+
#     theme_bw()
#     
# }
```
