==> scripts/04_model_parameterization.Rmd <==
---
output:
  html_document: default

title: "SubsidyExplorer - Step 3: Model Parameterization"
author: "Kat Millage, Vienna Saccomanno, Laura Lea Rubino, Christopher Costello"
date: "`r format(Sys.time(), '%B %d, %Y')`"
fontsize: 11pt
bibliography: "../pew-subsidies.bib"

---

# Introduction 

This script handles the preliminary data wrangling necessary to create the biological parameters for the bio-economic model underlying the SubsidyExplorer toolkit. It allows for specification of different regional configurations, though we have selected the one for which we believe the data is most supportive. 

```{r}
# Load packages
library(knitr) 
library(countrycode) # country name matching
library(janitor) # cleaning column names
library(readxl) # Load XLSX files
library(tidyverse)

# Path to Costello upsides data
costello_2016_upsides_dat_path <- here::here("data", "edited", "costello_et_al_2016_upsides_tidy.csv")

# Path to vessel list
vessel_list_path <- here::here("results", "03-management", "vessel_list_2018_management.csv")

# Code options
knitr::opts_chunk$set(echo = FALSE, error = FALSE, message = FALSE, warning = F)

# Directories for the output files generated by this script
results_dir <- here::here("results", "04-model-parameterization/")
if (dir.exists(results_dir) == F) {
  dir.create(results_dir, recursive = T)
}

app_dir <- here::here("app", "data/")
if (dir.exists(app_dir) == F) {
  dir.create(app_dir, recursive = T)
}

# Select number of regions (choices are 3 or 6)
n_regions <- 3

```

In addition to the data sources described in the inital data wrangling script, we also draw on the folloing sources: 

- Bioeconomic model parameters from both of the World Bank's Sunken Billions Reports (2009 and 2017). 
> World Bank, & FAO. (2009). The Sunken Billions: The Economic Justification for Fisheries Reform. Retrieved from https://siteresources.worldbank.org/EXTARD/Resources/336681-1224775570533/SunkenBillionsFinal.pdf
> World Bank. (2017). The Sunken Billions Revisited: Progress and Challenges in Global Marine Fisheries. Retrieved from https://openknowledge.worldbank.org/handle/10986/24056

```{r}
if(n_regions == 3){
  
  # Define regions 
  region_names <- sort(c("atlantic", "indian", "pacific"))
  
  # Create a tidy df with the cooresponding fao regions
  regions_tidy <- tibble(
    region = region_names,
    fao_region = list(c(21, 27, 31, 37, 34, 41, 47, 48), # atlantic (and med)
                      c(51, 57, 58), # indian
                      c(18, 61, 67, 71, 77, 81, 87, 88)) # pacific 
  ) %>%
    unnest(cols = c(fao_region))
  
  # Also create a blank tibble with regions as column names
  regions_spread <- setNames(data.frame(matrix(ncol = length(region_names), nrow = 0)), as.character(region_names))
  
  # Now make empty output tables for both our global parameters, and for our regional parameters
  global_model_parameters <- tibble(parameter = character(0), value = numeric(0), source = character(0))
  
  regional_model_parameters <- tibble(parameter = character(0), source = character(0)) %>%
    bind_cols(regions_spread) %>%
    mutate(atlantic = as.character(atlantic),
           indian = as.character(indian),
           pacific = as.character(pacific))
}
```


```{r setup, include=FALSE}
if(n_regions == 6){
  
# Define regions 
region_names <- sort(c("atlantic_e", "atlantic_w", "indian", "pacific_e", "pacific_w", "polar")) # ran into difficulties with the polar region. 6 regions needs more work

# Create a tidy df with the cooresponding fao regions
regions_tidy <- tibble(
  region = region_names,
  fao_region = list(c(27, 37, 34, 47), # atlantic east + med
                    c(21, 31, 41), # atlantic west
                    c(51, 57), # indian
                    c(67, 77, 87), # pacific e
                    c(61, 71, 81),# pacific w
                    c(18, 48, 58, 88)) # polar
) %>%
  unnest(cols = c(fao_region))

# Also create a blank tibble with regions as column names
regions_spread <- setNames(data.frame(matrix(ncol = length(region_names), nrow = 0)), region_names)

# Now make empty output tables for both our global parameters, and for our regional parameters
global_model_parameters <- tibble(parameter = character(0), value = numeric(0), source = character(0))

regional_model_parameters <- tibble(parameter = character(0), source = character(0)) %>%
  bind_cols(regions_spread)

}
```

```{r}
# Load data sets
# Load costello data
costello_2016_dat <- read_csv(costello_2016_upsides_dat_path)

# Load vessel list
vessel_list_mgmt <- read.csv(vessel_list_path, stringsAsFactors = F)

# Assign regions to vessels
vessel_list_out <- vessel_list_mgmt %>%
  left_join(regions_tidy, by = c("fao_region" = "fao_region"))

# Save updated vessel list
write_csv(vessel_list_out, paste0(results_dir, "vessel_list_2018_regions.csv"))
```


# Step 1: Biomass growth function

For our biomass growth function, $G(b)$, we use a form of the basic logistic surplus production model commonly referred to as the Pella-Tomlinson model. With discrete units of time (hereafter years for the purposes of our model), biomass in the next year, $b_{t+1}$, is described by: 

$$ b_{t+1} = b_{t} + \frac{\phi + 1}{\phi} g b_{t}\Bigg(1 - \bigg(\frac{b_{t}}{K}\bigg)^{\phi}\Bigg) - h_{t}$$
where $\phi$ is a scaling parameter, $g$ is the population growth rate, $K$ is the carrying capacity (maximum population size for growth to be positive), and $h_t$ is the total harvest in that year across all fleets. 

The intrinsic growth rate of the global stock ($r$) is given by: 
$$r = \frac{\phi + 1}{\phi}g$$

Certain attributes of the Pella-Tomlinson biomass growth function relate its parameters to stock characteristics. For instance, $g$ is equal to

$$ g = \bigg(\frac{MSY}{K} \bigg) (\phi + 1)^{\frac{1}{\phi}} $$
where $MSY$ is the maximum sustinable yield of the stock, and $K$ is the carrying capacity of the stock. 

Maximum sustainable yield is therefore equal to: 
$$MSY = \frac{gK}{(\phi + 1)^{\frac{1}{\phi}}} $$
and the biomass cooresponding to MSY as: 
$$B_{MSY} = \frac{K}{(\phi + 1)^{\frac{1}{\phi}}}$$

$B_{MSY}/K$ is given by: 
$$\frac{B_{MSY}}{K} = \frac{1}{(\phi + 1)^{\frac{1}{\phi}}}$$
where a value of $\phi = 1$ makes this equation equal to the Schaefer model (production function is symmetrical). 

We therefore need to assign the following parameters for the biomass growth function: 
- $\phi$ : Scaler parameter (no units)
- $MSY$ : Global maximum sustainable yield (tons)
- $K$ : Global biomass carrying capacity (tons)
- $r$ : Global biomass population growth rate (no units)
- $h_0$: Total harvest in the base year (tons)
- $b_0$ : Global biomass in the base year (tons)

### Alternate forms of the Pella-Tomlinson model

### Discrete

The Sunken Billions reports also used a form of the Pella-Tomlinson model, though they defined it differently. The form of the model used there may be written as:

$$ G(X) = r X - \beta X^{\gamma} $$
where $X$ is biomass, $\gamma$ is the Pella-Tomlinson exponent that defines the skewness of the biomass growth function, $r$ is the intrinsic growth rate, and $\beta$ is a parameter. 

In this form, values of $\gamma < 2$ skew the function to the left, values of $\gamma > 2$ skew the function to the right, and $\gamma = 2$ reproduces the Lotka-Volterra logistic function (Schaefer) model (symmetric). As $\gamma$ gets closer to 1, it converges on the Fox biomass growth function. 

$X_{MSY}/K \in [0.368,1]$ for $\gamma$ values ranging from unity to infinity. In general, there are biological reasons to expect this to occur at a value less than 0.5 cooreponding to a $\gamma < 2$. This ratio can be caclulated as: 

$$\frac{X_{MSY}}{K} = \gamma^{\frac{1}{1-\gamma}}$$

With this form of the model, carrying capacity ($K$) is given by:
$$K = \bigg(\frac{r}{\beta}\bigg)^{\frac{1}{\gamma-1}} $$
Maximum sustainable yield can then be calculated as:
$$ MSY = r \bigg(\frac{\gamma-1}{\gamma}\bigg)\bigg(\frac{r}{\gamma\beta}\bigg)^{\frac{1}{\gamma-1}} $$
and the biomass cooresponding to MSY as: 
$$ X_{MSY} = \bigg(\frac{r}{\gamma\beta}\bigg)^{\frac{1}{\gamma-1}}$$

## Global biological parameters

We first parameterize our bio-economic model of fisheries subsidies reform globally. 

We assume $\phi = 0.188$ based on literature for all (Thorson et al. 2012; World Bank 2017; Costello et al. 2016). 

> Thorson, J. T., Cope, J. M., Branch, T. A., & Jensen, O. P. (2012). Spawning biomass reference points for exploited marine fishes, incorporating taxonomic and body size information. Canadian Journal of Fisheries and Aquatic Sciences, 69(9), 1556–1568. https://doi.org/10.1139/f2012-077


```{r}
# Phi
phi <- 0.188

# Add phi to parameter table
global_model_parameters <- global_model_parameters %>%
  bind_rows(tibble(parameter = "phi", value = phi, source = "Thorson et al. 2012"))
```

We then need to calculate $r$, for which we need both $MSY$ and $K$. 

Many different estimates of the global marine $MSY$ exist, generally ranging between 80 and 115 million tons. Different sources have used different estimates: 
- The initial Sunken Billions report (World Bank and FAO 2009): 95 million tons
- Sumaila et al. 2012: 83 - 99 million tons
- Costello et al. 2012: 85 - 110 million tons
- The second Sunken Billions report (World Bank 2017): 102 million tons

For this analyis, we use the database of stock statuses for more than 4,000 fisheries (representing ~78% of global fisheries) created by Costello et al. (2016). MSY for our model is defined as the sum of the MSY of all stocks contained in this dataset, adjusted for the fraction of global biomass estimated to be represented by this data. 

```{r}
# Calculate global MSY
global_msy <- sum(costello_2016_dat$msy)/0.78

# Add MSY to parameter table
global_model_parameters <- global_model_parameters %>%
  bind_rows(tibble(parameter = "msy", value = global_msy, source = "Costello et al. 2016"))

```

We therefore assume that the global MSY is therefore equal to `r format(global_msy, big.mark = ",")` tons, which is comparable to the estimates used by other studies. 

Very little is known about the global carrying capacity of fish stocks, though many studies have suggested that it is between 5 and 15 times the MSY for different species of fish. The estimate of global carrying capacity obtained by summing the respective carrying capacities all all stocks from Costello et al. (2016) is higher than this (22 times MSY). We therefore instead assume global carrying capacity to be equal to 9.1 times MSY as found in the second Sunken Billions report.

```{r}
# One estimate of global carrying capacity
global_k <- global_msy*9.1

# Add K to parameter table
global_model_parameters <- global_model_parameters %>%
  bind_rows(tibble(parameter = "k", value = global_k, source = "Costello et al. 2016; World Bank 2017"))
```

We can then use all three parameters to estimate the global biomass intrinsic growth rate using the equation above.

```{r}
# Calculate population growth rate
global_g <- (global_msy/global_k)*(phi+1)^(1/phi) 

# Add g to parameter table
global_model_parameters <- global_model_parameters %>%
  bind_rows(tibble(parameter = "g", value = global_g, source = "Calculated"))

# Calculate intrinsic growth rate
global_r <- ((phi+1)/phi)*global_g

# Add g to parameter table
global_model_parameters <- global_model_parameters %>%
  bind_rows(tibble(parameter = "r", value = global_r, source = "Calculated"))
```

A great deal of uncertainty exists about the total amount of biomass currently in our oceans. We therefore estimate total biomass in the base year (2018) by assuming an annual rate of biomass growth of 8 million mt/year (World Bank 2017). We then solve for the biomass in 2018 that would have yielded the observed harvest after accounting for the assummed rate of biomass growth ($\xi = 8000000$)  $b_{t+1} = b_t + \xi$

$$ h_0+ \xi =\frac{\phi + 1}{\phi} g b_{0}\Bigg(1 - \bigg(\frac{b_{0}}{K}\bigg)^{\phi}\Bigg)$$

```{r}
# Calculate global total harvest
global_h0 <- sum(vessel_list_out$catch, na.rm = T)

# Add base year harvest to parameter table
global_model_parameters <- global_model_parameters %>%
  bind_rows(tibble(parameter = "h0", value = global_h0, source = "FAO 2020"))
```


```{r}
# Estimation of starting biomass
biomass_vec <- seq(0, 0.75*global_k, length.out = 1000000)
growth_approx <- (((phi+1)/phi)*global_g*biomass_vec*(1-((biomass_vec/global_k)^phi)))
biomass_pos <- which.min(abs(growth_approx - (global_h0+8000000)))
global_b0 <- biomass_vec[biomass_pos] 

# Add base year biomass to parameter table
global_model_parameters <- global_model_parameters %>%
  bind_rows(tibble(parameter = "b0", value = global_b0, source = "FAO 2019; World Bank 2017"))
```

## Regional biological parameters

We also apply our bio-economic model of fisheries subsidies reform on a regional basis. 

This regional breakdown was chosen with the goal of trying to reflect commonly used delimitations of the boundaries of the world's largest fish stocks. Regional characterizations used by other studies (see World Bank 2017) are often organized around continents and political boundaries, which is inherently problematic for fisheries as most of the world's major fish stocks straddle these boundaries. While our regional model has tried to overcome some of these problems by aggregating ocean areas with more similar physical characteristics, we aknowledge that our results by region are likely to be less dependable than our global results. 

For the regional portion of our analysis, we needed to calculate all of the same model inputs used in the global model for each of our six regions. 

First, we assume that $\phi$ remains constant across regions ($\phi = 1.188$, Thorson et al. 2012).

```{r}
# Apply phi regionally
regional_phi <- rep(phi, length = length(region_names))
names(regional_phi) <- region_names

# Add phi to the regional paramter table
regional_model_parameters <- regional_model_parameters %>%
  bind_rows(c("parameter" = "phi", "source" = "Thorson et al. 2012", regional_phi))
```

Then, regional values of MSY were estimated in the same way as for the global fishery. We summed the MSYs of all stocks in the FAO statistical areas comprising each region from Costello et al. 2016. In cases where stocks in this dataset were associated with multiple FAO statistical areas, we assume that the stock's MSY and carrying capacity is split evenly between areas. 

```{r}
# Calculate stats by region
costello_2016_dat_regional <- costello_2016_dat %>%
  dplyr::filter(region_fao_split != "") %>%
  mutate(msy_new = msy/n_regions,
         k_new = k/n_regions,
         catch_new = catch/n_regions) %>%
  left_join(regions_tidy, by = c("region_fao_split" = "fao_region")) %>%
  group_by(region) %>%
  summarize(msy = sum(msy_new, na.rm = T),
            k = sum(k_new, na.rm = T),
            catch = sum(catch_new, na.rm = T),
            price = median(price, na.rm = T)) %>%
  arrange(region)

# Extract msy
regional_msy <- round(costello_2016_dat_regional$msy/0.78)
names(regional_msy) <- region_names

# Add MSY to the regional paramter table
regional_model_parameters <- regional_model_parameters %>%
  bind_rows(c("parameter" = "msy", "source" = "Costello et al. 2016", regional_msy))
```

We estimate regional carrying capacities again by assuming K to be 9.1 times MSY. 

```{r}
# Extract carrying capacity
regional_k <- regional_msy*9.1

# Add MSY to the regional paramter table
regional_model_parameters <- regional_model_parameters %>%
  bind_rows(c("parameter" = "k", "source" = "Costello et al. 2016; World Bank 2017", regional_k))
```

We then calculate regional growth rates using those values of MSY and K. 

```{r}
# Calculate r regionally
regional_g <- (regional_msy/regional_k)*(regional_phi+1)^(1/regional_phi) 

# Add r to parameter table
regional_model_parameters <- regional_model_parameters %>%
  bind_rows(c("parameter" = "g", "source" = "Calculated", regional_g))

# Calculate r regionally
regional_r <- ((regional_phi+1)/(regional_phi))*regional_g

# Add r to parameter table
regional_model_parameters <- regional_model_parameters %>%
  bind_rows(c("parameter" = "r", "source" = "Calculated", regional_r))
```

```{r}
# Calculate global total harvest
regional_h0 <- (vessel_list_out %>%
  group_by(region) %>%
  summarize(catch = round(sum(catch, na.rm = T))))$catch
names(regional_h0) <- region_names

# Add base year harvest to parameter table
regional_model_parameters <- regional_model_parameters %>%
  bind_rows(c("parameter" = "h0", "source" = "FAO 2019", regional_h0))
```

We assume biomass growth to be equally divided among regions. 

```{r}
#Estimate starting biomass
regional_b0 <- numeric(length(region_names))
regional_biomass_growth <- c(2.66e6, 2.66e6, 2.66e6)

for(i in 1:length(regional_h0)){
  biomass_vec <- seq(0, 0.75*regional_k[i], length.out = 10000)
  growth_approx <- (((regional_phi[i]+1)/regional_phi[i])*regional_g[i]*biomass_vec*(1-((biomass_vec/regional_k[i])^regional_phi[i])))
  biomass_pos <- which.min(abs(growth_approx - (regional_h0 + regional_biomass_growth[i])))
  regional_b0[i] <- biomass_vec[biomass_pos]
}

names(regional_b0) <- region_names

# Add base year biomass to parameter table
regional_model_parameters <- regional_model_parameters %>%
  bind_rows(c("parameter" = "b0", "source" = "FAO 2019; World Bank 2017", regional_b0))
```


<!-- xxxxxxxxxxxxxxxxxxxx -->

# Step 2: Harvest function

Our harvest function, $Y(b,e)$, is really the sum of the individual harvest functions for our three fleet segments (managed, intermediate, and open-access). Harvest for each fleet in a given year, $h_{j,t}$, is represented by: 

$$ h_{j,t} = q_{j} b_{t} e_{j,t} $$

where $q_j$ is a fleet specific catchability parameter, and $e_{j,t}$ is the fishing effort expended by that fleet that year. We estimate catchability directly for each fleet based on base year data by calculating $q_j$ as follows:

$$ q_{j} = \frac{h_{j,0}}{b_{0}*e_{j,0}} $$

The total harvest is therefore the sum of harvests across all fleets: 

$$ h_{t} = \sum_{i=1}^{j} h_{j,t} $$

The following parameters are calculuated prior to every model run as they vary depending on the relative sizes of the fleet:
- $h_{j,0}$ : Base year harvest for each fleet (tons)
- $e_{j,0}$ : Base year effort for each fleet (kWh)
- $q_{j}$ : Catchability for each fleet (tons)

## Global harvest parameters

Though we don't calculate the fleet specific parameters here, we have already calculated total harvest in the base year, and can calculate total fishing effort. This data comes from GFW data (2018). Effort is given in units of kWh, which is the product of the hours spent engaged in fishing activity, and the engine capacity of the vessel in kW

```{r}
# Group by fleet and calculate base year harvest and effort for the global fleet
global_e0 <- sum(vessel_list_out$fishing_KWh_eez_fao_ter, na.rm = T)

# Add base year effort to parameter table
global_model_parameters <- global_model_parameters %>%
  bind_rows(tibble(parameter = "e0", value = global_e0, source = "GFW 2019"))
```

## Regional harvest paramters

We do the same process to calculate base year total effort by region

```{r}
# Group by fleet and calculate base year harvest and effort for the global fleet
regional_fleet <- vessel_list_out %>%
  group_by(region) %>%
  summarize(vessels = n_distinct(ssvid),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            bad_subs = sum(bad_subs, na.rm = T),
            catch = sum(catch, na.rm = T)) %>%
  arrange(region)

# Get effort
regional_e0 <- regional_fleet$fishing_KWh
names(regional_e0) <- region_names

# Add base year effort to parameter table
regional_model_parameters <- regional_model_parameters %>%
  bind_rows(c("parameter" = "e0", "source" = "GFW 2019", regional_e0))
```

<!-- xxxxxxxxxxxxxxxxxxxx -->

# Step 3: Global fish demand

In order to allow the price of fish to change as a function of harvest, we introduce a demand function, $D(h)$, for fish, and we assume it is downward-sloping. Total harvest in a given year, $h_{t}$, gives rise to fish price, $p_t$, through the constant elasticity of demand function: 

$$ p_{t} = \bigg(\frac{1}{\delta}\bigg)^{1/\epsilon} h_{t}^{1/\epsilon} $$

where $\epsilon$ is the constant elasticity of demand, and $\delta$ is a constant. We calculate the parameter $\delta$ using data as follows: 

$$ \delta = \frac{h_{0}}{(p_{0})^{\epsilon}}$$

We therefore need to estimate the following parameters for the demand function: 
- $\epsilon$ : price flexibility/constant elasticity of demand
- $p_0$: global price in the base year (2018)

## Global demand parameters

We assume that the price flexibility/elasticity ($\epsilon$) is equal to -1.15 based on Costello et al. (2016). 

```{r}
# Global price flexibility
global_epsilon <- -1.15

# Add base year biomass to parameter table
global_model_parameters <- global_model_parameters %>%
  bind_rows(tibble(parameter = "epsilon", value = global_epsilon, source = "Costello et al. 2016"))
```

## Regional demand parameters

For the regional analysis, we assume that the price flexibility/elasticity ($\epsilon$) is equal to -1.15 for all regions based on Costello et al. (2016). 

```{r}
# Regional price flexibility
regional_epsilon <- rep(-1.15, length = length(region_names))
names(regional_epsilon) <- region_names

# Add epsilon to parameter table
regional_model_parameters <- regional_model_parameters %>%
  bind_rows(c("parameter" = "epsilon", "source" = "Costello et al. 2016", regional_epsilon))
```

<!-- xxxxxxxxxxxxxxxxxxxx -->

# Step 4: Costs and Benefits

Our cost function, $C(e)$, assumes that each fleet's total costs ($c_{j,t}$) are a function of fishing effort in that time step ($e_{j,t}$) and a fleet-specific cost coefficient ($\alpha_{j}$):
$$ c_{j,t} = \alpha_{j}  e_{j,t}^{\beta}$$

where $\beta$ is a scalar cost parameter that determines the shape of the cost curve (i.e. how non-linear costs are). The cost coefficient $\alpha_{j}$ for each fleet is a function of how much each unit of effort is being subsidized for that fleet $s_{j,t}$. We estimate the costs for each fleet directly based on base year data by calculating $\alpha_{j}$ as follows: 
$$ \alpha_{j} = \frac{p_{0}h_{j,0} + s_{j,0} e_{j,0}}{(e_{j,0})^{\beta}}$$

Fisheries profits are then calculated as revenue less costs plus subsidies for each fleet, where revenue is the fish price $p_t$ times harvest $h_{j,t}$ and costs are calculated as above. Profits for each fleet in each year, $\pi_{j,t}$, are therefore equal to:

$$ \pi_{j,t} = p_{t} h_{j,t} - c_{j,t} + s_{j,t}e_{j,t} $$

where $s_{j,t}$ is the rate of subsidization (i.e. how much each unit of effort is being subsidized) for fleet $j$. 

Effort in the next year, $e_{j,t+1}$, adjusts in response to profits differently for each of our two fleets: 

1) For the managed fleet, we assume that effort is decoupled from costs, and therefore remains constant regardless of changes to profits: 

$$ e_{M,t+1} = e_{M,t}$$

2) For the open-access fleet, we assume that effort is coupled to costs, and therefore adjusts in response to profits:

$$ e_{OA,t+1} = \eta \pi_{OA,t} + e_{OA,t}$$
where $\eta$ is a parameter that regulates the speed at which effort is able to enter and exit the fishery.

We therefore need to estimate the following parameters for the cost and profit functions:
- $\beta$ : scalar cost parameter
- $p_{0}$ : base year fish price ($/ton)
- $\eta$ : speed at which effort is able to enter/exis a fishery

The following parameters are estimated prior to the start of each model run as they depend on the relative sizes of the fleets:
- $s_{j,0}$ : base year rate of subsidization for each fleet ($/kWh)
- $\alpha_{j}$ : cost coefficient for each fleet

## Global cost parameters

For our global analysis, we assume that $\beta$ is equal to 1.3 (Costello et al. 2016). 

```{r}
# Define global beta
global_beta <- 1.3

# Add to parameter table
global_model_parameters <- global_model_parameters %>%
  bind_rows(tibble(parameter = "beta", value = global_beta, source = "Costello et al. 2016"))
```

In order to calculate $\alpha$, we first need to determine fish price in the base year. We compare two alternatives - 1) using the median price from Costello et al. (2016) adjusted to 2018 US\$, and 2) using the global price from the latest Sunken Billions report adjusted to 2018 US\$. The first price looks a bit high (comparable global price estimates are in the range of $1200-1400 per ton), so we use the second 

```{r}
CPI_2012 <- 229.594  #BLS CPI All Items (Annual Ave), (BLS Series ID: CUUR0000SA0)
CPI_2018 <- 251.107  #BLS CPI All Items (Annual Ave), (BLS Series ID: CUUR0000SA0)

# Global price constant
global_price_wrangling <- costello_2016_dat %>%
  group_by(year) %>%
  summarize(price = median(price, na.rm = T))

global_base_price1 <- global_price_wrangling$price*(CPI_2018/CPI_2012)
global_base_price2 <- 1260*(CPI_2018/CPI_2012)

# Add to parameter table
global_model_parameters <- global_model_parameters %>%
  bind_rows(tibble(parameter = "p0", value = global_base_price2, source = "World Bank 2017"))
```

```{r}
global_delta <- global_h0/(global_base_price2^global_epsilon)

# Add price constant to parameter table
global_model_parameters <- global_model_parameters %>%
  bind_rows(tibble(parameter = "delta", value = global_delta, source = "Costello et al. 2016"))
```

Finally, we assume that $\eta$ is equal to 0.1 based on literature (Costello et al. 2016). 

```{r}
# Define global eta
global_eta <- 0.1

# Add to parameter table
global_model_parameters <- global_model_parameters %>%
  bind_rows(tibble(parameter = "eta", value = global_eta, source = "Calculated"))
```

We also define an omega which is equal to 0.001. 

```{r}
# Define global eta
global_omega <- 0.001

# Add to parameter table
global_model_parameters <- global_model_parameters %>%
  bind_rows(tibble(parameter = "omega", value = global_omega, source = "Calculated"))

```

## Regional cost parameters

As with the global analysis, we assume $\beta$ to be equal to 1.3 across all regions. 
```{r}
# Define global beta
regional_beta <- rep(1.3, length = length(region_names))
names(regional_beta) <- region_names

# Add to parameter table
regional_model_parameters <- regional_model_parameters %>%
  bind_rows(c("parameter" = "beta", "source" = "Costello et al. 2016", regional_beta))
```

For simplicity, we assume the starting price to be equal across all regions. 

```{r}
# Calculate base price regionally
regional_base_price <- rep(global_base_price2, length = length(region_names))
names(regional_base_price) <- region_names

# Add catchability to parameter table
regional_model_parameters <- regional_model_parameters %>%
  bind_rows(c("parameter" = "p0", "source" = "World Bank 2017", regional_base_price))

```

```{r}
# Regional price constant
regional_delta <- regional_h0/(regional_base_price^regional_epsilon)

# Add price constant to parameter table
regional_model_parameters <- regional_model_parameters %>%
  bind_rows(c("parameter" = "delta", "source" = "Costello et al. 2016", regional_delta))
```

And we again assume $\eta$ to be equal to 0.1 and $\omega$ to be equal to 0.001 across all regions

```{r}
regional_eta <- rep(0.1, length = length(region_names))
names(regional_eta) <- region_names

# Add subsidy base to parameter table
regional_model_parameters <- regional_model_parameters %>%
  bind_rows(c("parameter" = "eta", "source" = "Costello et al. 2016", regional_eta))
```

```{r}
regional_omega <- rep(0.001, length = length(region_names))
names(regional_omega) <- region_names

# Add subsidy base to parameter table
regional_model_parameters <- regional_model_parameters %>%
  bind_rows(c("parameter" = "omega", "source" = "Costello et al. 2016", regional_omega))
```

```{r}
# Save both parameter files
write_csv(global_model_parameters, paste0(results_dir, "model_parameters_global.csv"))
write_csv(global_model_parameters, paste0(app_dir, "model_parameters_global.csv"))

regional_name <- paste0("model_parameters_regional_", n_regions, "_regions.csv")

write_csv(regional_model_parameters, paste0(results_dir, regional_name))
write_csv(regional_model_parameters, paste0(app_dir, regional_name))
```

<!-- xxxxxxxxxxxxxxxxxxxx -->
