==> scripts/05_subsidy_reform_definitions.Rmd <==
---
output:
  html_document: default

title: "SubsidyExplorer - Step 5: Apply proposal 'definitions' to vessel list"
author: "Kat Millage, Vienna Saccomanno, Laura Lea Rubino, Christopher Costello"
date: "`r format(Sys.time(), '%B %d, %Y')`"
fontsize: 11pt
bibliography: "../pew-subsidies.bib"

---

# Introduction

The purpose of this script is to apply "definitions" to our global vessel list in order to make it easier to identify which vessels would be likely to be affected by each of the subsidy reform policies under consideration by the WTO. Storing whether each vessel meets a certain "definition" (i.e. qualification) speeds up selection of affected vessels within the SubsidyExplorer toolkit. 

```{r}
# Load packages
library(knitr) 
library(countrycode) # country name matching
library(janitor) # cleaning column names
library(tidyverse)

# Package to normalize vessel names in the same way GFW normalizes vessel names
#devtools::install_github("jcvdav/startR")
#library(startR)
# This used to live in the startR package, but now lives locally - the package developer is currently having some difficulties related to the update to R version 4.0
source(here::here("scripts", "functions", "normalize.R"))
devtools::install_github("rich-iannone/UnidecodeR")
library(UnidecodeR)

# Path to global vessel list
vessel_list_path <- here::here("results", "04-model-parameterization", "vessel_list_2018_regions.csv")

# Paths to other data files
iuu_dat_path <- here::here("data", "raw", "combined-IUU-vessel-list", "IUUList-20190430.csv")
costello_2016_path <- here::here("data", "edited", "costello_et_al_2016_upsides_tidy.csv")
ram_dat_path <- here::here("data", "edited", "RAMLDB_v4.491_bvmsy_tidy.csv")
eez_info_path <- here::here("data", "lookup-tables", "eez_info.csv")

```

```{r}
# Code options
knitr::opts_chunk$set(echo = FALSE, error = FALSE, message = FALSE, warning = F)

# Directories for the output files generated by this script
results_dir <- here::here("results", "05-subsidy-reform-definitions/")
if (dir.exists(results_dir) == F) {
  dir.create(results_dir, recursive = T)
}

app_dir <- here::here("app", "data/")
if (dir.exists(app_dir) == F) {
  dir.create(app_dir, recursive = T)
}

```

# Input vessel info table

Our global vessel info table was created in a previous script (`r here::here("scripts", "1-vessel-list.Rmd")`) and saved to `r vessel_list_path`. 

```{r}
vessel_list <- read.csv(vessel_list_path, stringsAsFactors = F)
```

```{r}
# Country lookup table
country_lookup <- read_csv(here::here("data", "lookup-tables", "country_dependencies.csv"))

# Countries participating in the Northern Agreement for fishing
eu_countries <- country_lookup$iso3[country_lookup$is_EU]
eu_northern_agreement <- c("NOR", "ISL", "SJM")
```

# Apply Definitions

## Country Development Status

We use the flag of the vessel to assign development status.

```{r}
# Sovereign development statuses
sov_country_development_status <- country_lookup %>%
  dplyr::filter(!is_overseas_territory) %>%
  group_by(iso3) %>%
  summarize(development_status = case_when(!is.na(development_status) ~ development_status,
                                                TRUE ~ "Developing"))
# Get development status
ter_development_status <- country_lookup %>%
  dplyr::filter(is_overseas_territory) %>%
  dplyr::select(-development_status) %>%
  left_join(sov_country_development_status, by = c("sovereign_iso3" = "iso3")) %>%
  dplyr::select(iso3, development_status)
  
# combine
development_status <- sov_country_development_status %>%
  bind_rows(ter_development_status)

# Add development status to the vessel list
temp_vessel_list <- vessel_list %>%
  left_join(development_status, by = c("flag_iso3" = "iso3")) 

```

## High seas fishing

In order to later be able to identify high seas fishing, we need to determine fraction of each vessel's total annual fishing effort spent on the high seas.  

```{r}
# Get high seas totals by vessel
vessel_totals_high_seas <- temp_vessel_list %>%
  dplyr::filter(eez_id == 0) %>%
  group_by(ssvid) %>%
  summarize(fishing_h_hs = sum(fishing_hours_eez_fao_ter, na.rm = T),
            fishing_KWh_hs = sum(fishing_KWh_eez_fao_ter, na.rm = T))

# Now calculate high seas proportions
vessel_totals_prop_hs <- temp_vessel_list %>%
  group_by(ssvid) %>%
  summarize(fishing_h = sum(fishing_hours_eez_fao_ter, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T)) %>%
  ungroup() %>%
  left_join(vessel_totals_high_seas, by = "ssvid") %>%
  mutate(prop_fishing_hours_high_seas = case_when(is.na(fishing_h_hs) ~ 0, 
                                           TRUE ~ fishing_h_hs/fishing_h),
         prop_fishing_KWh_high_seas = case_when(is.na(fishing_KWh_hs) ~ 0, 
                                         TRUE ~ fishing_KWh_hs/fishing_KWh)) %>%
  dplyr::select(ssvid, prop_fishing_hours_high_seas, prop_fishing_KWh_high_seas)

# Add back into vessel list
temp_vessel_list <- temp_vessel_list %>%
  left_join(vessel_totals_prop_hs, by = "ssvid")
```

## Distant water fishing 

In order to determine whether a vessel is engaging in distant water fishing, we need to know something about the sovereign state of the vessel's flag state. We have already assigned this. 

```{r}
# Get pairings of administering and sovereign states from the vessel data
sovereign_pairings <- temp_vessel_list %>%
  distinct(flag_iso3, flag_sovereign_iso3) %>%
  arrange(flag_iso3)

```

```{r}
# We fist exclude entries where the flag state of the vessel and the administering state of the EEZ in which it is fishing are the same
distant_water_fishing <- temp_vessel_list %>%
  mutate(territory1_match = case_when(!is.na(eez_territory1_iso3) & flag_iso3 == eez_territory1_iso3 ~ T,
                                      TRUE ~ F),
         territory2_match = case_when(!is.na(eez_territory2_iso3) & flag_iso3 == eez_territory2_iso3 ~ T,
                                      TRUE ~ F),
         territory3_match = case_when(!is.na(eez_territory3_iso3) & flag_iso3 == eez_territory3_iso3 ~ T,
                                      TRUE ~ F)) %>%
  dplyr::filter(!territory1_match & !territory2_match & !territory3_match)
```

### "Intra-EU fishing".

Vessels flagged to EU member states have the right to fish in the EEZs of other EU member states, therefore we do not include these instances as "distant water fishing". In order to remove these from our dataset, we first need create a list of EU member states, as well as non-member states with with the EU has fishing agreements as part of the northern agreement (i.e. Norway and Iceland). 

We define "intra-EU fishing" as fishing activity that meets one of the following criteria: 
1. The flag state of the vessel is a member state of the EU, Norway, Svalbard and Jan Mayen, or Iceland and the EEZ in which it is fishing is administered by a member state of the EU, Norway, Svalbard and Jan Mayen, or Iceland (e.g. a French-flagged vessel fishing in the EEZ of Spain);
2. The sovereign of the vessel's flag state is a member state of the EU, Norway, Svalbard and Jan Mayen, or Iceland and the EEZ in which it is fishing is administered by a member state of the EU, Norway, Svalbard and Jan Mayen, or Iceland (e.g. a XXX-flagged vessel fishing in the EEZ of Spain).

We do not consider the following to be "intra-EU fishing" (and is therefore distante water fishing): 
1. The flag state of the vessel is a member state of the EU, Norway, Svalbard and Jan Mayen, or Iceland and the EEZ in which it is fishing belongs to a territory or distant region of an EU member state (e.g. a French-flagged vessel fishing in the EEZ of French Polynesia). 

```{r}
# Exclude intra EU fishing
distant_water_fishing <- distant_water_fishing %>%
  mutate(is_intra_EU = case_when(flag_iso3 %in% c(eu_countries, eu_northern_agreement) & eez_territory1_iso3 %in% c(eu_countries, eu_northern_agreement) ~ T,
                                 flag_sovereign_iso3 %in% c(eu_countries, eu_northern_agreement) & eez_territory1_iso3 %in% c(eu_countries, eu_northern_agreement) ~ T,
                                 TRUE ~ F)) %>%
  dplyr::filter(!is_intra_EU) %>%
  mutate(distant_water = T) %>%
  dplyr::select(eez_id, ssvid, fao_region, is_territorial, distant_water)
```

#### A Note on Sovereignity

If the sovereign country of the flag state of a vessel is the same as the sovereign country of the EEZ in which it's fishing, this connection may still be considered to be distant water fishing, though this is generally not considered to be foreign distant water fishing.

We descibe "sovereign fishing" as fishing activity that meets one of the following criteria: 
1. The flag state of the vessel is the same as the sovereign of the administering territory of the EEZ in which it is fishing (e.g. a US flagged vessel fishing in the EEZ of Palmyra Atoll).
2. The sovereign of the flag state of the vessel is the same as the sovereign of the administering territory of the EEZ in which it is fishing (e.g. a Puerto-Rican flagged vessel fishing in the EEZ of Palmyra Atoll);

Though not foreign, we consider sovereign fishing to still be distant water fishing for the purposes of this analysis. 

```{r}
temp_vessel_list <- temp_vessel_list %>%
  left_join(distant_water_fishing, by = c("eez_id", "ssvid", "fao_region", "is_territorial"))
temp_vessel_list$distant_water[is.na(temp_vessel_list$distant_water)] <- F
```

## IUU

### IUU #1: RFMO IUU vessel lists

Using the combined list of IUU vessels (https://iuu-vessels.org/), we match as many vessels as possible to our vessel list. 

GFW uses a function called "normalize_shipname" from the `startR` package to normalize vessel names. We apply the same function to the IUU vessel list as one method of facilitating matching. 

```{r}
# Read in IUU list
iuu_list <- read_csv(iuu_dat_path) %>%
  clean_names(case = "snake")

# Extract only vessels currently listed, and create a new entry for each using the "name" field
# We also normalize shipname using the same function GFW uses to normalize their shipnames (ignore roman numeral error - double checked this and it's working fine)
iuu_clean_name <- iuu_list %>% 
   dplyr::filter(currently_listed) %>%
   mutate(name_new = str_split(name, ", "))%>%
   unnest(cols = c(name_new)) %>%
  dplyr::select(name, name_new, imo, gear_type, flag, ircs, ssvid, gt, length) %>%
  mutate(n_shipname = map_chr(name_new, normalize_shipname))
  
# Do the same thing, but for RFMO name since they differ
iuu_clean_rfmo_name <- iuu_list %>% 
   dplyr::filter(currently_listed) %>%
   mutate(name_new1 = str_split(rfmo_name, ", "),
          name_new2 = str_split(rfmo_name, "[()]"),
          name_new = case_when(name_new1 != rfmo_name ~ name_new1,
                               TRUE ~ name_new2)) %>%
   unnest(cols = c(name_new)) %>%
  dplyr::select(name, name_new, imo, gear_type, flag, ircs, ssvid, gt, length) %>%
  mutate(n_shipname = map_chr(name_new, normalize_shipname))

### Blank database for IUU vessels
iuu_matching_vessels <- tibble(ssvid = integer(0),
                               flag_iso3 = character(0),
                               broadcast_shipname = character(0))

### Abbreviated vessel list for matching
vessels_for_matching <- temp_vessel_list %>%
  distinct(ssvid, flag_iso3, vessel_class, length_m, tonnage_gt, engine_power_kw, broadcast_shipname, broadcast_callsign, broadcast_imo)

iuu <- iuu_clean_name %>%
  dplyr::select(-name, -name_new) %>%
  rename(iuu_imo = imo, iuu_gear_type = gear_type, iuu_flag = flag, iuu_callsign = ircs, iuu_ssvid = ssvid, iuu_tonnage_gt = gt, iuu_length = length, iuu_shipname = n_shipname)

iuu2 <- iuu_clean_rfmo_name %>%
  dplyr::select(-name, -name_new) %>%
  rename(iuu_imo = imo, iuu_gear_type = gear_type, iuu_flag = flag, iuu_callsign = ircs, iuu_ssvid = ssvid, iuu_tonnage_gt = gt, iuu_length = length, iuu_shipname = n_shipname)

### Get matches by ssvid
# Join by id number - 4 matching vessels, all appear to be a match
iuu_by_ssvid <- vessels_for_matching %>%
    inner_join(iuu, by = c("ssvid"="iuu_ssvid"))

iuu_matching_vessels <- iuu_matching_vessels %>%
  bind_rows(iuu_by_ssvid %>% distinct(ssvid, flag_iso3, broadcast_shipname))

### Get matches by call sign - 3 matching vessels, only two appear to be a match
iuu_by_callsign <- vessels_for_matching %>%
    inner_join(iuu %>% dplyr::filter(!is.na(iuu_callsign)), by = c("broadcast_callsign"="iuu_callsign"))

iuu_matching_vessels <- iuu_matching_vessels %>%
  bind_rows(iuu_by_callsign %>% distinct(ssvid, flag_iso3, broadcast_shipname) %>% dplyr::filter(broadcast_shipname != "ROSE4"))

### Get matches by IMO - 3 matching vessels, all appear to be enough for a match
iuu_by_imo <- vessels_for_matching %>%
    inner_join(iuu %>% dplyr::filter(!is.na(iuu_imo)), by = c("broadcast_imo"="iuu_imo"))

iuu_matching_vessels <- iuu_matching_vessels %>%
  bind_rows(iuu_by_imo %>% distinct(ssvid, flag_iso3, broadcast_shipname))

### Match by normalized shipname  - 52 matching records, most are not a match. 
iuu_by_name1 <- vessels_for_matching %>%
    inner_join(iuu %>% dplyr::filter(!is.na(iuu_shipname)), by = c("broadcast_shipname"="iuu_shipname")) %>%
  arrange(broadcast_shipname)

iuu_by_name2 <- vessels_for_matching %>%
    inner_join(iuu2 %>% dplyr::filter(!is.na(iuu_shipname)), by = c("broadcast_shipname"="iuu_shipname")) %>%
  arrange(broadcast_shipname)

# Manually determined matches
name_matches1 <- iuu_by_name1 %>%
  dplyr::filter(broadcast_shipname %in% c("EPHRAEEM", "HUALI8", "SHUNCHANG3"))

name_matches2 <- iuu_by_name2 %>%
  dplyr::filter(broadcast_shipname %in% c("HAILUNG", "HUALI8", "MYSMARII"))

iuu_matching_vessels <- iuu_matching_vessels %>%
  bind_rows(name_matches1 %>% distinct(ssvid, flag_iso3, broadcast_shipname)) %>%
  bind_rows(name_matches2 %>% distinct(ssvid, flag_iso3, broadcast_shipname))

# Get all matches - 7 vessels
iuu_matches <- iuu_matching_vessels %>%
  distinct(ssvid)

# Apply to vessel list
temp_vessel_list <- temp_vessel_list %>%
  mutate(iuu1 = ifelse(ssvid %in% iuu_matches$ssvid, T, F))
                    
```

### Disputed areas

We pull the version of marine regions used by GFW to characterize fishing activity. 

```{r}
# EEZ info lookup table from GFW
eez_info <- read_csv(eez_info_path) 

# Get disputed regions
disputed_areas <- eez_info %>%
  mutate(is_disputed = case_when(str_detect(reporting_name, "Disputed area") == T ~ T,
                                 TRUE ~ F)) %>%
  dplyr::filter(is_disputed)

# apply to vessel list
temp_vessel_list <- temp_vessel_list %>%
  mutate(is_disputed = ifelse(eez_id %in% disputed_areas$eez_id, T, F))
```

## Overfished and Unassessed Stocks

### OA #1/#2: B/Bmsy < 1/0.8 as determined by the RAM legacy stock assessment database

```{r}
# Load RAM data to get the status of stocks
costello_2016_dat <- read_csv(costello_2016_path) %>%
  clean_names(case = "snake")

# Separate RAM stocks 
ram_stocks <- costello_2016_dat %>%
  dplyr::filter(dbase == "RAM")

# Get statuses of multinational ram stocks by fao region
multinational_ram_stocks <- ram_stocks %>%
  dplyr::filter(is.na(iso3)) %>%
  group_by(region_fao_split) %>%
  summarize(n_stocks = n_distinct(id_orig),
            mean_bv_bmsy = mean(bv_bmsy, na.rm = T),
            median_bv_bmsy = mean(bv_bmsy, na.rm = T),
            w_mean_bv_bmsy = weighted.mean(bv_bmsy, w = (msy/n_regions), na.rm = T)) %>%
  mutate(oa1_median = ifelse(median_bv_bmsy >= 1, F, T),
         oa1_w_mean = ifelse(w_mean_bv_bmsy >= 1, F, T),
         oa2_median = ifelse(median_bv_bmsy >= 0.8, F, T),
         oa2_w_mean = ifelse(w_mean_bv_bmsy >= 0.8, F, T)) %>%
  dplyr::select(region_fao_split, oa1_median, oa1_w_mean, oa2_median, oa2_w_mean) %>%
  mutate(eez_id = 0)

# Get statuses of domestic stocks by flag state and fao region
eez_ram_stocks <- ram_stocks %>%
  dplyr::filter(!is.na(iso3)) 

# Get eez ids from fishing effort
eez_effort <- temp_vessel_list %>%
  distinct(eez_id, fao_region, eez_territory1_iso3, eez_territory2_iso3, eez_territory3_iso3, eez_sovereign1_iso3, eez_sovereign2_iso3, eez_sovereign3_iso3)

# Match
eez_ram_stocks <- eez_ram_stocks %>%
  left_join(eez_effort, by = c("iso3" = "eez_territory1_iso3", "region_fao_split" = "fao_region")) %>%
  group_by(eez_id, region_fao_split) %>%
  summarize(n_stocks = n_distinct(id_orig),
            mean_bv_bmsy = mean(bv_bmsy, na.rm = T),
            median_bv_bmsy = mean(bv_bmsy, na.rm = T),
            w_mean_bv_bmsy = weighted.mean(bv_bmsy, w = (msy/n_regions), na.rm = T)) %>%
  ungroup() %>%
  mutate(oa1_median = ifelse(median_bv_bmsy >= 1, F, T),
         oa1_w_mean = ifelse(w_mean_bv_bmsy >= 1, F, T),
         oa2_median = ifelse(median_bv_bmsy >= 0.8, F, T),
         oa2_w_mean = ifelse(w_mean_bv_bmsy >= 0.8, F, T)) %>%
  dplyr::select(eez_id, region_fao_split, oa1_median, oa1_w_mean, oa2_median, oa2_w_mean)

# Combine 
oa_1_2 <- multinational_ram_stocks %>%
  bind_rows(eez_ram_stocks)

# Join back to vessel list
temp_vessel_list <- temp_vessel_list %>%
  left_join(oa_1_2, by = c("eez_id" = "eez_id", "fao_region" = "region_fao_split"))

temp_vessel_list$oa1_median[is.na(temp_vessel_list$oa1_median)] <- F
temp_vessel_list$oa2_median[is.na(temp_vessel_list$oa2_median)] <- F
temp_vessel_list$oa1_w_mean[is.na(temp_vessel_list$oa1_w_mean)] <- F
temp_vessel_list$oa2_w_mean[is.na(temp_vessel_list$oa2_w_mean)] <- F

```

### OA #3/4: Overfished as determined by Costello et al. (2016)

```{r}
# Get statuses of multinational ram stocks by fao region
multinational_stocks <- costello_2016_dat %>%
  dplyr::filter(is.na(iso3)) %>%
  group_by(region_fao_split) %>%
  summarize(n_stocks = n_distinct(id_orig),
            mean_bv_bmsy = mean(bv_bmsy, na.rm = T),
            median_bv_bmsy = mean(bv_bmsy, na.rm = T),
            w_mean_bv_bmsy = weighted.mean(bv_bmsy, w = (msy/n_regions), na.rm = T)) %>%
  mutate(oa3_median = ifelse(median_bv_bmsy >= 1, F, T),
         oa3_w_mean = ifelse(w_mean_bv_bmsy >= 1, F, T),
         oa4_median = ifelse(median_bv_bmsy >= 0.8, F, T),
         oa4_w_mean = ifelse(w_mean_bv_bmsy >= 0.8, F, T)) %>%
  dplyr::select(region_fao_split, oa3_median, oa3_w_mean, oa4_median, oa4_w_mean) %>%
  mutate(eez_id = 0)

# Get statuses of domestic stocks by flag state and fao region
eez_stocks <- costello_2016_dat %>%
  dplyr::filter(!is.na(iso3)) %>%
  left_join(eez_effort, by = c("iso3" = "eez_territory1_iso3", "region_fao_split" = "fao_region")) %>%
  group_by(eez_id, region_fao_split) %>%
  summarize(n_stocks = n_distinct(id_orig),
            mean_bv_bmsy = mean(bv_bmsy, na.rm = T),
            median_bv_bmsy = mean(bv_bmsy, na.rm = T),
            w_mean_bv_bmsy = weighted.mean(bv_bmsy, w = (msy/n_regions), na.rm = T)) %>%
  ungroup() %>%
  mutate(oa3_median = ifelse(median_bv_bmsy >= 1, F, T),
         oa3_w_mean = ifelse(w_mean_bv_bmsy >= 1, F, T),
         oa4_median = ifelse(median_bv_bmsy >= 0.8, F, T),
         oa4_w_mean = ifelse(w_mean_bv_bmsy >= 0.8, F, T)) %>%
  dplyr::select(eez_id, region_fao_split, oa3_median, oa3_w_mean, oa4_median, oa4_w_mean)

# Combine 
oa_3_4 <- multinational_stocks %>%
  bind_rows(eez_stocks)

# Join back to vessel list
temp_vessel_list <- temp_vessel_list %>%
  left_join(oa_3_4, by = c("eez_id" = "eez_id", "fao_region" = "region_fao_split"))

temp_vessel_list$oa3_median[is.na(temp_vessel_list$oa3_median)] <- F
temp_vessel_list$oa4_median[is.na(temp_vessel_list$oa4_median)] <- F
temp_vessel_list$oa3_w_mean[is.na(temp_vessel_list$oa3_w_mean)] <- F
temp_vessel_list$oa4_w_mean[is.na(temp_vessel_list$oa4_w_mean)] <- F
```


# Export final vessel list

```{r}
# Export final vessel list with definitions
write_csv(temp_vessel_list, paste0(results_dir, "vessel_list_2018_final.csv"))
write_csv(temp_vessel_list, paste0(app_dir, "vessel_list_2018_final.csv"))
```
