==> scripts/06_run_model.Rmd <==
---
output:
  html_document: default

title: "SubsidyExplorer - Step 6: Run Model"
author: "Kat Millage, Vienna Saccomanno, Laura Lea Rubino, Christopher Costello"
date: "`r format(Sys.time(), '%B %d, %Y')`"
fontsize: 11pt
bibliography: "../pew-subsidies.bib"
---
# Introduction 

This script generates the most ambitious global and regional results obtained by removing all capacity enhancing subsidies from all WTO Members and Observers. This is effectively the reference case scenario for our model, which will be used later as a benchmark to which we can compare the results of specific proposals. 

```{r setup, include=FALSE}
# Load packages
library(knitr) 
library(tidyverse)

knitr::opts_chunk$set(echo = TRUE)

make_plots <- T

if(make_plots){
  library(RColorBrewer)
  library(scales)
}

# End year for simulations
end_year <- 2080

# Threshold to be considered managed (using FMI scores)
managed_cutoff <- 0.5

# Directories for the output files generated by this script
results_dir <- paste0(here::here("results", "06-run-model/"), "mgmt_threshold_", managed_cutoff, "/")
if (dir.exists(results_dir) == F) {
  dir.create(results_dir, recursive = T)
}

bad_subsidy_types <- c("B1", "B2", "B3", "B4", "B5", "B6", "B7")
ugly_subsidy_types <- c("C1", "C2", "C3")
```

```{r}
# Source model script
source(here::here("app", "scripts", "BioEconModel.R"))

# Source plotting script
source(here::here("scripts", "functions", "BioEconomicModelPlots.R"))
```

```{r}
# Load vessel list
vessel_dat <- read.csv(here::here("app", "data", "vessel_list_2018_final.csv"), stringsAsFactors = F)

# Create high seas code
vessel_dat <- vessel_dat %>%
  mutate(eez_hs_code = case_when(eez_id == 0 ~ paste0("HS-", fao_region),
                                 TRUE ~ as.character(eez_id)))

# Load country dependencies helper file
country_dependencies <- read_csv(here::here("data", "lookup-tables", "country_dependencies.csv"))
```

```{r}
# Load biological parameters for the model
bio_dat_regional <- read.csv(here::here("app", "data", "model_parameters_regional_3_regions.csv"))

# Regional parameter list
bio_dat_regional_list <- bio_dat_regional %>%
  gather(region, value, -c(1:2)) %>%
  group_by(region) %>%
  group_split()
```

```{r pressure, echo=FALSE}
# Assign fleets to vessels
fleet_vessels <- vessel_dat %>%
  left_join(country_dependencies %>% dplyr::select(iso3, is_WTO), by = c("flag_iso3" = "iso3")) %>%
  mutate(fleet = case_when(is_WTO & fmi_best >= managed_cutoff & bad_subs > 0 ~ "affected_managed",
                           is_WTO & fmi_best < managed_cutoff & bad_subs > 0 ~ "affected_oa",
                           (!is_WTO & fmi_best >= managed_cutoff) ~ "unaffected_managed",
                           (!is_WTO & fmi_best < managed_cutoff) ~ "unaffected_oa",
                           TRUE ~ "unaffected_oa")) %>%
    mutate(bad_ugly_subs = rowSums(select(., one_of(paste0(c(bad_subsidy_types, ugly_subsidy_types), "_subs")))))
```

# Section 1: Absolute subsidy values

This first section presents results obtained by using absolute subsidy values. 

## Scenario 1: Remove all bad + ugly

```{r}
# Create model input for removing all bad subsidies
abs_all_BC_fleet_summary <- fleet_vessels %>%
  group_by(region, fleet) %>%
  summarize(catch = sum(catch, na.rm = T),
            bad_subs = sum(bad_ugly_subs, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            removed_subs = sum(bad_ugly_subs, na.rm = T)) %>%
  ungroup()

# Turn into a list
abs_all_BC_list <- abs_all_BC_fleet_summary %>%
  group_by(region) %>%
  group_split()
names(abs_all_BC_list) <- colnames(bio_dat_regional)[-c(1:2)]

# Run model
abs_all_BC_results <- pmap_df(list(fleet = abs_all_BC_list, 
                                       region = names(abs_all_BC_list),
                                       bio_param = bio_dat_regional_list),
                                  BioEconModel,
                                  end_year = end_year,
                                  return = "all")

# Extract results timeseries
abs_all_BC_results_all <- abs_all_BC_results %>%
  dplyr::filter(Year > 2018) %>%
  group_by(Year, Variable, Fleet) %>%
  mutate(Diff = case_when(BAU != 0 ~ (Reform - BAU)/abs(BAU),
                          TRUE ~ 0)) %>%
  ungroup() %>%
  mutate(run_id = "Absolute",
         run_name = "All subsidies",
         run_group = "Reference: All capacity-enhancing & \nambiguous subsidies (absolute $)")

write_csv(abs_all_BC_results_all, paste0(results_dir, "BC_abs_ts_data.csv"))
```

```{r}
# Make plots
if(make_plots){
  
  # 1) global (aggregated) timeseries plot
  abs_all_BC_ts_plot <- BioEconomicModelPlots(dat = abs_all_BC_results_all,
                                              spatial_res = "global",
                                              temporal_res = "all",
                                              start_year = 2018,
                                              end_year = end_year,
                                              plot_name = "BC_abs_global_ts_plot",
                                              plot_width = 8,
                                              plot_height = 6.5,
                                              results_dir = results_dir)
  
  # 2) global (aggregated) "kobe" final year plot
  abs_all_BC_end_plot <- BioEconomicModelPlots(dat = abs_all_BC_results_all,
                                                      spatial_res = "global",
                                                      temporal_res = "last",
                                                      end_year = end_year,
                                                      plot_name = paste0("BC_abs_global_", end_year, "_plot"),
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
  
  # 3) regional timeseries plot
  regional_abs_all_BC_ts_plot <- BioEconomicModelPlots(dat = abs_all_BC_results_all,
                                                      spatial_res = "regional",
                                                      temporal_res = "all",
                                                      start_year = 2018,
                                                      end_year = end_year,
                                                      plot_name = "BC_abs_regional_ts_plot",
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
  
  # 4) regional "kobe" final year plot
  regional_abs_all_BC_end_plot <- BioEconomicModelPlots(dat = abs_all_BC_results_all,
                                                      spatial_res = "regional",
                                                      temporal_res = "last",
                                                      end_year = end_year,
                                                      plot_name = paste0("BC_abs_regional_", end_year, "_plot"),
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
}
```

## Scenario 2: Remove all bad

```{r}
# Create model input for removing all bad subsidies
abs_all_B_fleet_summary <- fleet_vessels %>%
  group_by(region, fleet) %>%
  summarize(catch = sum(catch, na.rm = T),
            bad_subs = sum(bad_ugly_subs, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            removed_subs = sum(bad_subs, na.rm = T)) %>%
  ungroup()

# Turn into a list
abs_all_B_list <- abs_all_B_fleet_summary %>%
  group_by(region) %>%
  group_split()
names(abs_all_B_list) <- colnames(bio_dat_regional)[-c(1:2)]

# Run model
abs_all_B_results <- pmap_df(list(fleet = abs_all_B_list, 
                                       region = names(abs_all_B_list),
                                       bio_param = bio_dat_regional_list),
                                  BioEconModel,
                                  end_year = end_year,
                                  return = "all")

# Extract results timeseries
abs_all_B_results_all <- abs_all_B_results %>%
  dplyr::filter(Year > 2018) %>%
  group_by(Year, Variable, Fleet) %>%
  mutate(Diff = case_when(BAU != 0 ~ (Reform - BAU)/abs(BAU),
                          TRUE ~ 0)) %>%
  ungroup() %>%
  mutate(run_id = "Absolute",
         run_name = "Capacity-enhancing subsidies",
         run_group = "Reference: All capacity-enhancing \nsubsidies (absolute $)")

write_csv(abs_all_B_results_all, paste0(results_dir, "B_abs_ts_data.csv"))
```

```{r}
# Make plots
if(make_plots){
  
  # 1) global (aggregated) timeseries plot
  abs_all_B_ts_plot <- BioEconomicModelPlots(dat = abs_all_B_results_all,
                                              spatial_res = "global",
                                              temporal_res = "all",
                                              start_year = 2018,
                                              end_year = end_year,
                                              plot_name = "B_abs_global_ts_plot",
                                              plot_width = 8,
                                              plot_height = 6.5,
                                              results_dir = results_dir)
  
  # 2) global (aggregated) "kobe" final year plot
  abs_all_B_end_plot <- BioEconomicModelPlots(dat = abs_all_B_results_all,
                                                      spatial_res = "global",
                                                      temporal_res = "last",
                                                      end_year = end_year,
                                                      plot_name = paste0("B_abs_global_", end_year, "_plot"),
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
  
  # 3) regional timeseries plot
  regional_abs_all_B_ts_plot <- BioEconomicModelPlots(dat = abs_all_B_results_all,
                                                      spatial_res = "regional",
                                                      temporal_res = "all",
                                                      start_year = 2018,
                                                      end_year = end_year,
                                                      plot_name = "B_abs_regional_ts_plot",
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
  
  # 4) regional "kobe" final year plot
  regional_abs_all_B_end_plot <- BioEconomicModelPlots(dat = abs_all_B_results_all,
                                                      spatial_res = "regional",
                                                      temporal_res = "last",
                                                      end_year = end_year,
                                                      plot_name = paste0("B_abs_regional_", end_year, "_plot"),
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
}
```

## Scenario 3: Remove fuel subsidies

```{r}
# Create model input for removing all bad subsidies
abs_fuel_fleet_summary <- fleet_vessels %>%
  group_by(region, fleet) %>%
  summarize(catch = sum(catch, na.rm = T),
            bad_subs = sum(bad_ugly_subs, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            removed_subs = sum(B7_subs, na.rm = T)) %>%
  ungroup()

# Turn into a list
abs_fuel_list <- abs_fuel_fleet_summary %>%
  group_by(region) %>%
  group_split()
names(abs_fuel_list) <- colnames(bio_dat_regional)[-c(1:2)]

# Run model
abs_fuel_results <- pmap_df(list(fleet = abs_fuel_list, 
                                       region = names(abs_fuel_list),
                                       bio_param = bio_dat_regional_list),
                                  BioEconModel,
                                  end_year = end_year,
                                  return = "all")

# Extract results timeseries
abs_fuel_results_all <- abs_fuel_results %>%
  dplyr::filter(Year > 2018) %>%
  group_by(Year, Variable, Fleet) %>%
  mutate(Diff = case_when(BAU != 0 ~ (Reform - BAU)/abs(BAU),
                          TRUE ~ 0)) %>%
  ungroup() %>%
  mutate(run_id = "Absolute",
         run_name = "Fuel subsidies",
         run_group = "Reference: All fuel \nsubsidies (absolute $)")

write_csv(abs_fuel_results_all, paste0(results_dir, "fuel_abs_ts_data.csv"))
```

```{r}
# Make plots
if(make_plots){
  
  # 1) global (aggregated) timeseries plot
  abs_fuel_ts_plot <- BioEconomicModelPlots(dat = abs_fuel_results_all,
                                              spatial_res = "global",
                                              temporal_res = "all",
                                              start_year = 2018,
                                              end_year = end_year,
                                              plot_name = "fuel_abs_global_ts_plot",
                                              plot_width = 8,
                                              plot_height = 6.5,
                                              results_dir = results_dir)
  
  # 2) global (aggregated) "kobe" final year plot
  abs_fuel_end_plot <- BioEconomicModelPlots(dat = abs_fuel_results_all,
                                                      spatial_res = "global",
                                                      temporal_res = "last",
                                                      end_year = end_year,
                                                      plot_name = paste0("fuel_abs_global_", end_year, "_plot"),
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
  
  # 3) regional timeseries plot
  regional_abs_fuel_ts_plot <- BioEconomicModelPlots(dat = abs_fuel_results_all,
                                                      spatial_res = "regional",
                                                      temporal_res = "all",
                                                      start_year = 2018,
                                                      end_year = end_year,
                                                      plot_name = "fuel_abs_regional_ts_plot",
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
  
  # 4) regional "kobe" final year plot
  regional_abs_fuel_end_plot <- BioEconomicModelPlots(dat = abs_fuel_results_all,
                                                      spatial_res = "regional",
                                                      temporal_res = "last",
                                                      end_year = end_year,
                                                      plot_name = paste0("fuel_abs_regional_", end_year, "_plot"),
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
}
```

## Combine absolute subsidy results 

```{r}
abs_results_all <- abs_all_BC_results_all %>%
  bind_rows(abs_all_B_results_all) %>%
  bind_rows(abs_fuel_results_all)
```

# Section 2: Effective subsidies

This second section presents results obtained by using effective subsidy values - based on suggestions from Roger Martini at the OECD.

```{r}
rel <- read_csv(here::here("data", "lookup-tables", "oecd_relative_effects_lookup_table.csv"))

rel_fleet_vessels <- fleet_vessels %>%
   mutate(B1_subs = B1_subs * rel$rel_effect_effort[rel$type == "B1"], 
           B2_subs = B2_subs * rel$rel_effect_effort[rel$type == "B2"],
           B3_subs = B3_subs * rel$rel_effect_effort[rel$type == "B3"],
           B4_subs = B4_subs * rel$rel_effect_effort[rel$type == "B4"],
           B5_subs = B5_subs * rel$rel_effect_effort[rel$type == "B5"], 
           B6_subs = B6_subs * rel$rel_effect_effort[rel$type == "B6"],
           B7_subs = B7_subs * rel$rel_effect_effort[rel$type == "B7"], 
           C1_subs = C1_subs * rel$rel_effect_effort[rel$type == "C1"], 
           C2_subs = C2_subs * rel$rel_effect_effort[rel$type == "C2"],
           C3_subs = C3_subs * rel$rel_effect_effort[rel$type == "C3"]) %>%
  ungroup() %>%
  mutate(bad_subs = rowSums(select(., one_of(paste0(bad_subsidy_types, "_subs")))),
         ugly_subs = rowSums(select(., one_of(paste0(ugly_subsidy_types, "_subs")))),
         bad_ugly_subs = rowSums(select(., one_of(paste0(c(bad_subsidy_types, ugly_subsidy_types), "_subs")))))
```

## Scenario 1: Remove all bad + ugly

```{r}
# Create model input for removing all bad subsidies
rel_all_BC_fleet_summary <- rel_fleet_vessels %>%
  group_by(region, fleet) %>%
  summarize(catch = sum(catch, na.rm = T),
            bad_subs = sum(bad_ugly_subs, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            removed_subs = sum(bad_ugly_subs, na.rm = T)) %>%
  ungroup()

# Turn into a list
rel_all_BC_list <- rel_all_BC_fleet_summary %>%
  group_by(region) %>%
  group_split()
names(rel_all_BC_list) <- colnames(bio_dat_regional)[-c(1:2)]

# Run model
rel_all_BC_results <- pmap_df(list(fleet = rel_all_BC_list, 
                                       region = names(rel_all_BC_list),
                                       bio_param = bio_dat_regional_list),
                                  BioEconModel,
                                  end_year = end_year,
                                  return = "all")

# Extract results timeseries
rel_all_BC_results_all <- rel_all_BC_results %>%
  dplyr::filter(Year > 2018) %>%
  group_by(Year, Variable, Fleet) %>%
  mutate(Diff = case_when(BAU != 0 ~ (Reform - BAU)/abs(BAU),
                          TRUE ~ 0)) %>%
  ungroup() %>%
  mutate(run_id = "Effective",
         run_name = "All subsidies",
         run_group = "Reference: All capacity-enhancing \n& ambiguous subsidies (effective $)")

write_csv(rel_all_BC_results_all, paste0(results_dir, "BC_rel_ts_data.csv"))
```

```{r}
# Make plots
if(make_plots){
  
  # 1) global (aggregated) timeseries plot
  rel_all_BC_ts_plot <- BioEconomicModelPlots(dat = rel_all_BC_results_all,
                                              spatial_res = "global",
                                              temporal_res = "all",
                                              start_year = 2018,
                                              end_year = end_year,
                                              plot_name = "BC_rel_global_ts_plot",
                                              plot_width = 8,
                                              plot_height = 6.5,
                                              results_dir = results_dir)
  
  # 2) global (aggregated) "kobe" final year plot
  rel_all_BC_end_plot <- BioEconomicModelPlots(dat = rel_all_BC_results_all,
                                                      spatial_res = "global",
                                                      temporal_res = "last",
                                                      end_year = end_year,
                                                      plot_name = paste0("BC_rel_global_", end_year, "_plot"),
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
  
  # 3) regional timeseries plot
  regional_rel_all_BC_ts_plot <- BioEconomicModelPlots(dat = rel_all_BC_results_all,
                                                      spatial_res = "regional",
                                                      temporal_res = "all",
                                                      start_year = 2018,
                                                      end_year = end_year,
                                                      plot_name = "BC_rel_regional_ts_plot",
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
  
  # 4) regional "kobe" final year plot
  regional_rel_all_BC_end_plot <- BioEconomicModelPlots(dat = rel_all_BC_results_all,
                                                      spatial_res = "regional",
                                                      temporal_res = "last",
                                                      end_year = end_year,
                                                      plot_name = paste0("BC_rel_regional_", end_year, "_plot"),
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
}
```

## Scenario 2: Remove all bad

```{r}
# Create model input for removing all bad subsidies
rel_all_B_fleet_summary <- rel_fleet_vessels %>%
  group_by(region, fleet) %>%
  summarize(catch = sum(catch, na.rm = T),
            bad_subs = sum(bad_ugly_subs, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            removed_subs = sum(bad_subs, na.rm = T)) %>%
  ungroup()

# Turn into a list
rel_all_B_list <- rel_all_B_fleet_summary %>%
  group_by(region) %>%
  group_split()
names(rel_all_B_list) <- colnames(bio_dat_regional)[-c(1:2)]

# Run model
rel_all_B_results <- pmap_df(list(fleet = rel_all_B_list, 
                                       region = names(rel_all_B_list),
                                       bio_param = bio_dat_regional_list),
                                  BioEconModel,
                                  end_year = end_year,
                                  return = "all")

# Extract results timeseries
rel_all_B_results_all <- rel_all_B_results %>%
  dplyr::filter(Year > 2018) %>%
  group_by(Year, Variable, Fleet) %>%
  mutate(Diff = case_when(BAU != 0 ~ (Reform - BAU)/abs(BAU),
                          TRUE ~ 0)) %>%
  ungroup() %>%
  mutate(run_id = "Effective",
         run_name = "Capacity-enhancing subsidies",
         run_group = "Reference: All capacity-enhancing \nsubsidies (effective $)")

write_csv(rel_all_B_results_all, paste0(results_dir, "B_rel_ts_data.csv"))
```

```{r}
# Make plots
if(make_plots){
  
  # 1) global (aggregated) timeseries plot
  rel_all_B_ts_plot <- BioEconomicModelPlots(dat = rel_all_B_results_all,
                                              spatial_res = "global",
                                              temporal_res = "all",
                                              start_year = 2018,
                                              end_year = end_year,
                                              plot_name = "B_rel_global_ts_plot",
                                              plot_width = 8,
                                              plot_height = 6.5,
                                              results_dir = results_dir)
  
  # 2) global (aggregated) "kobe" final year plot
  rel_all_B_end_plot <- BioEconomicModelPlots(dat = rel_all_B_results_all,
                                                      spatial_res = "global",
                                                      temporal_res = "last",
                                                      end_year = end_year,
                                                      plot_name = paste0("B_rel_global_", end_year, "_plot"),
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
  
  # 3) regional timeseries plot
  regional_rel_all_B_ts_plot <- BioEconomicModelPlots(dat = rel_all_B_results_all,
                                                      spatial_res = "regional",
                                                      temporal_res = "all",
                                                      start_year = 2018,
                                                      end_year = end_year,
                                                      plot_name = "B_rel_regional_ts_plot",
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
  
  # 4) regional "kobe" final year plot
  regional_rel_all_B_end_plot <- BioEconomicModelPlots(dat = rel_all_B_results_all,
                                                      spatial_res = "regional",
                                                      temporal_res = "last",
                                                      end_year = end_year,
                                                      plot_name = paste0("B_rel_regional_", end_year, "_plot"),
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
}
```

## Scenario 3: Remove fuel subsidies

```{r}
# Create model input for removing all bad subsidies
rel_fuel_fleet_summary <- rel_fleet_vessels %>%
  group_by(region, fleet) %>%
  summarize(catch = sum(catch, na.rm = T),
            bad_subs = sum(bad_ugly_subs, na.rm = T),
            fishing_KWh = sum(fishing_KWh_eez_fao_ter, na.rm = T),
            removed_subs = sum(B7_subs, na.rm = T)) %>%
  ungroup()

# Turn into a list
rel_fuel_list <- rel_fuel_fleet_summary %>%
  group_by(region) %>%
  group_split()
names(rel_fuel_list) <- colnames(bio_dat_regional)[-c(1:2)]

# Run model
rel_fuel_results <- pmap_df(list(fleet = rel_fuel_list, 
                                       region = names(rel_fuel_list),
                                       bio_param = bio_dat_regional_list),
                                  BioEconModel,
                                  end_year = end_year,
                                  return = "all")

# Extract results timeseries
rel_fuel_results_all <- rel_fuel_results %>%
  dplyr::filter(Year > 2018) %>%
  group_by(Year, Variable, Fleet) %>%
  mutate(Diff = case_when(BAU != 0 ~ (Reform - BAU)/abs(BAU),
                          TRUE ~ 0)) %>%
  ungroup() %>%
  mutate(run_id = "Effective",
         run_name = "Fuel subsidies",
         run_group = "Reference: All fuel \nsubsidies (effective $)")

write_csv(rel_fuel_results_all, paste0(results_dir, "fuel_rel_ts_data.csv"))
```

```{r}
# Make plots
if(make_plots){
  
  # 1) global (aggregated) timeseries plot
  rel_fuel_ts_plot <- BioEconomicModelPlots(dat = rel_fuel_results_all,
                                              spatial_res = "global",
                                              temporal_res = "all",
                                              start_year = 2018,
                                              end_year = end_year,
                                              plot_name = "fuel_rel_global_ts_plot",
                                              plot_width = 8,
                                              plot_height = 6.5,
                                              results_dir = results_dir)
  
  # 2) global (aggregated) "kobe" final year plot
  rel_fuel_end_plot <- BioEconomicModelPlots(dat = rel_fuel_results_all,
                                                      spatial_res = "global",
                                                      temporal_res = "last",
                                                      end_year = end_year,
                                                      plot_name = paste0("fuel_rel_global_", end_year, "_plot"),
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
  
  # 3) regional timeseries plot
  regional_fuel_BC_ts_plot <- BioEconomicModelPlots(dat = rel_fuel_results_all,
                                                      spatial_res = "regional",
                                                      temporal_res = "all",
                                                      start_year = 2018,
                                                      end_year = end_year,
                                                      plot_name = "fuel_rel_regional_ts_plot",
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
  
  # 4) regional "kobe" final year plot
  regional_rel_fuel_end_plot <- BioEconomicModelPlots(dat = rel_fuel_results_all,
                                                      spatial_res = "regional",
                                                      temporal_res = "last",
                                                      end_year = end_year,
                                                      plot_name = paste0("fuel_rel_regional_", end_year, "_plot"),
                                                      plot_width = 8,
                                                      plot_height = 6.5,
                                                      results_dir = results_dir)
}
```

## Combine absolute subsidy results 

```{r}
rel_results_all <- rel_all_BC_results_all %>%
  bind_rows(rel_all_B_results_all) %>%
  bind_rows(rel_fuel_results_all)
```


# Compare 

```{r}
results_all <- abs_results_all %>%
  bind_rows(rel_results_all)

# Make plots
if(make_plots){
  
  # 1) global (aggregated) timeseries plot
  compare_ts_plot <- BioEconomicModelPlots(dat = results_all,
                                           spatial_res = "global",
                                           temporal_res = "all",
                                           start_year = 2018,
                                           end_year = end_year,
                                           plot_name = "compare_global_ts_plot",
                                           plot_width = 8,
                                           plot_height = 6.5,
                                           results_dir = results_dir,
                                           group_var = "run_group",
                                           color_var = "run_group",
                                           legend_pos_manual = "bottom",
                                           legend_name = "Scenario")
  
  # 2) global (aggregated) "kobe" final year plot
  compare_end_plot <- BioEconomicModelPlots(dat = results_all,
                                            spatial_res = "global",
                                            temporal_res = "last",
                                            end_year = end_year,
                                            plot_name = paste0("compare_global_", end_year, "_plot"),
                                            plot_width = 8,
                                            plot_height = 6.5,
                                            results_dir = results_dir,
                                            group_var = "run_group",
                                            color_var = "run_group",
                                            legend_pos_manual = "bottom",
                                            legend_name = "Scenario")
  
  # 3) regional timeseries plot
  regional_compare_ts_plot <- BioEconomicModelPlots(dat = results_all,
                                                    spatial_res = "regional",
                                                    temporal_res = "all",
                                                    start_year = 2018,
                                                    end_year = end_year,
                                                    plot_name = "compare_regional_ts_plot",
                                                    plot_width = 8,
                                                    plot_height = 6.5,
                                                    results_dir = results_dir,
                                                    group_var = "run_group",
                                                    color_var = "run_group",
                                                    legend_pos_manual = "bottom",
                                                    legend_name = "Scenario")
  
  # 4) regional "kobe" final year plot
  regional_compare_end_plot <- BioEconomicModelPlots(dat = results_all,
                                                     spatial_res = "regional",
                                                     temporal_res = "last",
                                                     end_year = end_year,
                                                     plot_name = paste0("compare_regional_", end_year, "_plot"),
                                                     plot_width = 8,
                                                     plot_height = 6.5,
                                                     results_dir = results_dir,
                                                     group_var = "run_group",
                                                     color_var = "run_group",
                                                     legend_pos_manual = "bottom",
                                                     legend_name = "Scenario")
}
```
```
