==> scripts/08_subs_affected.Rmd <==
---
output:
  html_document: default

title: "SubsidyExplorer - Step 8: Subsidies affected by discipline type"
author: "Kat Millage, Vienna Saccomanno, Laura Lea Rubino, Christopher Costello"
date: "`r format(Sys.time(), '%B %d, %Y')`"
fontsize: 11pt
bibliography: "../pew-subsidies.bib"
---
# Introduction 

This script outputs the amounts of subsidies affected by each discipline type (and/or definition) in SubsidyExplorer. This was requested by Anthony Rogers to aid in his analyses for Pew. 

```{r setup, include=FALSE}
# Load packages
library(knitr) 
library(tidyverse)

knitr::opts_chunk$set(echo = TRUE)

make_plots <- T

if(make_plots){
  library(RColorBrewer)
  library(scales)
}

# Directories for the output files generated by this script
results_dir <- paste0(here::here("results", "08-subs-affected/"))
if (dir.exists(results_dir) == F) {
  dir.create(results_dir, recursive = T)
}

good_subsidy_types <- c("A1", "A2", "A3")
bad_subsidy_types <- c("B1", "B2", "B3", "B4", "B5", "B6", "B7")
ugly_subsidy_types <- c("C1", "C2", "C3")
```

```{r}
# Load vessel list
vessel_dat <- read.csv(here::here("app", "data", "vessel_list_2018_final.csv"), stringsAsFactors = F)

# Create high seas code
vessel_dat <- vessel_dat %>%
  mutate(eez_hs_code = case_when(eez_id == 0 ~ paste0("HS-", fao_region),
                                 TRUE ~ as.character(eez_id)))

```

# IUU

We really only have data on one measure of IUU fishing - vessels identified as IUU from RFMO lists. We do also have subsidies going to vessels that have ever fished in a disputed area flagged. 

```{r}
iuu_cols <- paste0(c(good_subsidy_types, bad_subsidy_types, ugly_subsidy_types), "_subs")

subs_affected_iuu <- vessel_dat %>%
  group_by(iuu1) %>%
  summarize(across(iuu_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "IUU - RFMO Lists") %>%
  relocate(description) %>%
  rename(affected = iuu1)

subs_affected_disputed <- vessel_dat %>%
  group_by(is_disputed) %>%
  summarize(across(iuu_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "IUU - Disputed Area") %>%
  relocate(description) %>%
  rename(affected = is_disputed)

subs_affected_iuu_out <- subs_affected_iuu %>%
  bind_rows(subs_affected_disputed)

write_csv(subs_affected_iuu_out, paste0(results_dir, "subs_affected_iuu_out.csv"))

# Now do the same by country
subs_affected_flag_iuu <- vessel_dat %>%
  group_by(iuu1, flag_iso3) %>%
  summarize(across(iuu_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "IUU - RFMO Lists") %>%
  relocate(description) %>%
  rename(affected = iuu1) %>%
  arrange(flag_iso3)

subs_affected_flag_disputed <- vessel_dat %>%
  group_by(is_disputed, flag_iso3) %>%
  summarize(across(iuu_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "IUU - Disputed Area") %>%
  relocate(description) %>%
  rename(affected = is_disputed) %>%
  arrange(flag_iso3)

subs_affected_flag_iuu_out <- subs_affected_flag_iuu %>%
  bind_rows(subs_affected_flag_disputed)

write_csv(subs_affected_flag_iuu_out, paste0(results_dir, "subs_affected_flag_iuu_out.csv"))
```

# Overfished

Here we could consider multiple metrics... Let's stick with the four from the tool for now, but provide alternative methods (mean/median/weighted mean)

```{r}
oa_cols <- paste0(c(good_subsidy_types, bad_subsidy_types, ugly_subsidy_types), "_subs")

subs_affected_oa1_median <- vessel_dat %>%
  group_by(oa1_median) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Median BvBmsy < 1 (RAM stocks only)") %>%
  relocate(description) %>%
  rename(affected = oa1_median)

subs_affected_oa1_w_mean <- vessel_dat %>%
  group_by(oa1_w_mean) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Weighted Mean BvBmsy < 1 (RAM stocks only)") %>%
  relocate(description) %>%
  rename(affected = oa1_w_mean)

subs_affected_oa2_median <- vessel_dat %>%
  group_by(oa2_median) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Median BvBmsy < 0.8 (RAM stocks only)") %>%
  relocate(description) %>%
  rename(affected = oa2_median)

subs_affected_oa2_w_mean <- vessel_dat %>%
  group_by(oa2_w_mean) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Weighted Mean BvBmsy < 0.8 (RAM stocks only)") %>%
  relocate(description) %>%
  rename(affected = oa2_w_mean)

subs_affected_oa3_median <- vessel_dat %>%
  group_by(oa3_median) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Median BvBmsy < 1 (All Costello et al. 2016 stocks)") %>%
  relocate(description) %>%
  rename(affected = oa3_median)

subs_affected_oa3_w_mean <- vessel_dat %>%
  group_by(oa3_w_mean) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Weighted Mean BvBmsy < 1 (All Costello et al. 2016 stocks)") %>%
  relocate(description) %>%
  rename(affected = oa3_w_mean)

subs_affected_oa4_median <- vessel_dat %>%
  group_by(oa4_median) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Median BvBmsy < 0.8 (All Costello et al. 2016 stocks)") %>%
  relocate(description) %>%
  rename(affected = oa4_median)

subs_affected_oa4_w_mean <- vessel_dat %>%
  group_by(oa4_w_mean) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Weighted Mean BvBmsy < 0.8 (All Costello et al. 2016 stocks)") %>%
  relocate(description) %>%
  rename(affected = oa4_w_mean)

subs_affected_oa_out <- subs_affected_oa1_median %>%
  bind_rows(subs_affected_oa1_w_mean) %>%
  bind_rows(subs_affected_oa2_median) %>%
  bind_rows(subs_affected_oa2_w_mean) %>%
  bind_rows(subs_affected_oa3_median) %>%
  bind_rows(subs_affected_oa3_w_mean) %>%
  bind_rows(subs_affected_oa4_median) %>%
  bind_rows(subs_affected_oa4_w_mean)

write_csv(subs_affected_oa_out, paste0(results_dir, "subs_affected_overfished_out.csv"))
```

```{r}
subs_affected_flag_oa1_median <- vessel_dat %>%
  group_by(oa1_median, flag_iso3) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Median BvBmsy < 1 (RAM stocks only)") %>%
  relocate(description) %>%
  rename(affected = oa1_median) %>%
  arrange(flag_iso3)

subs_affected_flag_oa1_w_mean <- vessel_dat %>%
  group_by(oa1_w_mean, flag_iso3) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Weighted Mean BvBmsy < 1 (RAM stocks only)") %>%
  relocate(description) %>%
  rename(affected = oa1_w_mean) %>%
  arrange(flag_iso3)

subs_affected_flag_oa2_median <- vessel_dat %>%
  group_by(oa2_median, flag_iso3) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Median BvBmsy < 0.8 (RAM stocks only)") %>%
  relocate(description) %>%
  rename(affected = oa2_median) %>%
  arrange(flag_iso3)

subs_affected_flag_oa2_w_mean <- vessel_dat %>%
  group_by(oa2_w_mean, flag_iso3) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Weighted Mean BvBmsy < 0.8 (RAM stocks only)") %>%
  relocate(description) %>%
  rename(affected = oa2_w_mean) %>%
  arrange(flag_iso3)

subs_affected_flag_oa3_median <- vessel_dat %>%
  group_by(oa3_median, flag_iso3) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Median BvBmsy < 1 (All Costello et al. 2016 stocks)") %>%
  relocate(description) %>%
  rename(affected = oa3_median) %>%
  arrange(flag_iso3)

subs_affected_flag_oa3_w_mean <- vessel_dat %>%
  group_by(oa3_w_mean, flag_iso3) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Weighted Mean BvBmsy < 1 (All Costello et al. 2016 stocks)") %>%
  relocate(description) %>%
  rename(affected = oa3_w_mean) %>%
  arrange(flag_iso3)

subs_affected_flag_oa4_median <- vessel_dat %>%
  group_by(oa4_median, flag_iso3) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Median BvBmsy < 0.8 (All Costello et al. 2016 stocks)") %>%
  relocate(description) %>%
  rename(affected = oa4_median) %>%
  arrange(flag_iso3)

subs_affected_flag_oa4_w_mean <- vessel_dat %>%
  group_by(oa4_w_mean, flag_iso3) %>%
  summarize(across(oa_cols, ~ sum(.x, na.rm = T))) %>%
  mutate(description = "Overfished - Weighted Mean BvBmsy < 0.8 (All Costello et al. 2016 stocks)") %>%
  relocate(description) %>%
  rename(affected = oa4_w_mean) %>%
  arrange(flag_iso3)

subs_affected_flag_oa_out <- subs_affected_flag_oa1_median %>%
  bind_rows(subs_affected_flag_oa1_w_mean) %>%
  bind_rows(subs_affected_flag_oa2_median) %>%
  bind_rows(subs_affected_flag_oa2_w_mean) %>%
  bind_rows(subs_affected_flag_oa3_median) %>%
  bind_rows(subs_affected_flag_oa3_w_mean) %>%
  bind_rows(subs_affected_flag_oa4_median) %>%
  bind_rows(subs_affected_flag_oa4_w_mean)

write_csv(subs_affected_flag_oa_out, paste0(results_dir, "subs_affected_flag_overfished_out.csv"))
```
