==> scripts/10_high_seas_vessel_breakdown.Rmd <==
---
output:
  html_document: default

title: "SubsidyExplorer - Step 10: Exploring high seas vessels"
author: "Kat Millage, Vienna Saccomanno, Laura Lea Rubino, Christopher Costello"
date: "`r format(Sys.time(), '%B %d, %Y')`"
fontsize: 11pt
bibliography: "../pew-subsidies.bib"
---

# Introduction 

This script explores a couple of different ways of differentiating between high seas vessels in order to explore possible ways of refining the high seas assumption for the final agreement. The following breakdowns are considered: 
- Gear type
- Size class
- GT class
- Engine power class

```{r setup, include=FALSE}
# Load packages
library(knitr) 
library(tidyverse)
library(countrycode)

knitr::opts_chunk$set(echo = TRUE)

make_plots <- T

if(make_plots){
  library(RColorBrewer)
  library(scales)
}

bad_subsidy_types <- c("B1", "B2", "B3", "B4", "B5", "B6", "B7")
ugly_subsidy_types <- c("C1", "C2", "C3")
good_subsidy_types <- c("A1", "A2", "A3")

subsidy_types_sorted_sumaila <- c(good_subsidy_types, bad_subsidy_types, ugly_subsidy_types)

# # Threshold to be considered managed (using FMI scores)
# managed_cutoff <- 0.5
# end_year <- 2080
# folder_name <- paste0("mgmt_threshold_", managed_cutoff, "_ALL/")

# Directories for the output files generated by this script
results_dir <- paste0(here::here("results", "10-high-seas-vessel-breakdowns/"))
if (dir.exists(results_dir) == F) {
  dir.create(results_dir, recursive = T)
}
```

```{r}
# # Source model script
# source(here::here("app", "scripts", "BioEconModel.R"))
# 
# # Source create fleets script
# source(here::here("app", "scripts", "CreateFleets.R"))
# 
# # Source plotting script
# source(here::here("scripts", "functions", "BioEconomicModelPlots.R"))
```

```{r}
# Load country dependencies helper file
country_dependencies <- read_csv(here::here("app", "data", "country_dependencies.csv")) %>%
   mutate(display_name = case_when(!is.na(WTO_name) ~ WTO_name,
                                  iso3 == "ANT" ~ "Netherlands Antilles",
                                  iso3 == "ASC" ~ "Ascension Island",
                                  iso3 == "CPT" ~ "Clipperton Island",
                                  iso3 == "TAA" ~ "Tristan da Cunha",
                                  TRUE ~ countrycode(iso3, "iso3c", "country.name"))) %>%
  arrange(sovereign_iso3)

# Vector of all WTO Members and Observers for use in the app [does not include the overseas territories associated with them]
wto_members_and_observers <- country_dependencies$iso3[country_dependencies$is_WTO & !country_dependencies$is_overseas_territory]
```

```{r}
# Load vessel list
vessel_dat_raw <- read.csv(here::here("app", "data", "vessel_list_2018_final.csv"), stringsAsFactors = F) %>%
  left_join(country_dependencies %>% dplyr::select(iso3, is_WTO), by = c("flag_iso3" = "iso3"))

rel <- read_csv(here::here("data", "lookup-tables", "oecd_relative_effects_lookup_table.csv"))

vessel_dat <- vessel_dat_raw %>%
   mutate(B1_subs = B1_subs * rel$rel_effect_effort[rel$type == "B1"], 
           B2_subs = B2_subs * rel$rel_effect_effort[rel$type == "B2"],
           B3_subs = B3_subs * rel$rel_effect_effort[rel$type == "B3"],
           B4_subs = B4_subs * rel$rel_effect_effort[rel$type == "B4"],
           B5_subs = B5_subs * rel$rel_effect_effort[rel$type == "B5"], 
           B6_subs = B6_subs * rel$rel_effect_effort[rel$type == "B6"],
           B7_subs = B7_subs * rel$rel_effect_effort[rel$type == "B7"], 
           C1_subs = C1_subs * rel$rel_effect_effort[rel$type == "C1"], 
           C2_subs = C2_subs * rel$rel_effect_effort[rel$type == "C2"],
           C3_subs = C3_subs * rel$rel_effect_effort[rel$type == "C3"]) %>%
  ungroup() %>%
  mutate(bad_subs = rowSums(select(., one_of(paste0(bad_subsidy_types, "_subs")))),
         ugly_subs = rowSums(select(., one_of(paste0(ugly_subsidy_types, "_subs")))),
         good_subs = rowSums(select(., one_of(paste0(good_subsidy_types, "_subs")))),
         bad_ugly_subs = rowSums(select(., one_of(paste0(c(bad_subsidy_types, ugly_subsidy_types), "_subs")))))

# Create high seas code
vessel_dat <- vessel_dat %>%
  mutate(eez_hs_code = case_when(eez_id == 0 ~ paste0("HS-", fao_region),
                                 TRUE ~ as.character(eez_id)))
```

## Get high seas vessels (at least 5% of annual effort on the high seas)

```{r}
hs_vessel_dat <- vessel_dat_raw %>%
  dplyr::filter(prop_fishing_KWh_high_seas >= 0.05)
```

## Gear type breakdown

```{r}
gear_type_breakdown <- hs_vessel_dat %>%
  group_by(vessel_class) %>%
  summarize(bad_subs = sum(bad_subs, na.rm = T),
            n_vessels = n_distinct(ssvid))

gear_type_breakdown_plot <- gear_type_breakdown %>%
  ggplot()+
  aes(x = vessel_class, y = bad_subs/1e9)+
  geom_bar(stat = "identity")+
  theme_linedraw()+
  coord_flip()+
  labs(x = "Gear type",
       y = "Capacity-enhancing subsidies ($, billions)")

write_csv(gear_type_breakdown, paste0(results_dir, "hs_gear_type_breakdown.csv"))
ggsave(paste0(results_dir, "hs_gear_type_breakdown.png"), gear_type_breakdown_plot, width = 7.5, height = 6)
```

## Length breakdown

```{r}
length_breakdown <- hs_vessel_dat %>%
  mutate(bin = cut(length_m, seq(min(length_m), max(length_m) + 10, 10))) %>%
  group_by(bin) %>%
  summarize(bad_subs = sum(bad_subs, na.rm = T),
            n_vessels = n_distinct(ssvid))

length_breakdown_plot <- length_breakdown %>%
  ggplot()+
  aes(x = bin, y = bad_subs/1e9)+
  geom_bar(stat = "identity")+
  theme_linedraw()+
  coord_flip()+
  labs(x = "Total length (m)",
       y = "Capacity-enhancing subsidies ($, billions)")

write_csv(length_breakdown, paste0(results_dir, "hs_length_breakdown.csv"))
ggsave(paste0(results_dir, "hs_length_breakdown.png"), length_breakdown_plot, width = 7.5, height = 6)
```

## Tonnage breakdown

```{r}
tonnage_breakdown <- hs_vessel_dat %>%
  mutate(bin = cut(tonnage_gt, seq(min(tonnage_gt), max(tonnage_gt) + 100, 100))) %>%
  group_by(bin) %>%
  summarize(bad_subs = sum(bad_subs, na.rm = T),
            n_vessels = n_distinct(ssvid))

tonnage_breakdown_plot <- tonnage_breakdown %>%
  ggplot()+
  aes(x = bin, y = bad_subs/1e9)+
  geom_bar(stat = "identity")+
  theme_linedraw()+
  coord_flip()+
  labs(x = "Tonnage (gt)",
       y = "Capacity-enhancing subsidies ($, billions)")

write_csv(tonnage_breakdown, paste0(results_dir, "hs_tonnage_breakdown.csv"))
ggsave(paste0(results_dir, "hs_tonnage_breakdown.png"), tonnage_breakdown_plot, width = 7.5, height = 8)
```

## Engine power breakdown

```{r}
engine_power_breakdown <- hs_vessel_dat %>%
  mutate(bin = cut(engine_power_kw, seq(min(engine_power_kw), max(engine_power_kw) + 100, 100))) %>%
  group_by(bin) %>%
  summarize(bad_subs = sum(bad_subs, na.rm = T),
            n_vessels = n_distinct(ssvid))

engine_power_breakdown_plot <- engine_power_breakdown %>%
  ggplot()+
  aes(x = bin, y = bad_subs/1e9)+
  geom_bar(stat = "identity")+
  theme_linedraw()+
  coord_flip()+
  labs(x = "Engine power (kW)",
       y = "Capacity-enhancing subsidies ($, billions)")

write_csv(engine_power_breakdown, paste0(results_dir, "hs_engine_power_breakdown.csv"))
ggsave(paste0(results_dir, "hs_engine_power_breakdown.png"), engine_power_breakdown_plot, width = 7.5, height = 8)
```

<!-- ```{r} -->
<!-- # Load biological parameters for the model -->
<!-- bio_dat_regional <- read.csv(here::here("app", "data", "model_parameters_regional_3_regions.csv")) -->

<!-- # Regional parameter list -->
<!-- bio_dat_regional_list <- bio_dat_regional %>% -->
<!--   gather(region, value, -c(1:2)) %>% -->
<!--   group_by(region) %>% -->
<!--   group_split() -->
<!-- ``` -->

<!-- ```{r} -->
<!-- # Load proposal settings -->
<!-- proposal_settings <- read.csv(here::here("app", "data", "wto_proposal_settings_FINAL_AGREEMENT.csv"), stringsAsFactors = F) %>% -->
<!--   dplyr::filter(include != "Hide") -->

<!-- # Filter to final list of proposals to be included in database -->
<!-- included_proposals <- proposal_settings %>% -->
<!--   dplyr::filter(include %in% c("Yes", "Alternate")) %>% -->
<!--   mutate(display_name = case_when(proposal == "Default" ~ "None", -->
<!--                                   TRUE ~ title_tool)) %>% -->
<!--   dplyr::select(category, proposal, display_name) %>% -->
<!--   arrange(category) %>% -->
<!--   dplyr::filter(!(proposal %in% c("RD/TN/RL/126/Rev.2", "RD/TN/RL/119/Rev.1"))) %>% # remove blank entries created for tool -->
<!--   dplyr::filter(proposal != "Default") -->

<!-- proposal_choices <- included_proposals$proposal -->
<!-- names(proposal_choices) <- included_proposals$display_name -->
<!-- ``` -->

<!-- ```{r} -->
<!-- # Load cap/tier lookup table -->
<!-- cap_tier_lookup_table <- read_csv(here::here("app","data","cap_tier_lookup_table.csv")) -->

<!-- # Load S&DT lookup table -->
<!-- sdt_lookup_table <- read_csv(here::here("app", "data", "sdt_lookup_table.csv")) -->
<!-- ``` -->
<!-- ## Model proposals -->

<!-- ```{r} -->
<!-- proposal_results <- NULL -->
<!-- affected_subs <- NULL -->

<!-- for(i in 1:length(proposal_choices)){ -->

<!--   # Choose proposal -->
<!--   proposal_name <- proposal_choices[i] -->

<!--   print(proposal_name) -->

<!--   # Get all data for that proposal -->
<!--   selected_proposal <- proposal_settings %>% -->
<!--       dplyr::filter(proposal == proposal_name) -->

<!--   # Get all IUU conditions -->
<!--   iuu <-  -->
<!--     list("definitions" = unlist(str_split(selected_proposal$iuu_definitions, ", ")), -->
<!--          "assumption" = selected_proposal$iuu_assumption, -->
<!--          "percent" = selected_proposal$iuu_percent, -->
<!--          "scope" = selected_proposal$iuu_scope, -->
<!--          "scope_select" = unlist(str_split(selected_proposal$iuu_scope_select, ", ")), -->
<!--          "scope_manual" = unlist(str_split(selected_proposal$iuu_scope_manual, ", ")), -->
<!--          "allow_sdt" = selected_proposal$iuu_allow_sdt, -->
<!--          "sdt_ldc" = selected_proposal$iuu_sdt_ldc, -->
<!--          "sdt_what_ldc" = unlist(str_split(selected_proposal$iuu_sdt_what_ldc, ", ")), -->
<!--          "sdt_time_delay_ldc" = selected_proposal$iuu_sdt_time_delay_ldc, -->
<!--          "sdt_developing" = selected_proposal$iuu_sdt_developing, -->
<!--          "sdt_what_developing" = unlist(str_split(selected_proposal$iuu_sdt_what_developing, ", ")), -->
<!--          "sdt_time_delay_developing" = selected_proposal$iuu_sdt_time_delay_developing, -->
<!--          "sdt_sve" = selected_proposal$iuu_sdt_sve, -->
<!--          "sdt_what_sve" = unlist(str_split(selected_proposal$iuu_sdt_what_sve, ", ")), -->
<!--          "sdt_time_delay_sve" = selected_proposal$iuu_sdt_time_delay_sve) -->

<!--    # Get all OA conditions -->
<!--    oa <-  -->
<!--      list("definitions" = unlist(str_split(selected_proposal$oa_definitions, ", ")), -->
<!--           "scope" = selected_proposal$oa_scope, -->
<!--           "scope_select" = unlist(str_split(selected_proposal$oa_scope_select, ", ")), -->
<!--           "scope_manual" = unlist(str_split(selected_proposal$oa_scope_manual, ", ")), -->
<!--           "hs_cutoff" = selected_proposal$oa_hs_cutoff, -->
<!--           "length_cutoff" = selected_proposal$oa_length_cutoff, -->
<!--           "tonnage_cutoff" = selected_proposal$oa_tonnage_cutoff, -->
<!--           "engine_cutoff" = selected_proposal$oa_engine_cutoff, -->
<!--           "allow_sdt" = selected_proposal$oa_allow_sdt, -->
<!--           "sdt_ldc" = selected_proposal$oa_sdt_ldc, -->
<!--           "sdt_what_ldc" = unlist(str_split(selected_proposal$oa_sdt_what_ldc, ", ")), -->
<!--           "sdt_hs_cutoff_ldc" = selected_proposal$oa_sdt_hs_cutoff_ldc, -->
<!--           "sdt_time_delay_ldc" = selected_proposal$oa_sdt_time_delay_ldc, -->
<!--           "sdt_developing" = selected_proposal$oa_sdt_developing, -->
<!--           "sdt_what_developing" = unlist(str_split(selected_proposal$oa_sdt_what_developing, ", ")), -->
<!--           "sdt_hs_cutoff_developing" = selected_proposal$oa_sdt_hs_cutoff_developing, -->
<!--           "sdt_time_delay_developing" = selected_proposal$oa_sdt_time_delay_developing, -->
<!--           "sdt_sve" = selected_proposal$oa_sdt_sve, -->
<!--           "sdt_what_sve" = unlist(str_split(selected_proposal$oa_sdt_what_sve, ", ")), -->
<!--           "sdt_hs_cutoff_sve" = selected_proposal$oa_sdt_hs_cutoff_sve, -->
<!--           "sdt_time_delay_sve" = selected_proposal$oa_sdt_time_delay_sve) -->

<!--    # Get all Overcap conditions -->
<!--    overcap <-  -->
<!--      list("definitions" = unlist(str_split(selected_proposal$overcap_definitions, ", ")), -->
<!--           "scope" = selected_proposal$overcap_scope, -->
<!--           "scope_select" = unlist(str_split(selected_proposal$overcap_scope_select, ", ")), -->
<!--           "scope_manual" = unlist(str_split(selected_proposal$overcap_scope_manual, ", ")), -->
<!--           "hs_cutoff" = selected_proposal$overcap_hs_cutoff, -->
<!--           "length_cutoff" = selected_proposal$overcap_length_cutoff, -->
<!--           "tonnage_cutoff" = selected_proposal$overcap_tonnage_cutoff, -->
<!--           "engine_cutoff" = selected_proposal$overcap_engine_cutoff, -->
<!--           "allow_sdt" = selected_proposal$overcap_allow_sdt, -->
<!--           "sdt_ldc" = selected_proposal$overcap_sdt_ldc, -->
<!--           "sdt_what_ldc" = unlist(str_split(selected_proposal$overcap_sdt_what_ldc, ", ")), -->
<!--           "sdt_hs_cutoff_ldc" = selected_proposal$overcap_sdt_hs_cutoff_ldc, -->
<!--           "sdt_time_delay_ldc" = selected_proposal$overcap_sdt_time_delay_ldc, -->
<!--           "sdt_developing" = selected_proposal$overcap_sdt_developing, -->
<!--           "sdt_what_developing" = unlist(str_split(selected_proposal$overcap_sdt_what_developing, ", ")), -->
<!--           "sdt_hs_cutoff_developing" = selected_proposal$overcap_sdt_hs_cutoff_developing, -->
<!--           "sdt_time_delay_developing" = selected_proposal$overcap_sdt_time_delay_developing, -->
<!--           "sdt_sve" = selected_proposal$overcap_sdt_sve, -->
<!--           "sdt_what_sve" = unlist(str_split(selected_proposal$overcap_sdt_what_sve, ", ")), -->
<!--           "sdt_hs_cutoff_sve" = selected_proposal$overcap_sdt_hs_cutoff_sve, -->
<!--           "sdt_time_delay_sve" = selected_proposal$overcap_sdt_time_delay_sve) -->

<!--    # Get all Cap/Tier conditions -->
<!--    cap_tier =  -->
<!--      list("on_off" = selected_proposal$cap_on_off, -->
<!--           "subsidy_types" = unlist(str_split(selected_proposal$cap_subsidy_types, ", ")), -->
<!--           "tier_number" = selected_proposal$cap_tier_number, -->
<!--           "tier_system" = selected_proposal$tier_system, -->
<!--           "two_tier_cutoff" = selected_proposal$two_tier_cutoff, -->
<!--           "three_tier_cutoff" = as.numeric(unlist(str_split(selected_proposal$three_tier_cutoff, ", "))), -->
<!--           "tier1_cap_rule" = selected_proposal$tier1_cap_rule, -->
<!--           "tier2_cap_rule" = selected_proposal$tier2_cap_rule, -->
<!--           "tier3_cap_rule" = selected_proposal$tier3_cap_rule, -->
<!--           "tier1_cap_value" = selected_proposal$tier1_cap_value, -->
<!--           "tier1_cap_percent" = selected_proposal$tier1_cap_percent, -->
<!--           "tier1_cap_best_percent_subs" = selected_proposal$tier1_cap_best_percent_subs, -->
<!--           "tier1_cap_best_percent_landed_value" = selected_proposal$tier1_cap_best_percent_landed_value, -->
<!--           "tier1_cap_best_percent_fishers" = selected_proposal$tier1_cap_best_percent_fishers, -->
<!--           "tier2_cap_value" = selected_proposal$tier2_cap_value, -->
<!--           "tier2_cap_percent" = selected_proposal$tier2_cap_percent, -->
<!--           "tier2_cap_best_percent_subs" = selected_proposal$tier2_cap_best_percent_subs, -->
<!--           "tier2_cap_best_percent_landed_value" = selected_proposal$tier2_cap_best_percent_landed_value, -->
<!--           "tier2_cap_best_percent_fishers" = selected_proposal$tier2_cap_best_percent_fishers, -->
<!--           "tier3_cap_value" = selected_proposal$tier3_cap_value, -->
<!--           "tier3_cap_percent" = selected_proposal$tier3_cap_percent, -->
<!--           "tier3_cap_best_percent_subs" = selected_proposal$tier3_cap_best_percent_subs, -->
<!--           "tier3_cap_best_percent_landed_value" = selected_proposal$tier3_cap_best_percent_landed_value, -->
<!--           "tier3_cap_best_percent_fishers" = selected_proposal$tier3_cap_best_percent_fishers) -->

<!--    # Find fleets -->
<!--     fleet <-  CreateFleets( -->
<!--       vessel_list = vessel_dat, -->
<!--       iuu = iuu, -->
<!--       oa = oa, -->
<!--       overcap = overcap, -->
<!--       cap_tier = cap_tier, -->
<!--       managed_threshold = managed_cutoff, -->
<!--       subsidy_types_all = subsidy_types_sorted_sumaila, -->
<!--       cap_tier_lookup = cap_tier_lookup_table, -->
<!--       country_lookup = country_dependencies, -->
<!--       sdt_lookup = sdt_lookup_table) -->

<!--     # Make list of fleets for bioeconomic model -->
<!--     fleet_list <- fleet$summary %>% -->
<!--       group_by(region) %>% -->
<!--       group_split() -->
<!--     names(fleet_list) <- colnames(bio_dat_regional)[-c(1:2)] -->

<!--     # Run model -->
<!--     results <- pmap_df(list(fleet = fleet_list, -->
<!--                             region = names(fleet_list), -->
<!--                             bio_param = bio_dat_regional_list), -->
<!--                        BioEconModel, -->
<!--                        end_year = end_year, -->
<!--                        return = "all") -->

<!--     # Extract model timeseries results -->
<!--     results_ts <- results %>% -->
<!--       dplyr::filter(Year > 2018) %>% -->
<!--       group_by(Year, Variable, Fleet) %>% -->
<!--       mutate(Diff = case_when(BAU != 0 ~ (Reform - BAU) / abs(BAU), -->
<!--       TRUE ~ 0)) %>% -->
<!--       ungroup() %>% -->
<!--       mutate(run_id = proposal_name, -->
<!--              run_name = names(proposal_name), -->
<!--              run_group = unique(proposal_settings$category[proposal_settings$proposal == proposal_name])) -->

<!--       write_csv(results_ts, paste0(results_dir, str_replace_all(names(proposal_name), "/", "_"), ".csv")) -->

<!--     # Add to data frame -->
<!--       proposal_results <- proposal_results %>% -->
<!--         bind_rows(results_ts) -->

<!--     # Extract affected/unaffected subsidies amounts and convert back to absolute dollars -->
<!--       affected_subs_raw <- fleet$vessels %>% -->
<!--         group_by(region) %>% -->
<!--         summarize_at(vars(A1_subs:C3_subs, A1_subs_removed:C3_subs_removed), sum, na.rm = T) %>% -->
<!--         gather("type", "value", -region) %>% -->
<!--         mutate(category = ifelse(grepl("removed", type), "removed", "base")) %>% -->
<!--         mutate(type = str_replace(type, "_subs_removed", "")) %>% -->
<!--         mutate(type = str_replace(type, "_subs", "")) %>% -->
<!--         left_join(rel, by = "type") %>% -->
<!--         mutate(rel_effect_effort = ifelse(is.na(rel_effect_effort), 1, rel_effect_effort)) %>% -->
<!--         mutate(value_real = value/rel_effect_effort) %>% -->
<!--         group_by(region, category) %>% -->
<!--         summarize(value_real = sum(value_real, na.rm = T)) %>% -->
<!--         ungroup() %>% -->
<!--         group_by(region) %>% -->
<!--         summarize(subs_base = value_real[category == "base"], -->
<!--                   subs_affected = value_real[category == "removed"], -->
<!--                   percent_subs_affected = (value_real[category == "removed"]/value_real[category == "base"])*100) -->

<!--       affected_subs_out <- data_frame(run_id = rep(proposal_name, length.out = 3), -->
<!--                                       run_name = rep(names(proposal_name), length.out = 3), -->
<!--                                       run_group = rep(unique(proposal_settings$category[proposal_settings$proposal == proposal_name]), length.out = 3), -->
<!--                                       region = c("atlantic", "indian", "pacific")) %>% -->
<!--         left_join(affected_subs_raw, by = "region") -->

<!--       # Add to data frame -->
<!--       affected_subs <- affected_subs %>% -->
<!--         bind_rows(affected_subs_out) -->

<!-- } -->
<!-- ``` -->

<!-- Add in most ambitious result and combine into single dataset -->

<!-- ```{r} -->
<!-- # Load most ambitous results from previous script -->
<!-- most_ambitious_results <- read_csv(paste0(here::here("results", "06-run-model/"), "mgmt_threshold_", managed_cutoff, "/", "BC_rel_ts_data.csv")) %>% -->
<!--   mutate(run_id = "Ambitious Reform", -->
<!--          run_group = "Ambitious Reform") -->

<!-- # Combine with ts data -->
<!-- all_ts_results <- proposal_results %>% -->
<!--   bind_rows(most_ambitious_results) -->

<!-- # Get subsidies affected for most ambitious scenario -->
<!-- most_ambitious_subs_affected <- vessel_dat_raw %>% -->
<!--   mutate(subs_base = bad_subs + good_subs + ugly_subs, -->
<!--          subs_affected = bad_subs + ugly_subs) %>% -->
<!--   group_by(region) %>% -->
<!--   summarize(subs_base = sum(subs_base, na.rm = T), -->
<!--             subs_affected = sum(subs_affected, na.rm = T)) %>% -->
<!--   mutate(percent_subs_affected = (subs_affected/subs_base)*100) -->

<!-- # Combine with subs affected data -->
<!-- all_subs_affected <- data_frame(run_id = rep("Ambitious Reform", 3), -->
<!--                                 run_name = rep("All bad and ugly subsidies", 3), -->
<!--                                 run_group = rep("Ambitious Reform", 3)) %>% -->
<!--   bind_cols(most_ambitious_subs_affected) %>% -->
<!--   bind_rows(affected_subs) -->
<!-- ``` -->

<!-- ## Make nice databases (global and regional) -->

<!-- ```{r} -->
<!-- # We want these to be ordered by date within each category, so we need to go back to our list -->
<!-- date_order <- proposal_settings %>% -->
<!--   dplyr::select(title_tool, date) %>% -->
<!--   rownames_to_column("order") %>% -->
<!--   mutate(order = as.numeric(order)) -->

<!-- # Collect global results -->
<!-- all_results_global_output_table <- all_ts_results %>% -->
<!--   dplyr::filter(Year %in% c(2050, end_year)) %>% -->
<!--   dplyr::filter(Variable %in% c("biomass", "catches_total", "revenue_total", "u_mort_total")) %>% -->
<!--   group_by(Year, Variable, Fleet, run_id, run_name, run_group) %>% -->
<!--   summarize(BAU_tot = sum(BAU, na.rm = T), -->
<!--             Reform_tot = sum(Reform, na.rm = T), -->
<!--             Diff_tot = case_when(BAU_tot != 0 ~ (Reform_tot - BAU_tot)/abs(BAU_tot), -->
<!--                             TRUE ~ 0)) %>% -->
<!--   ungroup() %>% -->
<!--   mutate(Variable = case_when(Variable == "biomass" ~ "Biomass", -->
<!--                                 Variable == "catches_total" ~ "Catches", -->
<!--                                 Variable == "revenue_total" ~ "Revenue", -->
<!--                                 Variable == "u_mort_total" ~ "Fishing Mortality")) %>% -->
<!--   dplyr::select(Id = run_id, -->
<!--                 Name = run_name, -->
<!--                 Category = run_group,  -->
<!--                 Year, -->
<!--                 Diff_tot, -->
<!--                 Variable) %>% -->
<!--   spread(Variable, Diff_tot) %>% -->
<!--   left_join(date_order %>% dplyr::select(Name = title_tool, order), by = "Name") %>% -->
<!--   mutate(Category = factor(Category, levels = c("Ambitious Reform", "IUU", "OFOC", "Overfished", "Other & Multiple"))) %>% -->
<!--   arrange(Category, order, Year) %>% -->
<!--   dplyr::select(-order) -->

<!-- write.csv(all_results_global_output_table, paste0(results_dir, "all_results_global_output_table.csv")) -->

<!-- # Now make a pretty version for 2050 only -->
<!-- all_results_global_output_table_2050_pretty <- all_results_global_output_table %>% -->
<!--   dplyr::filter(Year == 2050) %>% -->
<!--   ungroup() %>% -->
<!--   mutate(Biomass = round(Biomass*100, 2), -->
<!--          Catches = round(Catches*100, 2), -->
<!--          Revenue = round(Revenue*100, 2), -->
<!--          `Fishing Mortality` = round(`Fishing Mortality`*100, 2)) %>% -->
<!--   rename(`Change in biomass (%, 2050)` = Biomass, -->
<!--          `Change in catch (%, 2050)` = Catches, -->
<!--          `Change in revenue (%, 2050)` = Revenue, -->
<!--          `Change in fishing mortality (%, 2050)` = `Fishing Mortality`) %>% -->
<!--   dplyr::select(-Year) -->

<!-- write.csv(all_results_global_output_table_2050_pretty, paste0(results_dir, "all_results_global_output_table_2050_pretty.csv")) -->

<!-- # Now make a pretty version for 2080 only -->
<!-- all_results_global_output_table_2080_pretty <- all_results_global_output_table %>% -->
<!--   dplyr::filter(Year == 2080) %>% -->
<!--   ungroup() %>% -->
<!--   mutate(Biomass = round(Biomass*100, 2), -->
<!--          Catches = round(Catches*100, 2), -->
<!--          Revenue = round(Revenue*100, 2), -->
<!--          `Fishing Mortality` = round(`Fishing Mortality`*100, 2)) %>% -->
<!--   rename(`Change in biomass (%, 2080)` = Biomass, -->
<!--          `Change in catch (%, 2080)` = Catches, -->
<!--          `Change in revenue (%, 2080)` = Revenue, -->
<!--          `Change in fishing mortality (%, 2080)` = `Fishing Mortality`) %>% -->
<!--   dplyr::select(-Year) -->

<!-- write.csv(all_results_global_output_table_2080_pretty, paste0(results_dir, "all_results_global_output_table_2080_pretty.csv")) -->

<!-- # Combine -->
<!-- all_results_global_output_table_pretty <- all_results_global_output_table_2050_pretty %>% -->
<!--   left_join(all_results_global_output_table_2080_pretty %>% dplyr::select(-Name, -Category)) -->

<!-- write.csv(all_results_global_output_table_pretty, paste0(results_dir, "all_results_global_output_table_pretty.csv")) -->

<!-- # Now let's get affected subs globally -->
<!-- all_subs_affected_global <- all_subs_affected %>% -->
<!--   group_by(run_id, run_name, run_group) %>% -->
<!--   summarize(subs_base = sum(subs_base), -->
<!--             subs_affected = sum(subs_affected)) %>% -->
<!--   ungroup() %>% -->
<!--   mutate(percent_subs_affected = round((subs_affected/subs_base)*100, 2)) -->

<!-- # And add to table -->
<!-- global_database <- all_results_global_output_table_pretty %>% -->
<!--   left_join(all_subs_affected_global %>% dplyr::select(Id = run_id, subs_affected, percent_subs_affected)) %>% -->
<!--   rename(`Subsidies affected ($)` = subs_affected, -->
<!--          `Subsidies affected (%)` = percent_subs_affected) -->

<!-- write.csv(global_database, paste0(results_dir, "SubsidyExplorer_proposal_results_global.csv")) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- # Now do regional -->
<!-- all_results_regional_output_table <- all_ts_results %>% -->
<!--   dplyr::filter(Year %in% c(2050, end_year)) %>% -->
<!--   dplyr::filter(Variable %in% c("biomass", "catches_total", "revenue_total", "u_mort_total")) %>% -->
<!--   mutate(Variable = case_when(Variable == "biomass" ~ "Biomass", -->
<!--                                 Variable == "catches_total" ~ "Catches", -->
<!--                                 Variable == "revenue_total" ~ "Revenue", -->
<!--                                 Variable == "u_mort_total" ~ "Fishing Mortality")) %>% -->
<!--   dplyr::select(Id = run_id, -->
<!--                 Name = run_name, -->
<!--                 Category = run_group,  -->
<!--                 Year, -->
<!--                 Region, -->
<!--                 Diff, -->
<!--                 Variable) %>% -->
<!--   spread(Variable, Diff) %>% -->
<!--   left_join(date_order %>% dplyr::select(Name = title_tool, order), by = "Name") %>% -->
<!--   mutate(Category = factor(Category, levels = c("Ambitious Reform", "IUU", "OFOC", "Overfished", "Other & Multiple"))) %>% -->
<!--   arrange(Region, Category, order, Year) %>% -->
<!--   dplyr::select(-order) -->

<!-- write.csv(all_results_regional_output_table, paste0(results_dir, "all_results_regional_output_table.csv")) -->

<!-- # Now make a pretty version for 2050 only -->
<!-- all_results_regional_output_table_2050_pretty <- all_results_regional_output_table %>% -->
<!--   dplyr::filter(Year == 2050) %>% -->
<!--   ungroup() %>% -->
<!--   mutate(Biomass = round(Biomass*100, 2), -->
<!--          Catches = round(Catches*100, 2), -->
<!--          Revenue = round(Revenue*100, 2), -->
<!--          `Fishing Mortality` = round(`Fishing Mortality`*100, 2)) %>% -->
<!--   rename(`Change in biomass (%, 2050)` = Biomass, -->
<!--          `Change in catch (%, 2050)` = Catches, -->
<!--          `Change in revenue (%, 2050)` = Revenue, -->
<!--          `Change in fishing mortality (%, 2050)` = `Fishing Mortality`) %>% -->
<!--   dplyr::select(-Year) %>% -->
<!--   mutate(Region = case_when(Region == "atlantic" ~ "Atlantic Ocean", -->
<!--                             Region == "indian" ~ "Indian Ocean", -->
<!--                             Region == "pacific" ~ "Pacific Ocean")) -->

<!-- write.csv(all_results_regional_output_table_2050_pretty, paste0(results_dir, "all_results_regional_output_table_2050_pretty.csv")) -->

<!-- # Now make a pretty version for 2080 only -->
<!-- all_results_regional_output_table_2080_pretty <- all_results_regional_output_table %>% -->
<!--   dplyr::filter(Year == 2080) %>% -->
<!--   ungroup() %>% -->
<!--   mutate(Biomass = round(Biomass*100, 2), -->
<!--          Catches = round(Catches*100, 2), -->
<!--          Revenue = round(Revenue*100, 2), -->
<!--          `Fishing Mortality` = round(`Fishing Mortality`*100, 2)) %>% -->
<!--   rename(`Change in biomass (%, 2080)` = Biomass, -->
<!--          `Change in catch (%, 2080)` = Catches, -->
<!--          `Change in revenue (%, 2080)` = Revenue, -->
<!--          `Change in fishing mortality (%, 2080)` = `Fishing Mortality`) %>% -->
<!--   dplyr::select(-Year) %>% -->
<!--   mutate(Region = case_when(Region == "atlantic" ~ "Atlantic Ocean", -->
<!--                             Region == "indian" ~ "Indian Ocean", -->
<!--                             Region == "pacific" ~ "Pacific Ocean")) -->

<!-- write.csv(all_results_regional_output_table_2080_pretty, paste0(results_dir, "all_results_regional_output_table_2080_pretty.csv")) -->

<!-- # Combine -->
<!-- all_results_regional_output_table_pretty <- all_results_regional_output_table_2050_pretty %>% -->
<!--   left_join(all_results_regional_output_table_2080_pretty %>% dplyr::select(-Name, -Category)) -->

<!-- write.csv(all_results_regional_output_table_pretty, paste0(results_dir, "all_results_regional_output_table_pretty.csv")) -->

<!-- # And add affected subsidies to table -->
<!-- regional_database <- all_results_regional_output_table_pretty %>% -->
<!--   left_join(all_subs_affected %>% -->
<!--               mutate(region = case_when(region == "atlantic" ~ "Atlantic Ocean", -->
<!--                                         region == "indian" ~ "Indian Ocean", -->
<!--                                         region == "pacific" ~ "Pacific Ocean")) %>% -->
<!--               mutate(percent_subs_affected = round(percent_subs_affected, 2)) %>% -->
<!--               dplyr::select(Id = run_id, Region = region, subs_affected, percent_subs_affected)) %>% -->
<!--   rename(`Subsidies affected ($)` = subs_affected, -->
<!--          `Subsidies affected (%)` = percent_subs_affected) -->

<!-- write.csv(regional_database, paste0(results_dir, "SubsidyExplorer_proposal_results_regional.csv")) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- # # Now make a min, max, and average for each proposal type -->
<!-- # summary_results_global_2050_pretty <- all_results_global_output_table_2050_pretty %>% -->
<!-- #   gather(Variable, Value, c("Biomass", "Catches", "Revenue", "Fishing Mortality")) %>% -->
<!-- #   group_by(Category, Year, Variable) %>% -->
<!-- #   summarize(Minimum = min(Value, na.rm = T), -->
<!-- #             Maximum = max(Value, na.rm = T), -->
<!-- #             Median = median(Value, na.rm = T), -->
<!-- #             Mean = mean(Value, na.rm = T)) %>% -->
<!-- #   arrange(Category) -->
<!-- #  -->
<!-- # write.csv(summary_results_global_2050_pretty, paste0(results_dir, "summary_results_global_2050_pretty.csv")) -->
<!-- #  -->
<!-- # # Get rounded end values for the most ambitious scenario -->
<!-- # global_biomass_2050 <- round(all_results_global_output_table$Biomass[all_results_global_output_table$Year == 2050 & all_results_global_output_table$Id == "Ambitious Reform"]*100, 2) -->
<!-- # global_biomass_end <- round(all_results_global_output_table$Biomass[all_results_global_output_table$Year == end_year & all_results_global_output_table$Id == "Ambitious Reform"]*100, 2) -->
<!-- #  -->
<!-- # global_catches_2050 <- round(all_results_global_output_table$Catches[all_results_global_output_table$Year == 2050 & all_results_global_output_table$Id == "Ambitious Reform"]*100, 2) -->
<!-- # global_catches_end <- round(all_results_global_output_table$Catches[all_results_global_output_table$Year == end_year & all_results_global_output_table$Id == "Ambitious Reform"]*100, 2) -->

<!-- # # Now make a min, max, and average for each proposal type -->
<!-- # summary_results_regional_2050_pretty <- all_results_regional_output_table_2050_pretty %>% -->
<!-- #   gather(Variable, Value, c("Biomass", "Catches", "Revenue", "Fishing Mortality")) %>% -->
<!-- #   group_by(Region, Category, Year, Variable) %>% -->
<!-- #   summarize(Minimum = min(Value, na.rm = T), -->
<!-- #             Maximum = max(Value, na.rm = T), -->
<!-- #             Median = median(Value, na.rm = T), -->
<!-- #             Mean = mean(Value, na.rm = T)) %>% -->
<!-- #   arrange(Region, Category) -->
<!-- #  -->
<!-- # write.csv(summary_results_regional_2050_pretty, paste0(results_dir, "summary_results_regional_2050_pretty.csv")) -->
<!-- ``` -->

<!-- ## Plots -->

<!-- ```{r} -->
<!-- if(make_plots){ -->
<!--   # A) global (aggregated) "kobe" final year plot -->
<!--   box_2050_plot <- BioEconomicModelPlots(dat = all_ts_results, -->
<!--                                          spatial_res = "global", -->
<!--                                          temporal_res = "last", -->
<!--                                          end_year = 2050, -->
<!--                                          plot_name = "box_scenario_type_2050_plot", -->
<!--                                          plot_width = 7.5, -->
<!--                                          plot_height = 6, -->
<!--                                          results_dir = results_dir, -->
<!--                                          group_var = "run_group", -->
<!--                                          color_var = "run_group", -->
<!--                                          legend_pos_manual = "bottom", -->
<!--                                          legend_name = "Scenario Type", -->
<!--                                          plot_type = "box") -->

<!-- ggsave(paste0(results_dir, "proposal_effects_2050_boxplot.tiff"), box_2050_plot, width = 7.5, height = 6) -->
<!-- } -->
<!-- ``` -->

<!-- ## Do the same with absolute values -->

<!-- ```{r} -->
<!-- # # We want these to be ordered by date within each category, so we need to go back to our list -->
<!-- # date_order <- proposal_settings %>% -->
<!-- #   dplyr::select(title_tool, date) %>% -->
<!-- #   rownames_to_column("order") %>% -->
<!-- #   mutate(order = as.numeric(order)) -->
<!-- #  -->
<!-- # # Collect global results -->
<!-- # all_results_global_output_table_absolute <- plot_results %>% -->
<!-- #   dplyr::filter(Year %in% c(2050, end_year)) %>% -->
<!-- #   dplyr::filter(Variable %in% c("biomass", "catches_total", "revenue_total")) %>% -->
<!-- #   group_by(Year, Variable, Fleet, run_id, run_name, run_group) %>% -->
<!-- #   summarize(BAU_tot = sum(BAU, na.rm = T), -->
<!-- #             Reform_tot = sum(Reform, na.rm = T), -->
<!-- #             Diff_tot = case_when(BAU_tot != 0 ~ Reform_tot - BAU_tot, -->
<!-- #                             TRUE ~ 0)) %>% -->
<!-- #   ungroup() %>% -->
<!-- #   mutate(Variable = case_when(Variable == "biomass" ~ "Biomass", -->
<!-- #                                 Variable == "catches_total" ~ "Catches", -->
<!-- #                                 Variable == "revenue_total" ~ "Revenue")) %>% -->
<!-- #   dplyr::select(Id = run_id, -->
<!-- #                 Name = run_name, -->
<!-- #                 Category = run_group,  -->
<!-- #                 Year, -->
<!-- #                 Diff_tot_absolute = Diff_tot, -->
<!-- #                 Variable) %>% -->
<!-- #   spread(Variable, Diff_tot_absolute) %>% -->
<!-- #   left_join(date_order %>% dplyr::select(Name = title_tool, order), by = "Name") %>% -->
<!-- #   mutate(Category = factor(Category, levels = c("Ambitious Reform", "IUU", "OFOC", "Overfished", "Other & Multiple"))) %>% -->
<!-- #   arrange(Category, order, Year) %>% -->
<!-- #   dplyr::select(-order) -->
<!-- #  -->
<!-- # write.csv(all_results_global_output_table_absolute, paste0(results_dir, "all_results_global_output_table_absolute.csv")) -->

<!-- # write.csv(all_results_global_absolute_table, paste0(results_dir, "all_results_global_absolute_table.csv")) -->
<!-- #  -->
<!-- # all_results_regional_absolute_table <- plot_results %>% -->
<!-- #   dplyr::filter(Year %in% c(2050, end_year)) %>% -->
<!-- #   dplyr::filter(Variable %in% c("biomass", "catches_total")) %>% -->
<!-- #   mutate(Variable = case_when(Variable == "biomass" ~ "Biomass", -->
<!-- #                                 Variable == "catches_total" ~ "Catches")) %>% -->
<!-- #   mutate(abs_Diff = case_when(BAU != 0 ~ Reform-BAU, -->
<!-- #                               TRUE ~ 0)) %>% -->
<!-- #   dplyr::select(Id = run_id, -->
<!-- #                 Name = run_name, -->
<!-- #                 Category = run_group,  -->
<!-- #                 Year, -->
<!-- #                 Region, -->
<!-- #                 abs_Diff, -->
<!-- #                 Variable) %>% -->
<!-- #   spread(Variable, abs_Diff) -->
<!-- #  -->
<!-- # write.csv(all_results_regional_absolute_table, paste0(results_dir, "all_results_regional_absolute_table.csv")) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- # # Weight of a mature dairy cow in mt -->
<!-- # cow_weight <- 0.6803886 -->
<!-- #  -->
<!-- # all_results_global_cow_equivalent_table <- plot_results %>% -->
<!-- #   dplyr::filter(Year %in% c(2050, end_year)) %>% -->
<!-- #   dplyr::filter(Variable %in% c("biomass", "catches_total")) %>% -->
<!-- #   group_by(Year, Variable, Fleet, run_id, run_name, run_group) %>% -->
<!-- #   summarize(BAU_tot = sum(BAU, na.rm = T), -->
<!-- #             Reform_tot = sum(Reform, na.rm = T), -->
<!-- #             abs_Diff_tot = case_when(BAU_tot != 0 ~ Reform_tot-BAU_tot, -->
<!-- #                                      TRUE ~ 0), -->
<!-- #             cow_Diff = abs_Diff_tot*cow_weight) %>% -->
<!-- #   ungroup() %>% -->
<!-- #   mutate(Variable = case_when(Variable == "biomass" ~ "Biomass", -->
<!-- #                               Variable == "catches_total" ~ "Catches")) %>% -->
<!-- #   dplyr::select(Id = run_id, -->
<!-- #                 Name = run_name, -->
<!-- #                 Category = run_group,  -->
<!-- #                 Year, -->
<!-- #                 cow_Diff, -->
<!-- #                 Variable) %>% -->
<!-- #   spread(Variable, cow_Diff) -->
<!-- #  -->
<!-- # write.csv(all_results_global_cow_equivalent_table, paste0(results_dir, "all_results_global_cow_equivalent_table.csv")) -->
<!-- #  -->
<!-- # all_results_regional_cow_equivalent_table <- plot_results %>% -->
<!-- #   dplyr::filter(Year %in% c(2050, end_year)) %>% -->
<!-- #   dplyr::filter(Variable %in% c("biomass", "catches_total")) %>% -->
<!-- #   mutate(Variable = case_when(Variable == "biomass" ~ "Biomass", -->
<!-- #                                 Variable == "catches_total" ~ "Catches")) %>% -->
<!-- #   mutate(abs_Diff = case_when(BAU != 0 ~ Reform-BAU, -->
<!-- #                               TRUE ~ 0), -->
<!-- #          cow_Diff = abs_Diff*cow_weight) %>% -->
<!-- #   dplyr::select(Id = run_id, -->
<!-- #                 Name = run_name, -->
<!-- #                 Category = run_group,  -->
<!-- #                 Year, -->
<!-- #                 Region, -->
<!-- #                 cow_Diff, -->
<!-- #                 Variable) %>% -->
<!-- #   spread(Variable, cow_Diff) -->
<!-- #  -->
<!-- # write.csv(all_results_regional_cow_equivalent_table, paste0(results_dir, "all_results_regional_cow_equivalent_table.csv")) -->

<!-- ``` -->
